refactor(sharepoint): split spListRegistry definitions into domain-specific modules
