+++ src/sharepoint/definitions/attendance.ts
@@ -1,0 +1,120 @@
+import { ListKeys } from '@/sharepoint/fields';
+import { envOr, fromConfig } from '../spListRegistry.shared';
+import type { SpListEntry } from '../spListRegistry.shared';
+
+export const attendanceListEntries: readonly SpListEntry[] = [
+// ── 3. 出席管理系 ──────────────────────────────────────
+  {
+    key: 'daily_attendance',
+    displayName: '日次出欠',
+    resolve: () => envOr('VITE_SP_LIST_ATTENDANCE', fromConfig(ListKeys.DailyAttendance)),
+    operations: ['R', 'W'],
+    category: 'attendance',
+    lifecycle: 'required',
+    essentialFields: ['UserID', 'Date', 'Status'],
+    provisioningFields: [
+      { internalName: 'UserID', type: 'Text', displayName: 'User ID', required: true, candidates: ['UserID', 'User_x0020_ID'] },
+      { internalName: 'User_x0020_ID', type: 'Text', displayName: 'User ID (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['User_x0020_ID'] },
+      { internalName: 'Date', type: 'DateTime', displayName: 'Date', required: true, dateTimeFormat: 'DateOnly', indexed: true },
+      { internalName: 'Status', type: 'Choice', displayName: 'Status', choices: ['通常', '欠席', '振替', '休止'], required: true },
+      { internalName: 'IsTrial', type: 'Boolean', displayName: 'Trial', candidates: ['IsTrial', 'Trial'] },
+      { internalName: 'Trial', type: 'Boolean', displayName: 'Trial (Legacy)', governance: 'allow', isSilent: true, candidates: ['Trial'] },
+      { internalName: 'Notes', type: 'Note', displayName: 'Notes', richText: false },
+    ],
+  },
+  {
+    key: 'attendance_users',
+    displayName: '出席管理ユーザー',
+    resolve: () => envOr('VITE_SP_LIST_ATTENDANCE_USERS', fromConfig(ListKeys.AttendanceUsers)),
+    operations: ['R'],
+    category: 'attendance',
+    lifecycle: 'required',
+    essentialFields: [
+      'UserCode', 'IsActive', 'UsageStatus', 'ServiceEndDate'
+    ],
+    provisioningFields: [
+      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true },
+      { internalName: 'IsActive', type: 'Boolean', displayName: 'Is Active', default: true },
+      { internalName: 'UsageStatus', type: 'Text', displayName: 'Usage Status' },
+      { internalName: 'ServiceEndDate', type: 'DateTime', displayName: 'Service End Date', dateTimeFormat: 'DateOnly' },
+      { internalName: 'Service_x0020_End_x0020_Date', type: 'DateTime', displayName: 'Service End Date (Legacy SP-encoded)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['Service_x0020_End_x0020_Date'] },
+      { internalName: 'IsTransportTarget', type: 'Boolean', displayName: 'Is Transport Target' },
+      { internalName: 'StandardMinutes', type: 'Number', displayName: 'Standard Minutes' },
+      { internalName: 'DefaultTransportToMethod', type: 'Text', displayName: 'Default Transport To Method (Legacy)', governance: 'allow', isSilent: true, candidates: ['DefaultTransportToMethod'] },
+      { internalName: 'DefaultTransportFromMethod', type: 'Text', displayName: 'Default Transport From Method (Legacy)', governance: 'allow', isSilent: true, candidates: ['DefaultTransportFromMethod'] },
+      { internalName: 'DefaultTransportToNote', type: 'Note', displayName: 'Default Transport To Note (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['DefaultTransportToNote'] },
+      { internalName: 'DefaultTransportFromNote', type: 'Note', displayName: 'Default Transport From Note (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['DefaultTransportFromNote'] },
+    ],
+  },
+  {
+    key: 'attendance_daily',
+    displayName: '日次出席詳細',
+    resolve: () => envOr('VITE_SP_LIST_ATTENDANCE_DAILY', fromConfig(ListKeys.AttendanceDaily)),
+    operations: ['R', 'W'],
+    category: 'attendance',
+    lifecycle: 'required',
+    // ATTENDANCE_DAILY_ESSENTIALS と同期
+    essentialFields: ['UserCode', 'RecordDate', 'Status', 'CheckInAt'],
+    provisioningFields: [
+      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true, candidates: ['UserCode'] },
+      { internalName: 'RecordDate', type: 'DateTime', displayName: 'Record Date', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['RecordDate'] },
+      { internalName: 'Status', type: 'Text', displayName: 'Status', required: true },
+      { internalName: 'CheckInAt', type: 'DateTime', displayName: 'Check-in Time', dateTimeFormat: 'DateTime', candidates: ['CheckInAt'] },
+      { internalName: 'CheckOutAt', type: 'DateTime', displayName: 'Check-out Time', dateTimeFormat: 'DateTime', candidates: ['CheckOutAt'] },
+      { internalName: 'CntAttendIn', type: 'Number', displayName: 'Attendance In Count', isSilent: true, candidates: ['CntAttendIn'] },
+      { internalName: 'CntAttendOut', type: 'Number', displayName: 'Attendance Out Count', isSilent: true, candidates: ['CntAttendOut'] },
+      { internalName: 'TransportTo', type: 'Boolean', displayName: 'Transport To', candidates: ['TransportTo'] },
+      { internalName: 'TransportFrom', type: 'Boolean', displayName: 'Transport From', candidates: ['TransportFrom'] },
+      { internalName: 'ProvidedMinutes', type: 'Number', displayName: 'Provided Minutes', candidates: ['ProvidedMinutes'] },
+      { internalName: 'IsEarlyLeave', type: 'Boolean', displayName: 'Early Leave', isSilent: true, candidates: ['IsEarlyLeave'] },
+      { internalName: 'UserConfirmedAt', type: 'DateTime', displayName: 'User Confirmed Time', dateTimeFormat: 'DateTime', isSilent: true, candidates: ['UserConfirmedAt'] },
+      { internalName: 'AbsentMorningContacted', type: 'Boolean', displayName: 'Absent Morning Contacted', isSilent: true, candidates: ['AbsentMorningContacted'] },
+      { internalName: 'AbsentMorningMethod', type: 'Text', displayName: 'Absent Morning Method', isSilent: true, candidates: ['AbsentMorningMethod'] },
+      { internalName: 'EveningChecked', type: 'Boolean', displayName: 'Evening Checked', candidates: ['EveningChecked'] },
+      { internalName: 'EveningNote', type: 'Note', displayName: 'Evening Note', candidates: ['EveningNote'] },
+      { internalName: 'IsAbsenceAddonClaimable', type: 'Boolean', displayName: 'Absence Add-on Claimable', isSilent: true, candidates: ['IsAbsenceAddonClaimable'] },
+    ],
+  },
+  {
+    key: 'staff_attendance',
+    displayName: '職員出勤管理',
+    resolve: () => envOr('VITE_SP_LIST_STAFF_ATTENDANCE', fromConfig(ListKeys.StaffAttendance)),
+    operations: ['R', 'W'],
+    category: 'attendance',
+    lifecycle: 'required',
+    essentialFields: ['StaffId', 'RecordDate', 'Status'],
+    provisioningFields: [
+      { internalName: 'StaffId', type: 'Text', displayName: 'Staff ID', required: true, indexed: true, candidates: ['StaffId', 'Staff_x0020_ID'] },
+      { internalName: 'Staff_x0020_ID', type: 'Text', displayName: 'Staff ID (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Staff_x0020_ID'] },
+      { internalName: 'RecordDate', type: 'DateTime', displayName: 'Record Date', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['RecordDate', 'Record_x0020_Date'] },
+      { internalName: 'Status', type: 'Text', displayName: 'Status', required: true },
+      { internalName: 'CheckInAt', type: 'DateTime', displayName: 'Check-in', dateTimeFormat: 'DateTime', candidates: ['CheckInAt', 'Check_x002d_in'] },
+      { internalName: 'CheckOutAt', type: 'DateTime', displayName: 'Check-out', dateTimeFormat: 'DateTime', candidates: ['CheckOutAt', 'Check_x002d_out'] },
+      { internalName: 'IsFinalized', type: 'Boolean', displayName: 'Finalized', default: false, candidates: ['IsFinalized', 'Finalized'] },
+      { internalName: 'Finalized', type: 'Boolean', displayName: 'Finalized (Legacy)', governance: 'allow', isSilent: true, candidates: ['Finalized'] },
+      { internalName: 'Note', type: 'Note', displayName: 'Note' },
+    ],
+  },
+  {
+    key: 'transport_log',
+    displayName: '送迎ステータスログ',
+    resolve: () => envOr('VITE_SP_LIST_TRANSPORT_LOG', fromConfig(ListKeys.TransportLog)),
+    operations: ['R', 'W'],
+    category: 'attendance',
+    lifecycle: 'required',
+    essentialFields: ['UserCode', 'RecordDate', 'Direction', 'Status'],
+    provisioningFields: [
+      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true, candidates: ['UserCode', 'User_x0020_Code'] },
+      { internalName: 'RecordDate', type: 'DateTime', displayName: 'Record Date', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['RecordDate', 'Record_x0020_Date'] },
+      { internalName: 'Direction', type: 'Choice', displayName: 'Direction', choices: ['to', 'from'], required: true },
+      { internalName: 'Status', type: 'Choice', displayName: 'Status', choices: ['pending', 'in-progress', 'arrived', 'absent', 'self'], required: true },
+      { internalName: 'Method', type: 'Choice', displayName: 'Method', choices: ['facility-vehicle', 'family', 'taxi', 'walk', 'self', 'other'] },
+      { internalName: 'ScheduledTime', type: 'Text', displayName: 'Scheduled Time', candidates: ['ScheduledTime', 'Scheduled_x0020_Time'] },
+      { internalName: 'ActualTime', type: 'Text', displayName: 'Actual Time', candidates: ['ActualTime', 'Actual_x0020_Time'] },
+      { internalName: 'DriverName', type: 'Text', displayName: 'Driver Name', candidates: ['DriverName', 'Driver_x0020_Name'] },
+      { internalName: 'Notes', type: 'Note', displayName: 'Notes', candidates: ['Notes', 'Note'] },
+      { internalName: 'UpdatedBy', type: 'Text', displayName: 'Updated By' },
+      { internalName: 'UpdatedAt', type: 'DateTime', displayName: 'Updated At', dateTimeFormat: 'DateTime' },
+    ],
+  },
+];
