+++ src/sharepoint/definitions/handoff.ts
@@ -1,0 +1,271 @@
+import { ListKeys } from '@/sharepoint/fields';
+import { envOr, fromConfig } from '../spListRegistry.shared';
+import type { SpListEntry } from '../spListRegistry.shared';
+
+export const handoffListEntries: readonly SpListEntry[] = [
+// ── 6. 引き継ぎ・支援計画系 ────────────────────────────
+  {
+    key: 'handoff',
+    displayName: '引き継ぎ',
+    resolve: () => envOr('VITE_SP_HANDOFF_LIST_TITLE', 'Handoff'),
+    operations: ['R', 'W', 'D'],
+    category: 'handoff',
+    lifecycle: 'required',
+    essentialFields: ['Message', 'UserCode', 'Category'],
+    provisioningFields: [
+      { internalName: 'Message', type: 'Note', displayName: 'Message', required: true },
+      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true },
+      { internalName: 'Category', type: 'Text', displayName: 'Category', required: true },
+      { internalName: 'Severity', type: 'Text', displayName: 'Severity' },
+      { internalName: 'Status', type: 'Text', displayName: 'Status' },
+      { internalName: 'TimeBand', type: 'Text', displayName: 'Time Band' },
+      { internalName: 'SourceType', type: 'Text', displayName: 'Source Type' },
+      { internalName: 'SourceKey', type: 'Text', displayName: 'Source Key' },
+      { internalName: 'UserDisplayName', type: 'Text', displayName: 'User Display Name', governance: 'allow', isSilent: true, candidates: ['UserDisplayName'] },
+      { internalName: 'CreatedAt', type: 'DateTime', displayName: 'Created At', governance: 'allow', isSilent: true, candidates: ['CreatedAt'] },
+      { internalName: 'CreatedByName', type: 'Text', displayName: 'Created By Name', governance: 'allow', isSilent: true, candidates: ['CreatedByName'] },
+      { internalName: 'IsDraft', type: 'Boolean', displayName: 'Is Draft', governance: 'allow', isSilent: true, candidates: ['IsDraft'] },
+      { internalName: 'MeetingSessionKey', type: 'Text', displayName: 'Meeting Session Key', governance: 'allow', isSilent: true, candidates: ['MeetingSessionKey'] },
+      { internalName: 'SourceId', type: 'Text', displayName: 'Source ID', governance: 'allow', isSilent: true, candidates: ['SourceId'] },
+      { internalName: 'SourceUrl', type: 'Text', displayName: 'Source URL', governance: 'allow', isSilent: true, candidates: ['SourceUrl'] },
+      { internalName: 'SourceLabel', type: 'Text', displayName: 'Source Label', governance: 'allow', isSilent: true, candidates: ['SourceLabel'] },
+      { internalName: 'CarryOverDate', type: 'DateTime', displayName: 'Carry Over Date', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['CarryOverDate'] },
+      { internalName: 'CreatedBy', type: 'Text', displayName: 'Created By (Legacy)', governance: 'allow', isSilent: true, candidates: ['CreatedBy'] },
+      { internalName: 'ModifiedBy', type: 'Text', displayName: 'Modified By (Legacy)', governance: 'allow', isSilent: true, candidates: ['ModifiedBy'] },
+      { internalName: 'ModifiedAt', type: 'DateTime', displayName: 'Modified At (Legacy)', governance: 'allow', isSilent: true, candidates: ['ModifiedAt'] },
+    ],
+  },
+  {
+    key: 'support_templates',
+    displayName: '支援手順テンプレート',
+    resolve: () => envOr('VITE_SP_LIST_SUPPORT_TEMPLATES', fromConfig(ListKeys.SupportTemplates)),
+    operations: ['R'],
+    category: 'handoff',
+    lifecycle: 'required',
+    essentialFields: ['UserCode0', 'RowNo0', 'TimeSlot0', 'Activity0'],
+    provisioningFields: [
+      { internalName: 'UserCode0', type: 'Text', displayName: '利用者コード', indexed: true, candidates: ['UserCode0', 'UserCode'] },
+      { internalName: 'RowNo0', type: 'Number', displayName: '行番号', candidates: ['RowNo0', 'RowNo'] },
+      { internalName: 'TimeSlot0', type: 'Text', displayName: '時間帯', candidates: ['TimeSlot0', 'TimeSlot'] },
+      { internalName: 'Activity0', type: 'Text', displayName: '活動内容', candidates: ['Activity0', 'Activity'] },
+      { internalName: 'PersonManual0', type: 'Note', displayName: '本人手順', candidates: ['PersonManual0', 'PersonManual'] },
+      { internalName: 'SupporterManual0', type: 'Note', displayName: '支援者手順', candidates: ['SupporterManual0', 'SupporterManual'] },
+      { internalName: 'UserID', type: 'Text', displayName: 'User ID (Legacy)', governance: 'allow', isSilent: true },
+      { internalName: 'IsActive', type: 'Boolean', displayName: 'Is Active (Legacy)', governance: 'allow', isSilent: true },
+      { internalName: 'modifiedBy', type: 'Text', displayName: 'Modified By (Legacy)', governance: 'allow', isSilent: true, candidates: ['modifiedBy'] },
+      { internalName: 'Note', type: 'Note', displayName: 'Note (Legacy)', governance: 'allow', isSilent: true },
+      { internalName: 'Notes', type: 'Note', displayName: 'Notes (Legacy)', governance: 'allow', isSilent: true },
+    ],
+  },
+  {
+    key: 'plan_goals',
+    displayName: '支援計画目標',
+    resolve: () => envOr('VITE_SP_LIST_PLAN_GOAL', fromConfig(ListKeys.PlanGoal)),
+    operations: ['R', 'W'],
+    category: 'handoff',
+    lifecycle: 'required',
+    essentialFields: ['UserCode', 'GoalType', 'GoalText', 'PlanStatus'],
+    provisioningFields: [
+      { internalName: 'UserCode', type: 'Text', displayName: '利用者コード', required: true, indexed: true, candidates: ['UserCode', 'cr013_usercode', 'UserID'] },
+      { internalName: 'GoalType', type: 'Text', displayName: '目標種別', required: true, candidates: ['GoalType', 'cr013_goaltype'] },
+      { internalName: 'GoalLabel', type: 'Text', displayName: '目標ラベル', isSilent: true, candidates: ['GoalLabel', 'cr013_goallabel', 'Title'] },
+      { internalName: 'GoalText', type: 'Note', displayName: '目標内容', richText: false, candidates: ['GoalText', 'cr013_goaltext'] },
+      { internalName: 'Domains', type: 'Text', displayName: '領域', isSilent: true, candidates: ['Domains', 'cr013_domains'] },
+      { internalName: 'PlanPeriod', type: 'Text', displayName: '計画期間', isSilent: true, candidates: ['PlanPeriod', 'cr013_planperiod'] },
+      { internalName: 'PlanStatus', type: 'Text', displayName: '計画状態', candidates: ['PlanStatus', 'cr013_planstatus'] },
+      { internalName: 'CertExpiry', type: 'Text', displayName: '受給者証期限', isSilent: true, candidates: ['CertExpiry', 'cr013_certexpiry'] },
+      { internalName: 'SortOrder', type: 'Number', displayName: '表示順', isSilent: true, candidates: ['SortOrder', 'cr013_sortorder'] },
+    ],
+  },
+  {
+    key: 'support_plans',
+    displayName: '個別支援計画',
+    resolve: () => envOr('VITE_SP_LIST_SUPPORT_PLANS', fromConfig(ListKeys.SupportPlans)),
+    operations: ['R', 'W', 'D'],
+    category: 'handoff',
+    lifecycle: 'required', // 制度上必須のため required
+    essentialFields: ['DraftId', 'UserCode', 'FormDataJson'],
+    provisioningFields: [
+      { internalName: 'DraftId', type: 'Text', displayName: 'Draft ID', required: true, indexed: true },
+      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true },
+      { internalName: 'DraftName', type: 'Text', displayName: 'Draft Name' },
+      { internalName: 'FormDataJson', type: 'Note', displayName: 'Form Data (JSON)', required: true, richText: false },
+      { internalName: 'Status', type: 'Text', displayName: 'Status' },
+      { internalName: 'SchemaVersion', type: 'Number', displayName: 'Schema Version', default: 2 },
+      { internalName: 'UserCode0', type: 'Text', displayName: 'User Code (Legacy)', governance: 'allow', isSilent: true, candidates: ['UserCode0'] },
+      { internalName: 'Status0', type: 'Text', displayName: 'Status (Legacy)', governance: 'allow', isSilent: true, candidates: ['Status0'] },
+    ],
+  },
+  {
+    key: 'iceberg_pdca',
+    displayName: '氷山モデルPDCA',
+    resolve: () => envOr('VITE_SP_LIST_ICEBERG_PDCA', fromConfig(ListKeys.IcebergPdca)),
+    operations: ['R', 'W', 'D'],
+    category: 'handoff',
+    lifecycle: 'optional',
+    essentialFields: ['Title', 'UserID0', 'Phase0'],
+    provisioningFields: [
+      { internalName: 'Title', type: 'Text', displayName: 'Title', required: true },
+      { internalName: 'UserID0', type: 'Text', displayName: 'User ID', required: true, indexed: true, candidates: ['UserID0', 'userId', 'UserID', 'UserCode'] },
+      { internalName: 'PlanningSheetId', type: 'Text', displayName: 'Planning Sheet ID' },
+      { internalName: 'Summary0', type: 'Note', displayName: 'Summary', richText: false, candidates: ['Summary0', 'summary'] },
+      { internalName: 'Phase0', type: 'Text', displayName: 'Phase', candidates: ['Phase0', 'phase'] },
+      { internalName: 'userId', type: 'Text', displayName: 'User ID (Legacy)', governance: 'allow', isSilent: true },
+      { internalName: 'summary', type: 'Note', displayName: 'Summary (Legacy)', governance: 'allow', richText: false, isSilent: true },
+      { internalName: 'phase', type: 'Text', displayName: 'Phase (Legacy)', governance: 'allow', isSilent: true },
+    ],
+  },
+  {
+    key: 'iceberg_analysis',
+    displayName: '氷山モデル分析',
+    resolve: () => envOr('VITE_SP_LIST_ICEBERG_ANALYSIS', fromConfig(ListKeys.IcebergAnalysis)),
+    operations: ['R', 'W'],
+    category: 'handoff',
+    lifecycle: 'optional',
+    essentialFields: ['Title', 'EntryHash', 'UserId', 'PayloadJson'],
+    provisioningFields: [
+      { internalName: 'Title', type: 'Text', displayName: 'Title', required: true },
+      { internalName: 'EntryHash', type: 'Text', displayName: 'Entry Hash', required: true, indexed: true },
+      { internalName: 'SessionId', type: 'Text', displayName: 'Session ID', indexed: true },
+      { internalName: 'UserId', type: 'Text', displayName: 'User ID', indexed: true },
+      { internalName: 'PayloadJson', type: 'Note', displayName: 'Payload JSON', required: true, richText: false },
+      { internalName: 'SchemaVersion', type: 'Number', displayName: 'Schema Version' },
+      { internalName: 'UpdatedAt', type: 'DateTime', displayName: 'Updated At', indexed: true },
+    ],
+  },
+  {
+    key: 'isp_master',
+    displayName: '個別支援計画（ISP）',
+    resolve: () => envOr('VITE_SP_LIST_ISP_MASTER', fromConfig(ListKeys.IspMaster)),
+    operations: ['R', 'W'],
+    category: 'handoff',
+    lifecycle: 'required', // 三層モデルの核となるため required
+    essentialFields: ['UserCode', 'PlanStartDate', 'Status'],
+    provisioningFields: [
+      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true, candidates: ['UserCode', 'UserID', 'User_ID', 'cr013_userCode'] },
+      { internalName: 'PlanStartDate', type: 'DateTime', displayName: 'Plan Start Date', required: true, dateTimeFormat: 'DateOnly', candidates: ['PlanStartDate', 'StartDate', 'cr013_planStartDate'] },
+      { internalName: 'PlanEndDate', type: 'DateTime', displayName: 'Plan End Date', dateTimeFormat: 'DateOnly', candidates: ['PlanEndDate', 'EndDate', 'cr013_planEndDate'] },
+      { internalName: 'Status', type: 'Text', displayName: 'Status', required: true, candidates: ['Status', 'UsageStatus', 'cr013_status'] },
+      { internalName: 'NextReviewAt', type: 'DateTime', displayName: 'Next Review At (Legacy)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['NextReviewAt'] },
+      { internalName: 'UserIntent', type: 'Note', displayName: 'User Intent (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['UserIntent'] },
+      { internalName: 'Precautions', type: 'Note', displayName: 'Precautions (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['Precautions'] },
+      { internalName: 'FamilyIntent', type: 'Note', displayName: 'Family Intent (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['FamilyIntent'] },
+      { internalName: 'OverallSupportPolicy', type: 'Note', displayName: 'Overall Support Policy (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['OverallSupportPolicy'] },
+      { internalName: 'QolIssues', type: 'Note', displayName: 'QOL Issues (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['QolIssues'] },
+      { internalName: 'LongTermGoalsJson', type: 'Note', displayName: 'Long Term Goals JSON (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['LongTermGoalsJson'] },
+      { internalName: 'ShortTermGoalsJson', type: 'Note', displayName: 'Short Term Goals JSON (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['ShortTermGoalsJson'] },
+      { internalName: 'SupportSummary', type: 'Note', displayName: 'Support Summary (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['SupportSummary'] },
+      { internalName: 'MonitoringSummary', type: 'Note', displayName: 'Monitoring Summary (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['MonitoringSummary'] },
+      { internalName: 'ConsentAt', type: 'DateTime', displayName: 'Consent At (Legacy)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['ConsentAt'] },
+      { internalName: 'DeliveredAt', type: 'DateTime', displayName: 'Delivered At (Legacy)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['DeliveredAt'] },
+      { internalName: 'LastMonitoringAt', type: 'DateTime', displayName: 'Last Monitoring At (Legacy)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['LastMonitoringAt'] },
+      { internalName: 'VersionNo', type: 'Number', displayName: 'Version No', default: 1, governance: 'allow' },
+      { internalName: 'IsCurrent', type: 'Boolean', displayName: 'Is Current', default: true },
+      { internalName: 'FormDataJson', type: 'Note', displayName: 'Form Data (JSON)', richText: false },
+      { internalName: 'UserSnapshotJson', type: 'Note', displayName: 'User Snapshot (JSON)', richText: false },
+    ],
+  },
+  {
+    key: 'planning_sheet_master',
+    displayName: '支援計画シート',
+    resolve: () => envOr('VITE_SP_LIST_PLANNING_SHEET', fromConfig(ListKeys.PlanningSheetMaster)),
+    operations: ['R', 'W'],
+    category: 'handoff',
+    lifecycle: 'optional',
+    essentialFields: ['Status', 'UserCode', 'UserId'],
+    provisioningFields: [
+      { internalName: 'Status', type: 'Text', displayName: 'Status', required: true, candidates: ['Status', 'UsageStatus', 'cr013_status'] },
+      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true, candidates: ['UserCode', 'User_ID', 'cr013_userCode'] },
+      { internalName: 'UserId', type: 'Text', displayName: 'User ID', required: true, indexed: true, candidates: ['UserId', 'UserID', 'cr013_userId'] },
+      { internalName: 'DecisionStatus', type: 'Text', displayName: 'Decision Status', candidates: ['DecisionStatus'] },
+      { internalName: 'IsCurrent', type: 'Boolean', displayName: 'Is Current', default: true },
+      { internalName: 'ISPId', type: 'Number', displayName: 'ISP ID', candidates: ['ISPId', 'ISP_x0020_ID', 'cr013_ispId'] },
+      { internalName: 'TargetScene', type: 'Text', displayName: 'Target Scene', candidates: ['TargetScene', 'Scene', 'cr013_targetScene'] },
+      { internalName: 'TargetDomain', type: 'Text', displayName: 'Target Domain', candidates: ['TargetDomain', 'Domain', 'cr013_targetDomain'] },
+      { internalName: 'ObservationFacts', type: 'Note', displayName: 'Observation Facts', richText: false, candidates: ['ObservationFacts', 'cr013_observationFacts'] },
+      { internalName: 'CollectedInformation', type: 'Note', displayName: 'Collected Information', richText: false, candidates: ['CollectedInformation', 'cr013_collectedInformation'] },
+      { internalName: 'InterpretationHypothesis', type: 'Note', displayName: 'Interpretation Hypothesis', richText: false, candidates: ['InterpretationHypothesis', 'cr013_interpretationHypothesis'] },
+      { internalName: 'SupportIssues', type: 'Note', displayName: 'Support Issues', richText: false, candidates: ['SupportIssues', 'cr013_supportIssues'] },
+      { internalName: 'SupportPolicy', type: 'Note', displayName: 'Support Policy', richText: false, candidates: ['SupportPolicy', 'cr013_supportPolicy'] },
+      { internalName: 'EnvironmentalAdjustments', type: 'Note', displayName: 'Environmental Adjustments', richText: false, candidates: ['EnvironmentalAdjustments', 'cr013_environmentalAdjustments'] },
+      { internalName: 'ConcreteApproaches', type: 'Note', displayName: 'Concrete Approaches', richText: false, candidates: ['ConcreteApproaches', 'cr013_concreteApproaches'] },
+      { internalName: 'AppliedFrom', type: 'DateTime', displayName: 'Applied From', dateTimeFormat: 'DateOnly', candidates: ['AppliedFrom', 'cr013_appliedFrom'] },
+      { internalName: 'NextReviewAt', type: 'DateTime', displayName: 'Next Review At', dateTimeFormat: 'DateOnly', candidates: ['NextReviewAt', 'cr013_nextReviewAt'] },
+      { internalName: 'AuthoredByStaffId', type: 'Text', displayName: 'Authored By Staff ID', candidates: ['AuthoredByStaffId', 'cr013_authoredByStaffId'] },
+      { internalName: 'AuthoredByQualification', type: 'Text', displayName: 'Authored By Qualification', candidates: ['AuthoredByQualification', 'cr013_authoredByQualification'] },
+      { internalName: 'AuthoredAt', type: 'DateTime', displayName: 'Authored At', dateTimeFormat: 'DateTime', candidates: ['AuthoredAt', 'cr013_authoredAt'] },
+      { internalName: 'ApplicableServiceType', type: 'Text', displayName: 'Applicable Service Type', candidates: ['ApplicableServiceType', 'cr013_applicableServiceType'] },
+      { internalName: 'ApplicableAddOnTypesJson', type: 'Note', displayName: 'Applicable Add-on Types JSON', richText: false, candidates: ['ApplicableAddOnTypesJson', 'cr013_applicableAddOnTypesJson'] },
+      { internalName: 'DeliveredToUserAt', type: 'DateTime', displayName: 'Delivered To User At', dateTimeFormat: 'DateOnly', candidates: ['DeliveredToUserAt', 'cr013_deliveredToUserAt'] },
+      { internalName: 'ReviewedAt', type: 'DateTime', displayName: 'Reviewed At', dateTimeFormat: 'DateOnly', candidates: ['ReviewedAt', 'cr013_reviewedAt'] },
+      { internalName: 'HasMedicalCoordination', type: 'Boolean', displayName: 'Has Medical Coordination', candidates: ['HasMedicalCoordination', 'cr013_hasMedicalCoordination'] },
+      { internalName: 'HasEducationCoordination', type: 'Boolean', displayName: 'Has Education Coordination', candidates: ['HasEducationCoordination', 'cr013_hasEducationCoordination'] },
+      { internalName: 'RegulatoryBasisSnapshotJson', type: 'Note', displayName: 'Regulatory Basis Snapshot JSON', richText: false, candidates: ['RegulatoryBasisSnapshotJson', 'cr013_regulatoryBasisSnapshotJson'] },
+      { internalName: 'IntakeJson', type: 'Note', displayName: 'Intake JSON', richText: false, candidates: ['IntakeJson', 'cr013_intakeJson'] },
+      { internalName: 'AssessmentJson', type: 'Note', displayName: 'Assessment JSON', richText: false, candidates: ['AssessmentJson', 'cr013_assessmentJson'] },
+      { internalName: 'PlanningJson', type: 'Note', displayName: 'Planning JSON', richText: false, candidates: ['PlanningJson', 'cr013_planningJson'] },
+      { internalName: 'GoalId', type: 'Text', displayName: 'Goal ID', candidates: ['GoalId'] },
+      { internalName: 'GoalLabel', type: 'Text', displayName: 'Goal Label', candidates: ['GoalLabel'] },
+      { internalName: 'DecisionNote', type: 'Note', displayName: 'Decision Note', richText: false, candidates: ['DecisionNote'] },
+      { internalName: 'DecisionBy', type: 'Text', displayName: 'Decision By', candidates: ['DecisionBy'] },
+      { internalName: 'DecisionAt', type: 'DateTime', displayName: 'Decision At', dateTimeFormat: 'DateTime', candidates: ['DecisionAt'] },
+      { internalName: 'RecommendationLevel', type: 'Text', displayName: 'Recommendation Level', candidates: ['RecommendationLevel'] },
+      { internalName: 'SnapshotJson', type: 'Note', displayName: 'Snapshot JSON', richText: false, candidates: ['SnapshotJson'] },
+      { internalName: 'FormDataJson', type: 'Note', displayName: 'Form Data (JSON)', richText: false, candidates: ['FormDataJson', 'cr013_formDataJson'] },
+    ],
+  },
+  {
+    key: 'behavior_monitoring_master',
+    displayName: '行動モニタリング記録',
+    resolve: () => envOr('VITE_SP_LIST_BEHAVIOR_MONITORING', fromConfig(ListKeys.BehaviorMonitoringMaster)),
+    operations: ['R', 'W'],
+    category: 'handoff',
+    lifecycle: 'required',
+    essentialFields: ['UserId', 'PlanningSheetId', 'PeriodEnd'],
+    provisioningFields: [
+      { internalName: 'UserId', type: 'Text', displayName: 'User ID', required: true, indexed: true, isSilent: true, candidates: ['UserId', 'UserCode', 'User_x0020_ID', 'cr013_userId'] },
+      { internalName: 'PlanningSheetId', type: 'Text', displayName: 'Planning Sheet ID', required: true, isSilent: true, candidates: ['PlanningSheetId', 'Planning_x0020_Sheet_x0020_ID', 'cr013_planningSheetId'] },
+      { internalName: 'PeriodStart', type: 'DateTime', displayName: 'Period Start', dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['PeriodStart', 'PeriodStart0', 'cr013_periodStart'] },
+      { internalName: 'PeriodEnd', type: 'DateTime', displayName: 'Period End', required: true, dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['PeriodEnd', 'Period_x0020_End', 'cr013_periodEnd'] },
+      { internalName: 'SupportEvaluationsJson', type: 'Note', displayName: 'Support Evaluations JSON', richText: false, isSilent: true, candidates: ['SupportEvaluationsJson', 'SupportEvaluationsJson0', 'cr013_supportEvaluationsJson'] },
+      { internalName: 'EnvironmentFindingsJson', type: 'Note', displayName: 'Environment Findings JSON', richText: false, isSilent: true, candidates: ['EnvironmentFindingsJson', 'EnvironmentFindingsJson0', 'cr013_environmentFindingsJson'] },
+      { internalName: 'EffectiveSupports', type: 'Note', displayName: 'Effective Supports', richText: false, isSilent: true, candidates: ['EffectiveSupports', 'EffectiveSupports0', 'cr013_effectiveSupports'] },
+      { internalName: 'DifficultiesObserved', type: 'Note', displayName: 'Difficulties Observed', richText: false, isSilent: true, candidates: ['DifficultiesObserved', 'DifficultiesObserved0', 'cr013_difficultiesObserved'] },
+      { internalName: 'NewTriggersJson', type: 'Note', displayName: 'New Triggers JSON', richText: false, isSilent: true, candidates: ['NewTriggersJson', 'NewTriggersJson0', 'cr013_newTriggersJson'] },
+      { internalName: 'MedicalSafetyNotes', type: 'Note', displayName: 'Medical Safety Notes', richText: false, isSilent: true, candidates: ['MedicalSafetyNotes', 'MedicalSafetyNotes0', 'cr013_medicalSafetyNotes'] },
+      { internalName: 'UserFeedback', type: 'Note', displayName: 'User Feedback', richText: false, isSilent: true, candidates: ['UserFeedback', 'UserFeedback0', 'cr013_userFeedback'] },
+      { internalName: 'FamilyFeedback', type: 'Note', displayName: 'Family Feedback', richText: false, isSilent: true, candidates: ['FamilyFeedback', 'FamilyFeedback0', 'cr013_familyFeedback'] },
+      { internalName: 'RecommendedChangesJson', type: 'Note', displayName: 'Recommended Changes JSON', richText: false, isSilent: true, candidates: ['RecommendedChangesJson', 'RecommendedChangesJson0', 'cr013_recommendedChangesJson'] },
+      { internalName: 'Summary', type: 'Note', displayName: 'Summary', richText: false, isSilent: true, candidates: ['Summary', 'Summary0', 'cr013_summary'] },
+      { internalName: 'RecordedBy', type: 'Text', displayName: 'Recorded By', isSilent: true, candidates: ['RecordedBy', 'RecordedBy0', 'cr013_recordedBy'] },
+      { internalName: 'RecordedAt', type: 'DateTime', displayName: 'Recorded At', isSilent: true, candidates: ['RecordedAt', 'RecordedAt0', 'cr013_recordedAt'] },
+    ],
+  },
+  {
+    key: 'planning_sheet_reassessment_master',
+    displayName: 'アセスメント・再アセスメント履歴',
+    resolve: () => envOr('VITE_SP_LIST_PLANNING_REASSESSMENT', fromConfig(ListKeys.PlanningSheetReassessmentMaster)),
+    operations: ['R', 'W'],
+    category: 'handoff',
+    lifecycle: 'required',
+    essentialFields: ['PlanningSheetId', 'ReassessmentDate'],
+    provisioningFields: [
+      { internalName: 'PlanningSheetId', type: 'Text', displayName: 'Planning Sheet ID', required: true, indexed: true, isSilent: true, candidates: ['PlanningSheetId', 'Planning_x0020_Sheet_x0020_ID', 'cr013_planningSheetId'] },
+      { internalName: 'UserId', type: 'Text', displayName: 'User ID', isSilent: true, candidates: ['UserId', 'UserCode', 'cr013_userId'] },
+      { internalName: 'ReassessmentTrigger', type: 'Text', displayName: 'Reassessment Trigger', isSilent: true, candidates: ['ReassessmentTrigger', 'ReassessmentTrigger0', 'cr013_reassessmentTrigger'] },
+      { internalName: 'ReassessmentDate', type: 'DateTime', displayName: 'Reassessment Date', required: true, dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['ReassessmentDate', 'Reassessment_x0020_Date', 'cr013_reassessmentDate'] },
+      { internalName: 'Summary', type: 'Note', displayName: 'Summary', richText: false, isSilent: true, candidates: ['Summary', 'Summary0', 'cr013_summary'] },
+      { internalName: 'CreatedByText', type: 'Text', displayName: 'Created By (Text)', isSilent: true, candidates: ['CreatedByText', 'CreatedBy', 'cr013_createdByText'] },
+      { internalName: 'CreatedAtText', type: 'Text', displayName: 'Created At (Text)', isSilent: true, candidates: ['CreatedAtText', 'CreatedAt', 'cr013_createdAtText'] },
+      { internalName: 'VersionNo', type: 'Number', displayName: 'Version No', governance: 'allow', isSilent: true, candidates: ['VersionNo', 'Version', 'cr013_versionNo'] },
+      { internalName: 'PlanChangeDecision', type: 'Text', displayName: 'Plan Change Decision', isSilent: true, candidates: ['PlanChangeDecision', 'PlanChangeDecision0', 'cr013_planChangeDecision'] },
+      { internalName: 'AbcSummary', type: 'Note', displayName: 'ABC Summary', richText: false, isSilent: true, candidates: ['AbcSummary', 'AbcSummary0', 'cr013_abcSummary'] },
+      { internalName: 'HypothesisReview', type: 'Note', displayName: 'Hypothesis Review', richText: false, isSilent: true, candidates: ['HypothesisReview', 'HypothesisReview0', 'cr013_hypothesisReview'] },
+      { internalName: 'ProcedureEffectiveness', type: 'Note', displayName: 'Procedure Effectiveness', richText: false, isSilent: true, candidates: ['ProcedureEffectiveness', 'ProcedureEffectiveness0', 'cr013_procedureEffectiveness'] },
+      { internalName: 'EnvironmentChange', type: 'Note', displayName: 'Environment Change', richText: false, isSilent: true, candidates: ['EnvironmentChange', 'EnvironmentChange0', 'cr013_environmentChange'] },
+      { internalName: 'NextReassessmentAt', type: 'DateTime', displayName: 'Next Reassessment At', dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['NextReassessmentAt', 'NextReassessmentAt0', 'cr013_nextReassessmentAt'] },
+      { internalName: 'Notes', type: 'Note', displayName: 'Notes', richText: false, isSilent: true, candidates: ['Notes', 'Notes0', 'cr013_notes'] },
+      { internalName: 'ReassessedBy', type: 'Text', displayName: 'Reassessed By', isSilent: true, candidates: ['ReassessedBy', 'ReassessedBy0', 'cr013_reassessedBy'] },
+    ],
+  },
+];
