+++ src/sharepoint/definitions/index.ts
@@ -1,0 +1,8 @@
+export * from './master';
+export * from './daily';
+export * from './attendance';
+export * from './schedule';
+export * from './meeting';
+export * from './handoff';
+export * from './compliance';
+export * from './other';
