+++ src/sharepoint/definitions/master.ts
@@ -1,0 +1,196 @@
+import { ListKeys } from '@/sharepoint/fields';
+import { envOr, fromConfig } from '../spListRegistry.shared';
+import type { SpListEntry } from '../spListRegistry.shared';
+
+export const masterListEntries: readonly SpListEntry[] = [
+// ── 1. マスタ系 ──────────────────────────────────────────
+  {
+    key: 'users_master',
+    displayName: '利用者マスタ',
+    resolve: () => envOr('VITE_SP_LIST_USERS', fromConfig(ListKeys.UsersMaster)),
+    operations: ['R', 'W'],
+    category: 'master',
+    lifecycle: 'required', // 制度管理の基盤となるため required
+    essentialFields: [
+      'UserID', 'FullName', 'IsActive', 'UsageStatus'
+    ],
+    provisioningFields: [
+      { internalName: 'UserID', type: 'Text', displayName: 'User Code', required: true, indexed: true, candidates: ['UserID', 'userCode', 'UserCode'] },
+      { internalName: 'FullName', type: 'Text', displayName: 'Full Name', required: true, candidates: ['FullName', 'Full_x0020_Name', 'Title'] },
+      { internalName: 'Furigana', type: 'Text', displayName: 'Furigana' },
+      { internalName: 'FullNameKana', type: 'Text', displayName: 'Full Name Kana', candidates: ['FullNameKana', 'Full_x0020_Name_x0020_Kana'] },
+      { internalName: 'ContractDate', type: 'DateTime', displayName: 'Contract Date', dateTimeFormat: 'DateOnly', candidates: ['ContractDate', 'Contract_x0020_Date'] },
+      { internalName: 'ServiceStartDate', type: 'DateTime', displayName: 'Service Start Date', dateTimeFormat: 'DateOnly', candidates: ['ServiceStartDate', 'Service_x0020_Start_x0020_Date'] },
+      { internalName: 'ServiceEndDate', type: 'DateTime', displayName: 'Service End Date', dateTimeFormat: 'DateOnly', candidates: ['ServiceEndDate', 'Service_x0020_End_x0020_Date', 'EndDate'] },
+      { internalName: 'IsActive', type: 'Boolean', displayName: 'Is Active', default: true, candidates: ['IsActive', 'Active'] },
+      { internalName: 'UsageStatus', type: 'Text', displayName: 'Usage Status', candidates: ['UsageStatus', 'Usage_x0020_Status'] },
+      { internalName: 'IsHighIntensitySupportTarget', type: 'Boolean', displayName: 'High Intensity Target', candidates: ['IsHighIntensitySupportTarget', 'HighIntensityTarget'] },
+      { internalName: 'IsSupportProcedureTarget', type: 'Boolean', displayName: 'Support Procedure Target', candidates: ['IsSupportProcedureTarget', 'SupportProcedureTarget'] },
+      { internalName: 'severeFlag', type: 'Boolean', displayName: 'Severe Flag', candidates: ['severeFlag', 'HighSeverityFlag', 'cr013_severeFlag'] },
+      { internalName: 'TransportToDays', type: 'MultiChoice', displayName: 'Transport To Days (Legacy)', choices: ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'], governance: 'allow', isSilent: true, candidates: ['TransportToDays', 'Transport_x0020_To_x0020_Days', 'cr013_transportToDays'] },
+      { internalName: 'TransportFromDays', type: 'MultiChoice', displayName: 'Transport From Days (Legacy)', choices: ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'], governance: 'allow', isSilent: true, candidates: ['TransportFromDays', 'Transport_x0020_From_x0020_Days', 'cr013_transportFromDays'] },
+      { internalName: 'LastAssessmentDate', type: 'DateTime', displayName: 'Last Assessment Date', dateTimeFormat: 'DateOnly', candidates: ['LastAssessmentDate', 'Last_x0020_Assessment_x0020_Date'] },
+      { internalName: 'BehaviorScore', type: 'Number', displayName: 'Behavior Score', candidates: ['BehaviorScore', 'Behavior_x0020_Score'] },
+      { internalName: 'ChildBehaviorScore', type: 'Number', displayName: 'Child Behavior Score', candidates: ['ChildBehaviorScore', 'Child_x0020_Behavior_x0020_Score'] },
+      { internalName: 'ServiceTypesJson', type: 'Note', displayName: 'Service Types JSON', richText: false, candidates: ['ServiceTypesJson', 'Service_x0020_Types_x0020_JSON'] },
+      { internalName: 'EligibilityCheckedAt', type: 'DateTime', displayName: 'Eligibility Checked At', candidates: ['EligibilityCheckedAt', 'Eligibility_x0020_Checked_x0020_'] },
+      { internalName: 'Last_x0020_Assessment_x0020_Date', type: 'DateTime', displayName: 'Last Assessment Date (Legacy SP-encoded)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['Last_x0020_Assessment_x0020_Date'] },
+      { internalName: 'Behavior_x0020_Score', type: 'Number', displayName: 'Behavior Score (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Behavior_x0020_Score'] },
+      { internalName: 'Child_x0020_Behavior_x0020_Score', type: 'Number', displayName: 'Child Behavior Score (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Child_x0020_Behavior_x0020_Score'] },
+      { internalName: 'Service_x0020_Types_x0020_JSON', type: 'Note', displayName: 'Service Types JSON (Legacy SP-encoded)', richText: false, governance: 'allow', isSilent: true, candidates: ['Service_x0020_Types_x0020_JSON'] },
+      { internalName: 'Eligibility_x0020_Checked_x0020_', type: 'DateTime', displayName: 'Eligibility Checked At (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Eligibility_x0020_Checked_x0020_'] },
+      { internalName: 'GrantMunicipality', type: 'Text', displayName: 'Grant Municipality', isSilent: true, candidates: ['GrantMunicipality', 'GrantMunicipality0', 'Grant_x0020_Municipality'] },
+      { internalName: 'GrantPeriodStart', type: 'DateTime', displayName: 'Grant Period Start', dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['GrantPeriodStart', 'GrantPeriodStart0', 'Grant_x0020_Period_x0020_Start'] },
+      { internalName: 'GrantPeriodEnd', type: 'DateTime', displayName: 'Grant Period End', dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['GrantPeriodEnd', 'GrantPeriodEnd0', 'Grant_x0020_Period_x0020_End'] },
+      { internalName: 'UserCopayLimit', type: 'Text', displayName: 'User Copay Limit', isSilent: true, candidates: ['UserCopayLimit', 'UserCopayLimit0', 'User_x0020_Copay_x0020_Limit'] },
+      { internalName: 'MealAddition', type: 'Text', displayName: 'Meal Addition', isSilent: true, candidates: ['MealAddition', 'MealAddition0', 'Meal_x0020_Addition'] },
+      { internalName: 'CopayPaymentMethod', type: 'Text', displayName: 'Copay Payment Method', isSilent: true, candidates: ['CopayPaymentMethod', 'CopayPaymentMethod0', 'Copay_x0020_Payment_x0020_Method'] },
+      { internalName: 'AttendanceDays', type: 'Number', displayName: 'Attendance Days (Legacy)', governance: 'allow', isSilent: true },
+      { internalName: 'RecipientCertNumber', type: 'Text', displayName: 'Recipient Cert Number (Legacy)', governance: 'allow', isSilent: true },
+      { internalName: 'RecipientCertExpiry', type: 'DateTime', displayName: 'Recipient Cert Expiry (Legacy)', governance: 'allow', isSilent: true },
+      { internalName: 'email', type: 'Text', displayName: 'Email (Legacy)', governance: 'allow', isSilent: true },
+      { internalName: 'phone', type: 'Text', displayName: 'Phone (Legacy)', governance: 'allow', isSilent: true, candidates: ['phone', 'Phone'] },
+    ],
+  },
+  {
+    key: 'user_feature_flags',
+    displayName: 'ユーザー機能フラグ',
+    resolve: () => envOr('VITE_SP_LIST_USER_FLAGS', fromConfig(ListKeys.User_Feature_Flags)),
+    operations: ['R', 'W'],
+    category: 'master',
+    lifecycle: 'optional',
+    essentialFields: ['UserCode', 'FlagKey'],
+    provisioningFields: [
+      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true, candidates: ['UserCode', 'User_x0020_Code'] },
+      { internalName: 'FlagKey', type: 'Text', displayName: 'Flag Key', required: true, candidates: ['FlagKey', 'Flag_x0020_Key'] },
+      { internalName: 'FlagValue', type: 'Text', displayName: 'Flag Value', candidates: ['FlagValue', 'Flag_x0020_Value'] },
+      { internalName: 'ExpiresAt', type: 'DateTime', displayName: 'Expires At', dateTimeFormat: 'DateTime', candidates: ['ExpiresAt', 'Expires_x0020_At'] },
+    ],
+  },
+  {
+    key: 'user_transport_settings',
+    displayName: '利用者送迎設定',
+    resolve: () => envOr('VITE_SP_LIST_USER_TRANSPORT', fromConfig(ListKeys.UserTransportSettings)),
+    operations: ['R', 'W'],
+    category: 'master',
+    lifecycle: 'required',
+    essentialFields: [
+      'UserID', 'TransportToDays', 'TransportFromDays', 'TransportAdditionType'
+    ],
+    provisioningFields: [
+      { internalName: 'UserID', type: 'Text', displayName: 'User ID', required: true, indexed: true, candidates: ['UserID'] },
+      { internalName: 'TransportToDays', type: 'MultiChoice', displayName: 'Transport To Days', choices: ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'], candidates: ['TransportToDays'] },
+      { internalName: 'TransportFromDays', type: 'MultiChoice', displayName: 'Transport From Days', choices: ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'], candidates: ['TransportFromDays'] },
+      { internalName: 'TransportCourse', type: 'Text', displayName: 'Transport Course', candidates: ['TransportCourse'] },
+      { internalName: 'TransportSchedule', type: 'Note', displayName: 'Transport Schedule', richText: false, candidates: ['TransportSchedule'] },
+      { internalName: 'TransportAdditionType', type: 'Text', displayName: 'Transport Addition Type', required: true, candidates: ['TransportAdditionType', 'Transport_x0020_Addition_x0020_Type', 'Transport_x0020_Addition_x0020_T', 'cr013_transportAdditionType', 'cr013_transportadditiontype'] },
+    ],
+  },
+  {
+    key: 'user_benefit_profile',
+    displayName: '利用者支給量プロファイル',
+    resolve: () => envOr('VITE_SP_LIST_USER_BENEFIT', fromConfig(ListKeys.UserBenefitProfile)),
+    operations: ['R', 'W'],
+    category: 'master',
+    lifecycle: 'required',
+    essentialFields: [
+      'UserID'
+    ],
+    provisioningFields: [
+      { internalName: 'UserID', type: 'Text', displayName: 'User ID', required: true, indexed: true, isSilent: true, candidates: ['UserID', 'Title'] },
+      // RecipientCertNumber moved to _Ext to avoid 8KB limit
+      { internalName: 'RecipientCertExpiry', type: 'DateTime', displayName: 'Recipient Cert Expiry', dateTimeFormat: 'DateOnly', isSilent: true },
+      { internalName: 'GrantMunicipality', type: 'Text', displayName: 'Grant Municipality', isSilent: true, candidates: ['GrantMunicipality', 'Grant_x0020_Municipality'] },
+      { internalName: 'GrantPeriodStart', type: 'DateTime', displayName: 'Grant Period Start', dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['GrantPeriodStart', 'Grant_x0020_Period_x0020_Start'] },
+      { internalName: 'GrantPeriodEnd', type: 'DateTime', displayName: 'Grant Period End', dateTimeFormat: 'DateOnly', indexed: true, isSilent: true, candidates: ['GrantPeriodEnd', 'Grant_x0020_Period_x0020_End'] },
+      { internalName: 'DisabilitySupportLevel', type: 'Text', displayName: 'Disability Support Level', isSilent: true },
+      { internalName: 'GrantedDaysPerMonth', type: 'Text', displayName: 'Granted Days Per Month', isSilent: true },
+      { internalName: 'UserCopayLimit', type: 'Text', displayName: 'User Copay Limit', isSilent: true, candidates: ['UserCopayLimit', 'User_x0020_Copay_x0020_Limit'] },
+      { internalName: 'MealAddition', type: 'Text', displayName: 'Meal Addition', isSilent: true, candidates: ['MealAddition', 'Meal_x0020_Addition'] },
+      { internalName: 'CopayPaymentMethod', type: 'Text', displayName: 'Copay Payment Method', isSilent: true, candidates: ['CopayPaymentMethod', 'Copay_x0020_Payment_x0020_Method'] },
+    ],
+  },
+  {
+    key: 'user_benefit_profile_ext',
+    displayName: '利用者支給量プロファイル (拡充)',
+    resolve: () => envOr('VITE_SP_LIST_USER_BENEFIT_EXT', fromConfig(ListKeys.UserBenefitProfileExt)),
+    operations: ['R', 'W'],
+    category: 'master',
+    lifecycle: 'required',
+    essentialFields: [
+      'UserID', 'RecipientCertNumber'
+    ],
+    provisioningFields: [
+      { internalName: 'UserID', type: 'Text', displayName: 'User ID', required: true, indexed: true, candidates: ['UserID', 'Title'] },
+      { internalName: 'RecipientCertNumber', type: 'Text', displayName: 'Recipient Cert Number', required: true, isSilent: true, candidates: ['RecipientCertNumber', 'RecipientCertNumber0', 'Recipient_x0020_Cert_x0020_Numbe'] },
+    ],
+  },
+  {
+    key: 'staff_master',
+    displayName: '職員マスタ',
+    resolve: () => envOr('VITE_SP_LIST_STAFF', fromConfig(ListKeys.StaffMaster)),
+    operations: ['R'],
+    category: 'master',
+    lifecycle: 'required',
+    essentialFields: [
+      'StaffID', 'FullName', 'Role', 'RBACRole', 'IsActive', 'Department'
+    ],
+    provisioningFields: [
+      { internalName: 'StaffID', type: 'Text', displayName: 'Staff ID', required: true, candidates: ['StaffID', 'Staff_x0020_ID', 'StaffCode'] },
+      { internalName: 'FullName', type: 'Text', displayName: 'Full Name', required: true, candidates: ['FullName', 'Full_x0020_Name', 'Title'] },
+      { internalName: 'Furigana', type: 'Text', displayName: 'Furigana' },
+      { internalName: 'FullNameKana', type: 'Text', displayName: 'Full Name Kana', candidates: ['FullNameKana', 'Full_x0020_Name_x0020_Kana'] },
+      { internalName: 'JobTitle', type: 'Text', displayName: 'Job Title', candidates: ['JobTitle', 'Job_x0020_Title'] },
+      { internalName: 'EmploymentType', type: 'Text', displayName: 'Employment Type', candidates: ['EmploymentType', 'Employment_x0020_Type'] },
+      { internalName: 'RBACRole', type: 'Text', displayName: 'RBAC Role', candidates: ['RBACRole', 'RBAC_x0020_Role'] },
+      { internalName: 'Role', type: 'Text', displayName: 'Role' },
+      { internalName: 'IsActive', type: 'Boolean', displayName: 'Is Active', default: true, candidates: ['IsActive', 'Active'] },
+      { internalName: 'Department', type: 'Text', displayName: 'Department' },
+      { internalName: 'Work_x0020_Days', type: 'Text', displayName: 'Work Days (Legacy SP-encoded)', isSilent: true, governance: 'allow' },
+      { internalName: 'WorkDays', type: 'MultiChoice', displayName: 'Work Days', choices: ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'] },
+      { internalName: 'BaseWorkingDays', type: 'Number', displayName: 'Base Working Days', governance: 'allow', isSilent: true, candidates: ['BaseWorkingDays', 'BaseWorkDays'] },
+      { internalName: 'StaffCode', type: 'Text', displayName: 'Staff Code (Legacy)', governance: 'allow', isSilent: true, candidates: ['StaffCode'] },
+      { internalName: 'ActiveFlag', type: 'Boolean', displayName: 'Active Flag (Legacy)', governance: 'allow', isSilent: true, candidates: ['ActiveFlag'] },
+      { internalName: 'BaseShiftStartTime', type: 'Text', displayName: 'Base Shift Start' },
+      { internalName: 'BaseShiftEndTime', type: 'Text', displayName: 'Base Shift End' },
+      { internalName: 'HireDate', type: 'DateTime', displayName: 'Hire Date', dateTimeFormat: 'DateOnly' },
+      { internalName: 'ResignDate', type: 'DateTime', displayName: 'Resign Date', dateTimeFormat: 'DateOnly' },
+      { internalName: 'Email', type: 'Text', displayName: 'Email' },
+      { internalName: 'Phone', type: 'Text', displayName: 'Phone' },
+      { internalName: 'Certifications', type: 'Note', displayName: 'Certifications' },
+    ],
+  },
+  {
+    key: 'org_master',
+    displayName: '組織マスタ',
+    resolve: () => envOr('VITE_SP_LIST_ORG_MASTER', fromConfig(ListKeys.OrgMaster)),
+    operations: ['R'],
+    category: 'master',
+    lifecycle: 'required',
+    essentialFields: ['OrgCode', 'OrgType', 'Audience'],
+    provisioningFields: [
+      { internalName: 'OrgCode', type: 'Text', displayName: 'Organization Code', required: true },
+      { internalName: 'OrgType', type: 'Text', displayName: 'Organization Type' },
+      { internalName: 'Audience', type: 'Text', displayName: 'Audience' },
+      { internalName: 'SortOrder', type: 'Number', displayName: 'Sort Order' },
+      { internalName: 'IsActive', type: 'Boolean', displayName: 'Active', default: true },
+      { internalName: 'Notes', type: 'Note', displayName: 'Notes' },
+    ],
+  },
+  {
+    key: 'holiday_master',
+    displayName: '祝日・休業日マスタ',
+    resolve: () => envOr('VITE_SP_LIST_HOLIDAY_MASTER', fromConfig(ListKeys.HolidayMaster)),
+    operations: ['R'],
+    category: 'master',
+    lifecycle: 'required',
+    essentialFields: ['Date', 'Type'],
+    provisioningFields: [
+      { internalName: 'Date', type: 'DateTime', displayName: 'Date', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['Date', 'HolidayDate', 'TargetDate', 'RecordDate'] },
+      { internalName: 'Label', type: 'Text', displayName: 'Label', candidates: ['Label', 'Title', 'Name'] },
+      { internalName: 'Type', type: 'Text', displayName: 'Type', candidates: ['Type', 'Category', 'HolidayType'] },
+      { internalName: 'FiscalYear', type: 'Text', displayName: 'Fiscal Year', candidates: ['FiscalYear', 'Year', 'BusinessYear'] },
+      { internalName: 'IsActive', type: 'Boolean', displayName: 'Is Active', default: true, candidates: ['IsActive', 'Active', 'Enabled'] },
+    ],
+  },
+];
