+++ src/sharepoint/definitions/meeting.ts
@@ -1,0 +1,125 @@
+import { ListKeys } from '@/sharepoint/fields';
+import { envOr, fromConfig } from '../spListRegistry.shared';
+import type { SpListEntry } from '../spListRegistry.shared';
+
+export const meetingListEntries: readonly SpListEntry[] = [
+// ── 5. 会議系 ──────────────────────────────────────────
+  {
+    key: 'meeting_sessions',
+    displayName: '会議セッション',
+    resolve: () => envOr('VITE_SP_LIST_MEETING_SESSIONS', 'MeetingSessions'),
+    operations: ['R', 'W', 'D'],
+    category: 'meeting',
+    lifecycle: 'optional',
+    essentialFields: ['SessionKey', 'MeetingKind', 'MeetingDate'],
+    provisioningFields: [
+      { internalName: 'SessionKey', type: 'Text', displayName: 'Session Key', required: true, indexed: true, candidates: ['SessionKey', 'Session_x0020_Key'] },
+      { internalName: 'MeetingKind', type: 'Choice', displayName: 'Meeting Kind', choices: ['morning', 'evening'], required: true, candidates: ['MeetingKind', 'Meeting_x0020_Kind'] },
+      { internalName: 'MeetingDate', type: 'DateTime', displayName: 'Meeting Date', required: true, dateTimeFormat: 'DateOnly', candidates: ['MeetingDate', 'Meeting_x0020_Date'] },
+      { internalName: 'StartTime', type: 'Text', displayName: 'Start Time', candidates: ['StartTime', 'Start_x0020_Time'] },
+      { internalName: 'EndTime', type: 'Text', displayName: 'End Time', candidates: ['EndTime', 'End_x0020_Time'] },
+      { internalName: 'Session_x0020_Key', type: 'Text', displayName: 'Session Key (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Session_x0020_Key'] },
+      { internalName: 'Meeting_x0020_Kind', type: 'Text', displayName: 'Meeting Kind (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Meeting_x0020_Kind'] },
+      { internalName: 'Meeting_x0020_Date', type: 'DateTime', displayName: 'Meeting Date (Legacy SP-encoded)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['Meeting_x0020_Date'] },
+      { internalName: 'Start_x0020_Time', type: 'Text', displayName: 'Start Time (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Start_x0020_Time'] },
+      { internalName: 'End_x0020_Time', type: 'Text', displayName: 'End Time (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['End_x0020_Time'] },
+      { internalName: 'ChairpersonUserId', type: 'Text', displayName: 'Chairperson ID', isSilent: true, candidates: ['ChairpersonUserId', 'Chairperson_x0020_ID'] },
+      { internalName: 'ChairpersonName', type: 'Text', displayName: 'Chairperson Name', candidates: ['ChairpersonName', 'Chairperson_x0020_Name'] },
+      { internalName: 'Chairperson_x0020_Name', type: 'Text', displayName: 'Chairperson Name (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Chairperson_x0020_Name'] },
+      { internalName: 'Status', type: 'Choice', displayName: 'Status', choices: ['scheduled', 'in-progress', 'completed', 'cancelled'] },
+      { internalName: 'TotalParticipants', type: 'Number', displayName: 'Total Participants', candidates: ['TotalParticipants', 'Total_x0020_Participants'] },
+      { internalName: 'Total_x0020_Participants', type: 'Number', displayName: 'Total Participants (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Total_x0020_Participants'] },
+    ],
+  },
+  {
+    key: 'meeting_steps',
+    displayName: '会議ステップ',
+    resolve: () => envOr('VITE_SP_LIST_MEETING_STEPS', 'MeetingSteps'),
+    operations: ['R', 'W', 'D'],
+    category: 'meeting',
+    lifecycle: 'optional',
+    essentialFields: ['SessionId', 'StepId', 'StepTitle'],
+    provisioningFields: [
+      { internalName: 'SessionId', type: 'Number', displayName: 'Session ID', required: true, indexed: true, isSilent: true, candidates: ['SessionId', 'Session_x0020_ID'] },
+      { internalName: 'SessionKey', type: 'Text', displayName: 'Session Key', required: true, isSilent: true, candidates: ['SessionKey', 'Session_x0020_Key'] },
+      { internalName: 'StepId', type: 'Number', displayName: 'Step ID', required: true, isSilent: true, candidates: ['StepId', 'Step_x0020_ID'] },
+      { internalName: 'StepTitle', type: 'Text', displayName: 'Step Title', required: true, isSilent: true, candidates: ['StepTitle', 'Step_x0020_Title'] },
+      { internalName: 'Completed', type: 'Boolean', displayName: 'Completed', default: false },
+      { internalName: 'TimeSpentMinutes', type: 'Number', displayName: 'Minutes Spent', isSilent: true, candidates: ['TimeSpentMinutes', 'Minutes_x0020_Spent'] },
+      { internalName: 'StepNotes', type: 'Note', displayName: 'Step Notes', richText: false, isSilent: true, candidates: ['StepNotes', 'Step_x0020_Notes'] },
+      { internalName: 'Session_x0020_ID', type: 'Number', displayName: 'Session ID (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Session_x0020_ID'] },
+      { internalName: 'Session_x0020_Key', type: 'Text', displayName: 'Session Key (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Session_x0020_Key'] },
+      { internalName: 'Step_x0020_ID', type: 'Number', displayName: 'Step ID (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Step_x0020_ID'] },
+      { internalName: 'Step_x0020_Title', type: 'Text', displayName: 'Step Title (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Step_x0020_Title'] },
+      { internalName: 'Step_x0020_Notes', type: 'Note', displayName: 'Step Notes (Legacy SP-encoded)', richText: false, governance: 'allow', isSilent: true, candidates: ['Step_x0020_Notes'] },
+    ],
+  },
+  {
+    key: 'meeting_minutes',
+    displayName: '議事録',
+    resolve: () => envOr('VITE_SP_LIST_MEETING_MINUTES', fromConfig(ListKeys.MeetingMinutes)),
+    operations: ['R', 'W', 'D'],
+    category: 'meeting',
+    lifecycle: 'required',
+    essentialFields: ['MeetingDate', 'Category'],
+    provisioningFields: [
+      { internalName: 'MeetingDate', type: 'DateTime', displayName: 'Meeting Date', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['MeetingDate', 'meetingDate', 'cr013_meetingDate', 'Date'] },
+      { internalName: 'Category', type: 'Text', displayName: 'Category', required: true, candidates: ['Category', 'category', 'cr013_category', 'MeetingCategory'] },
+      { internalName: 'Summary', type: 'Note', displayName: 'Summary', candidates: ['Summary', 'summary', 'cr013_summary', 'Notes'] },
+      { internalName: 'Decisions', type: 'Note', displayName: 'Decisions', candidates: ['Decisions', 'decisions', 'cr013_decisions', 'DecisionSummary'] },
+      { internalName: 'Actions', type: 'Note', displayName: 'Actions', candidates: ['Actions', 'actions', 'cr013_actions', 'ActionItems'] },
+      { internalName: 'Attendees', type: 'Note', displayName: 'Attendees', candidates: ['Attendees', 'attendees', 'cr013_attendees'] },
+      { internalName: 'StaffAttendance', type: 'Note', displayName: 'Staff Attendance', candidates: ['StaffAttendance', 'staffAttendance', 'cr013_staffAttendance'] },
+      { internalName: 'Tags', type: 'Text', displayName: 'Tags (Legacy)', governance: 'allow', isSilent: true, candidates: ['Tags'] },
+      { internalName: 'UserHealthNotes', type: 'Note', displayName: 'User Health Notes (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['UserHealthNotes'] },
+      { internalName: 'RelatedLinks', type: 'Note', displayName: 'Related Links (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['RelatedLinks'] },
+      { internalName: 'Chair', type: 'Text', displayName: 'Chair (Legacy)', governance: 'allow', isSilent: true, candidates: ['Chair'] },
+      { internalName: 'Scribe', type: 'Text', displayName: 'Scribe (Legacy)', governance: 'allow', isSilent: true, candidates: ['Scribe'] },
+      { internalName: 'IsPublished', type: 'Boolean', displayName: 'Published', default: false, candidates: ['IsPublished', 'isPublished', 'cr013_isPublished', 'Published'] },
+    ],
+  },
+  {
+    key: 'monitoring_meetings',
+    displayName: 'モニタリング会議',
+    resolve: () => envOr('VITE_SP_LIST_MONITORING_MEETINGS', 'MonitoringMeetings'),
+    operations: ['R', 'W', 'D'],
+    category: 'meeting',
+    lifecycle: 'required',
+    essentialFields: ['cr014_recordId', 'cr014_userId', 'cr014_meetingDate'],
+    provisioningFields: [
+      { internalName: 'cr014_recordId', type: 'Text', displayName: 'Record ID', required: true, indexed: true, candidates: ['cr014_recordId', 'RecordId', 'recordId', 'Record_x0020_Id', 'Title'] },
+      { internalName: 'cr014_userId', type: 'Text', displayName: 'User ID', required: true, indexed: true, candidates: ['cr014_userId', 'UserId', 'UserID', 'UserCode', 'userId', 'User_x0020_Id', 'cr013_userId'] },
+      { internalName: 'cr014_ispId', type: 'Text', displayName: 'ISP ID', required: true, candidates: ['cr014_ispId', 'IspId', 'ISPId', 'ispId'] },
+      { internalName: 'cr014_planningSheetId', type: 'Text', displayName: 'Planning Sheet ID', candidates: ['cr014_planningSheetId', 'PlanningSheetId', 'planningSheetId'] },
+      { internalName: 'cr014_meetingType', type: 'Text', displayName: 'Meeting Type', candidates: ['cr014_meetingType', 'MeetingType', 'meetingType'] },
+      { internalName: 'cr014_status', type: 'Text', displayName: 'Status', candidates: ['cr014_status', 'Status', 'status'] },
+      { internalName: 'cr014_meetingDate', type: 'DateTime', displayName: 'Meeting Date', required: true, dateTimeFormat: 'DateOnly', candidates: ['cr014_meetingDate', 'MeetingDate', 'Meeting_x0020_Date', 'meetingDate', 'Date', 'RecordDate', 'cr013_date'] },
+      { internalName: 'cr014_venue', type: 'Text', displayName: 'Venue', candidates: ['cr014_venue', 'Venue', 'venue'] },
+      { internalName: 'cr014_attendeesJson', type: 'Note', displayName: 'Attendees (JSON)', richText: false, candidates: ['cr014_attendeesJson', 'AttendeesJson', 'Attendees', 'attendeesJson'] },
+      { internalName: 'cr014_goalEvaluationsJson', type: 'Note', displayName: 'Goal Evaluations (JSON)', richText: false, candidates: ['cr014_goalEvaluationsJson', 'GoalEvaluationsJson', 'GoalEvaluations', 'goalEvaluationsJson'] },
+      { internalName: 'cr014_overallAssessment', type: 'Note', displayName: 'Overall Assessment', richText: false, candidates: ['cr014_overallAssessment', 'OverallAssessment', 'overallAssessment'] },
+      { internalName: 'cr014_userFeedback', type: 'Note', displayName: 'User Feedback', richText: false, candidates: ['cr014_userFeedback', 'UserFeedback', 'userFeedback'] },
+      { internalName: 'cr014_familyFeedback', type: 'Note', displayName: 'Family Feedback', richText: false, candidates: ['cr014_familyFeedback', 'FamilyFeedback', 'familyFeedback'] },
+      { internalName: 'cr014_planChangeDecision', type: 'Text', displayName: 'Plan Change Decision', candidates: ['cr014_planChangeDecision', 'PlanChangeDecision', 'planChangeDecision'] },
+      { internalName: 'cr014_changeReason', type: 'Note', displayName: 'Change Reason', richText: false, candidates: ['cr014_changeReason', 'ChangeReason', 'changeReason'] },
+      { internalName: 'cr014_decisionsJson', type: 'Note', displayName: 'Decisions (JSON)', richText: false, candidates: ['cr014_decisionsJson', 'DecisionsJson', 'Decisions', 'decisionsJson'] },
+      { internalName: 'cr014_nextMonitoringDate', type: 'DateTime', displayName: 'Next Monitoring Date', dateTimeFormat: 'DateOnly', candidates: ['cr014_nextMonitoringDate', 'NextMonitoringDate', 'Next_x0020_Monitoring_x0020_Date', 'nextMonitoringDate'] },
+      { internalName: 'cr014_recordedBy', type: 'Text', displayName: 'Recorded By', candidates: ['cr014_recordedBy', 'RecordedBy', 'recordedBy'] },
+      { internalName: 'cr014_recordedAt', type: 'DateTime', displayName: 'Recorded At', candidates: ['cr014_recordedAt', 'RecordedAt', 'recordedAt'] },
+      { internalName: 'cr014_implSummary', type: 'Note', displayName: 'Implementation Summary (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['cr014_implSummary'] },
+      { internalName: 'cr014_behaviorChange', type: 'Note', displayName: 'Behavior Change (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['cr014_behaviorChange'] },
+      { internalName: 'cr014_effSupport', type: 'Note', displayName: 'Effective Support (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['cr014_effSupport'] },
+      { internalName: 'cr014_issueSummary', type: 'Note', displayName: 'Issue Summary (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['cr014_issueSummary'] },
+      { internalName: 'cr014_discussionSummary', type: 'Note', displayName: 'Discussion Summary (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['cr014_discussionSummary'] },
+      { internalName: 'cr014_reqPlanSheetUpd', type: 'Boolean', displayName: 'Request Plan Sheet Update (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_reqPlanSheetUpd'] },
+      { internalName: 'cr014_reqIspUpd', type: 'Boolean', displayName: 'Request ISP Update (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_reqIspUpd'] },
+      { internalName: 'cr014_hasBasicTrained', type: 'Boolean', displayName: 'Has Basic Trained (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_hasBasicTrained'] },
+      { internalName: 'cr014_hasPractTrained', type: 'Boolean', displayName: 'Has Practical Trained (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_hasPractTrained'] },
+      { internalName: 'cr014_qualCheckStatus', type: 'Text', displayName: 'Qualification Check Status (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_qualCheckStatus'] },
+      { internalName: 'cr014_nextActions', type: 'Note', displayName: 'Next Actions (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['cr014_nextActions'] },
+      { internalName: 'cr014_finalizedAt', type: 'DateTime', displayName: 'Finalized At (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_finalizedAt'] },
+      { internalName: 'cr014_finalizedBy', type: 'Text', displayName: 'Finalized By (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_finalizedBy'] },
+      { internalName: 'cr014_prevMeetingId', type: 'Text', displayName: 'Previous Meeting ID (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_prevMeetingId'] },
+    ],
+  },
+];
