+++ src/sharepoint/definitions/other.ts
@@ -1,0 +1,215 @@
+import { ListKeys } from '@/sharepoint/fields';
+import { envOr, fromConfig } from '../spListRegistry.shared';
+import type { SpListEntry } from '../spListRegistry.shared';
+
+export const otherListEntries: readonly SpListEntry[] = [
+// ── 8. その他 ─────────────────────────────────────────
+  {
+    key: 'survey_tokusei',
+    displayName: '特性アンケート',
+    resolve: () => envOr('VITE_SP_LIST_SURVEY_TOKUSEI', fromConfig(ListKeys.SurveyTokusei)),
+    operations: ['R'],
+    category: 'other',
+    lifecycle: 'optional',
+    provisioningFields: [
+      { internalName: 'Notes', type: 'Note', displayName: 'Notes', richText: false },
+      { internalName: 'StartTime', type: 'DateTime', displayName: 'Start Time' },
+      { internalName: 'EndTime', type: 'DateTime', displayName: 'End Time' },
+      { internalName: 'ResponderEmail', type: 'Text', displayName: 'Responder Email', candidates: ['ResponderEmail', 'Responder_x0020_Email'] },
+      { internalName: 'ResponderName', type: 'Text', displayName: 'Responder Name', candidates: ['ResponderName', 'Responder_x0020_Name'] },
+      { internalName: 'FillDate', type: 'DateTime', displayName: 'Fill Date', candidates: ['FillDate', 'Fill_x0020_Date'] },
+      { internalName: 'TargetUserName', type: 'Text', displayName: 'Target User Name', candidates: ['TargetUserName', 'Target_x0020_User_x0020_Name'] },
+      { internalName: 'FormRowId', type: 'Text', displayName: 'Form Row ID', candidates: ['FormRowId', 'Form_x0020_Row_x0020_Id'] },
+      { internalName: 'Strengths', type: 'Note', displayName: 'Strengths', richText: false, candidates: ['Strengths', 'UserStrengths'] },
+      { internalName: 'RelationalDifficulties', type: 'Note', displayName: 'Relational Difficulties', richText: false, candidates: ['RelationalDifficulties', 'Relational_x0020_Difficulties'] },
+      { internalName: 'SituationalUnderstanding', type: 'Note', displayName: 'Situational Understanding', richText: false, candidates: ['SituationalUnderstanding', 'Situational_x0020_Understanding'] },
+      { internalName: 'Hearing', type: 'Note', displayName: 'Hearing', richText: false, candidates: ['Hearing', 'Sensory_x0020_Hearing'] },
+      { internalName: 'Vision', type: 'Note', displayName: 'Vision', richText: false, candidates: ['Vision', 'Sensory_x0020_Vision'] },
+      { internalName: 'Touch', type: 'Note', displayName: 'Touch', richText: false, candidates: ['Touch', 'Sensory_x0020_Touch'] },
+      { internalName: 'Smell', type: 'Note', displayName: 'Smell', richText: false, candidates: ['Smell', 'Sensory_x0020_Smell'] },
+      { internalName: 'Taste', type: 'Note', displayName: 'Taste', richText: false, candidates: ['Taste', 'Sensory_x0020_Taste'] },
+      { internalName: 'SensoryMultiSelect', type: 'Text', displayName: 'Sensory Multi Select', candidates: ['SensoryMultiSelect', 'Sensory_x0020_Multiple'] },
+      { internalName: 'SensoryFreeText', type: 'Note', displayName: 'Sensory Free Text', richText: false, candidates: ['SensoryFreeText', 'Sensory_x0020_Notes'] },
+      { internalName: 'DifficultyWithChanges', type: 'Note', displayName: 'Difficulty With Changes', richText: false, candidates: ['DifficultyWithChanges', 'Change_x0020_Difficulties'] },
+      { internalName: 'InterestInParts', type: 'Note', displayName: 'Interest In Parts', richText: false, candidates: ['InterestInParts', 'Part_x0020_Interests'] },
+      { internalName: 'RepetitiveBehaviors', type: 'Note', displayName: 'Repetitive Behaviors', richText: false, candidates: ['RepetitiveBehaviors', 'Repetitive_x0020_Actions'] },
+      { internalName: 'FixedHabits', type: 'Note', displayName: 'Fixed Habits', richText: false, candidates: ['FixedHabits', 'Fixed_x0020_Routines'] },
+      { internalName: 'ComprehensionDifficulty', type: 'Note', displayName: 'Comprehension Difficulty', richText: false, candidates: ['ComprehensionDifficulty', 'Comprehension_x0020_Issues'] },
+      { internalName: 'ExpressionDifficulty', type: 'Note', displayName: 'Expression Difficulty', richText: false, candidates: ['ExpressionDifficulty', 'Expression_x0020_Issues'] },
+      { internalName: 'InteractionDifficulty', type: 'Note', displayName: 'Interaction Difficulty', richText: false, candidates: ['InteractionDifficulty', 'Interaction_x0020_Issues'] },
+      { internalName: 'BehaviorMultiSelect', type: 'Text', displayName: 'Behavior Multi Select', candidates: ['BehaviorMultiSelect', 'Behavior_x0020_Multiple'] },
+      { internalName: 'BehaviorEpisodes', type: 'Note', displayName: 'Behavior Episodes', richText: false, candidates: ['BehaviorEpisodes', 'Behavior_x0020_Details'] },
+    ],
+  },
+  {
+    key: 'survey_tokusei_v2',
+    displayName: '特性確認シート',
+    resolve: () => envOr('VITE_SP_LIST_SURVEY_TOKUSEI_V2', fromConfig(ListKeys.SurveyTokuseiV2)),
+    operations: ['R'],
+    category: 'other',
+    lifecycle: 'optional',
+    provisioningFields: [
+      { internalName: 'Notes', type: 'Note', displayName: 'Notes', richText: false },
+      { internalName: 'StartTime', type: 'DateTime', displayName: 'Start Time' },
+      { internalName: 'EndTime', type: 'DateTime', displayName: 'End Time' },
+      { internalName: 'ResponderEmail', type: 'Text', displayName: 'Responder Email', candidates: ['ResponderEmail', 'Responder_x0020_Email'] },
+      { internalName: 'ResponderName', type: 'Text', displayName: 'Responder Name', candidates: ['ResponderName', 'Responder_x0020_Name'] },
+      { internalName: 'FillDate', type: 'DateTime', displayName: 'Fill Date', candidates: ['FillDate', 'Fill_x0020_Date'] },
+      { internalName: 'TargetUserName', type: 'Text', displayName: 'Target User Name', candidates: ['TargetUserName', 'Target_x0020_User_x0020_Name'] },
+      { internalName: 'FormRowId', type: 'Text', displayName: 'Form Row ID', candidates: ['FormRowId', 'Form_x0020_Row_x0020_Id'] },
+      { internalName: 'Strengths', type: 'Note', displayName: 'Strengths', richText: false, candidates: ['Strengths', 'UserStrengths'] },
+      { internalName: 'RelationalDifficulties', type: 'Note', displayName: 'Relational Difficulties', richText: false, candidates: ['RelationalDifficulties', 'Relational_x0020_Difficulties'] },
+      { internalName: 'SituationalUnderstanding', type: 'Note', displayName: 'Situational Understanding', richText: false, candidates: ['SituationalUnderstanding', 'Situational_x0020_Understanding'] },
+      { internalName: 'Hearing', type: 'Note', displayName: 'Hearing', richText: false, candidates: ['Hearing', 'Sensory_x0020_Hearing'] },
+      { internalName: 'Vision', type: 'Note', displayName: 'Vision', richText: false, candidates: ['Vision', 'Sensory_x0020_Vision'] },
+      { internalName: 'Touch', type: 'Note', displayName: 'Touch', richText: false, candidates: ['Touch', 'Sensory_x0020_Touch'] },
+      { internalName: 'Smell', type: 'Note', displayName: 'Smell', richText: false, candidates: ['Smell', 'Sensory_x0020_Smell'] },
+      { internalName: 'Taste', type: 'Note', displayName: 'Taste', richText: false, candidates: ['Taste', 'Sensory_x0020_Taste'] },
+      { internalName: 'SensoryMultiSelect', type: 'Text', displayName: 'Sensory Multi Select', candidates: ['SensoryMultiSelect', 'Sensory_x0020_Multiple'] },
+      { internalName: 'SensoryFreeText', type: 'Note', displayName: 'Sensory Free Text', richText: false, candidates: ['SensoryFreeText', 'Sensory_x0020_Notes'] },
+      { internalName: 'DifficultyWithChanges', type: 'Note', displayName: 'Difficulty With Changes', richText: false, candidates: ['DifficultyWithChanges', 'Change_x0020_Difficulties'] },
+      { internalName: 'InterestInParts', type: 'Note', displayName: 'Interest In Parts', richText: false, candidates: ['InterestInParts', 'Part_x0020_Interests'] },
+      { internalName: 'RepetitiveBehaviors', type: 'Note', displayName: 'Repetitive Behaviors', richText: false, candidates: ['RepetitiveBehaviors', 'Repetitive_x0020_Actions'] },
+      { internalName: 'FixedHabits', type: 'Note', displayName: 'Fixed Habits', richText: false, candidates: ['FixedHabits', 'Fixed_x0020_Routines'] },
+      { internalName: 'ComprehensionDifficulty', type: 'Note', displayName: 'Comprehension Difficulty', richText: false, candidates: ['ComprehensionDifficulty', 'Comprehension_x0020_Issues'] },
+      { internalName: 'ExpressionDifficulty', type: 'Note', displayName: 'Expression Difficulty', richText: false, candidates: ['ExpressionDifficulty', 'Expression_x0020_Issues'] },
+      { internalName: 'InteractionDifficulty', type: 'Note', displayName: 'Interaction Difficulty', richText: false, candidates: ['InteractionDifficulty', 'Interaction_x0020_Issues'] },
+      { internalName: 'BehaviorMultiSelect', type: 'Text', displayName: 'Behavior Multi Select', candidates: ['BehaviorMultiSelect', 'Behavior_x0020_Multiple'] },
+      { internalName: 'BehaviorEpisodes', type: 'Note', displayName: 'Behavior Episodes', richText: false, candidates: ['BehaviorEpisodes', 'Behavior_x0020_Details'] },
+    ],
+  },
+  {
+    key: 'nurse_observations',
+    displayName: '看護観察記録',
+    resolve: () => envOr('VITE_SP_LIST_NURSE_OBSERVATION', fromConfig(ListKeys.NurseObservations)),
+    operations: ['R', 'W'],
+    category: 'other',
+    lifecycle: 'optional',
+    essentialFields: ['UserLookupId', 'ObservedAt', 'Temperature'],
+    provisioningFields: [
+      { internalName: 'UserLookupId', type: 'Number', displayName: 'User Lookup ID', required: true, isSilent: true, candidates: ['UserLookupId', 'User_x0020_Lookup_x0020_ID'] },
+      { internalName: 'ObservedAt', type: 'DateTime', displayName: 'Observed At', required: true, dateTimeFormat: 'DateTime', isSilent: true, candidates: ['ObservedAt', 'Observed_x0020_At', 'Created'] },
+      { internalName: 'Temperature', type: 'Number', displayName: 'Temperature', required: true, candidates: ['Temperature', 'Temperature_x0020_Value'] },
+      { internalName: 'Pulse', type: 'Number', displayName: 'Pulse', candidates: ['Pulse'] },
+      { internalName: 'Systolic', type: 'Number', displayName: 'Systolic BP', isSilent: true, candidates: ['Systolic', 'Systolic_x0020_BP'] },
+      { internalName: 'Diastolic', type: 'Number', displayName: 'Diastolic BP', isSilent: true, candidates: ['Diastolic', 'Diastolic_x0020_BP'] },
+      { internalName: 'SpO2', type: 'Number', displayName: 'SpO2', isSilent: true, candidates: ['SpO2', 'SPO2', '_x0053_pO2'] },
+      { internalName: 'Weight', type: 'Number', displayName: 'Weight', candidates: ['Weight'] },
+      { internalName: 'Memo', type: 'Note', displayName: 'Memo', richText: false, candidates: ['Memo', 'Notes'] },
+      { internalName: 'Tags', type: 'Text', displayName: 'Tags', candidates: ['Tags'] },
+      { internalName: 'IdempotencyKey', type: 'Text', displayName: 'Idempotency Key', isSilent: true, candidates: ['IdempotencyKey', 'Idempotency_x0020_Key'] },
+      { internalName: 'Source', type: 'Text', displayName: 'Source', candidates: ['Source'] },
+      { internalName: 'LocalTimeZone', type: 'Text', displayName: 'Local Time Zone', isSilent: true, candidates: ['LocalTimeZone', 'Local_x0020_Time_x0020_Zone'] },
+      { internalName: 'CreatedBy', type: 'Text', displayName: 'Created By', isSilent: true, candidates: ['CreatedBy', 'Created_x0020_By'] },
+      { internalName: 'DeviceId', type: 'Text', displayName: 'Device ID', isSilent: true, candidates: ['DeviceId', 'Device_x0020_ID'] },
+      { internalName: 'VitalsJson', type: 'Note', displayName: 'Vitals JSON', richText: false, isSilent: true, candidates: ['VitalsJson', 'Vitals_x0020_JSON'] },
+    ],
+  },
+  {
+    key: 'official_forms',
+    displayName: '公式帳票ライブラリ',
+    resolve: () => envOr('VITE_SP_LIST_OFFICIAL_FORMS', 'OfficialForms'),
+    operations: ['W'],
+    category: 'other',
+    lifecycle: 'optional',
+    baseTemplate: 101, // Document Library
+    provisioningFields: [
+      { internalName: 'Title', type: 'Text', displayName: 'Title' },
+      { internalName: 'TemplateName', type: 'Text', displayName: 'Template Name', candidates: ['Template_x0020_Name', 'TemplateName'] },
+      { internalName: 'GeneratedAt', type: 'DateTime', displayName: 'Generated At', candidates: ['Generated_x0020_At', 'GeneratedAt'] },
+    ],
+  },
+  {
+    key: 'billing_orders',
+    displayName: '請求オーダー',
+    resolve: () => envOr('VITE_SP_LIST_BILLING_ORDERS', 'BillingOrders'),
+    operations: ['R'],
+    category: 'other',
+    lifecycle: 'optional',
+    provisioningFields: [
+      { internalName: 'Title', type: 'Text', displayName: 'Order Date' },
+      { internalName: 'OrdererCode', type: 'Text', displayName: 'Orderer Code', required: true, governance: 'allow', isSilent: true, candidates: ['OrdererCode', 'Orderer_x0020_Code'] },
+      { internalName: 'OrdererName', type: 'Text', displayName: 'Orderer Name', governance: 'allow', isSilent: true, candidates: ['OrdererName', 'Orderer_x0020_Name'] },
+      { internalName: 'OrderCount', type: 'Number', displayName: 'Order Count', governance: 'allow', isSilent: true, candidates: ['OrderCount', 'Order_x0020_Count'] },
+      { internalName: 'Item', type: 'Text', displayName: 'Item', governance: 'allow' },
+    ],
+  },
+  {
+    key: 'billing_summary',
+    displayName: '月次請求サマリー',
+    resolve: () => envOr('VITE_SP_LIST_BILLING_SUMMARY', fromConfig(ListKeys.MonthlyRecordSummary)),
+    operations: ['R', 'W'],
+    category: 'other',
+    lifecycle: 'optional',
+    essentialFields: ['UserId', 'YearMonth', 'KPI_TotalDays'],
+    provisioningFields: [
+      { internalName: 'UserId', type: 'Text', displayName: 'User ID', required: true, indexed: true, candidates: ['UserId', 'User_x0020_ID'] },
+      { internalName: 'YearMonth', type: 'Text', displayName: 'Year Month', required: true, indexed: true, candidates: ['YearMonth', 'Year_x0020_Month'] },
+      { internalName: 'DisplayName', type: 'Text', displayName: 'Display Name', isSilent: true, candidates: ['Display_x0020_Name', 'DisplayName', 'cr013_displayName'] },
+      { internalName: 'LastUpdated', type: 'DateTime', displayName: 'Last Updated', isSilent: true, candidates: ['LastAggregatedAt', 'LastUpdated', 'Last_x0020_Updated'] },
+      { internalName: 'KPI_TotalDays', type: 'Number', displayName: 'Total Days', candidates: ['KPI_TotalDays', 'TotalDays', 'WorkingDays', 'Total_x0020_Days'], legacyCandidates: ['TotalDays', 'Total_x0020_Days'] },
+      { internalName: 'KPI_PlannedRows', type: 'Number', displayName: 'Planned Rows', isSilent: true, candidates: ['Planned_x0020_Rows', 'KPI_PlannedRows', 'PlannedRows', 'cr013_plannedRows'] },
+      { internalName: 'KPI_CompletedRows', type: 'Number', displayName: 'Completed Rows', isSilent: true, candidates: ['Completed_x0020_Rows', 'KPI_CompletedRows', 'CompletedRows', 'CompletedCount', 'cr013_completedRows'] },
+      { internalName: 'KPI_InProgressRows', type: 'Number', displayName: 'In Progress Rows', isSilent: true, candidates: ['In_x0020_Progress_x0020_Rows', 'KPI_InProgressRows', 'InProgressRows', 'PendingCount', 'cr013_inProgressRows'] },
+      { internalName: 'KPI_EmptyRows', type: 'Number', displayName: 'Empty Rows', isSilent: true, candidates: ['Empty_x0020_Rows', 'KPI_EmptyRows', 'EmptyRows', 'EmptyCount', 'cr013_emptyRows'] },
+      { internalName: 'KPI_SpecialNotes', type: 'Number', displayName: 'Special Notes', isSilent: true, candidates: ['Special_x0020_Notes', 'KPI_SpecialNotes', 'SpecialNotes', 'SpecialNoteCount', 'cr013_specialNotes'] },
+      { internalName: 'KPI_Incidents', type: 'Number', displayName: 'Incidents', isSilent: true, candidates: ['Incidents', 'KPI_Incidents', 'IncidentCount', 'cr013_incidents'] },
+      { internalName: 'CompletionRate', type: 'Number', displayName: 'Completion Rate' },
+      { internalName: 'Key', type: 'Text', displayName: 'Idempotency Key (Legacy)', governance: 'allow', isSilent: true, candidates: ['Key'] },
+      { internalName: 'IsLocked', type: 'Boolean', displayName: 'Locked Flag (Legacy)', governance: 'allow', isSilent: true, candidates: ['IsLocked'] },
+      { internalName: 'TotalDays', type: 'Number', displayName: 'Total Days (Legacy)', governance: 'allow', isSilent: true, candidates: ['TotalDays'] },
+      { internalName: 'Total_x0020_Days', type: 'Number', displayName: 'Total Days (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Total_x0020_Days'] },
+      { internalName: 'CompletedCount', type: 'Number', displayName: 'Completed Count (Legacy)', governance: 'allow', isSilent: true, candidates: ['CompletedCount'] },
+      { internalName: 'PendingCount', type: 'Number', displayName: 'Pending Count (Legacy)', governance: 'allow', isSilent: true, candidates: ['PendingCount'] },
+      { internalName: 'EmptyCount', type: 'Number', displayName: 'Empty Count (Legacy)', governance: 'allow', isSilent: true, candidates: ['EmptyCount'] },
+      { internalName: 'IncidentCount', type: 'Number', displayName: 'Incident Count (Legacy)', governance: 'allow', isSilent: true, candidates: ['IncidentCount'] },
+      { internalName: 'SpecialNoteCount', type: 'Number', displayName: 'Special Note Count (Legacy)', governance: 'allow', isSilent: true, candidates: ['SpecialNoteCount'] },
+      { internalName: 'FirstEntryDate', type: 'DateTime', displayName: 'First Entry Date', dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['First_x0020_Entry_x0020_Date', 'FirstEntryDate', 'cr013_firstEntryDate'] },
+      { internalName: 'LastEntryDate', type: 'DateTime', displayName: 'Last Entry Date', dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['Last_x0020_Entry_x0020_Date', 'LastEntryDate', 'cr013_lastEntryDate'] },
+      { internalName: 'IdempotencyKey', type: 'Text', displayName: 'Idempotency Key', isSilent: true, candidates: ['Idempotency_x0020_Key', 'IdempotencyKey', 'Key', 'cr013_idempotencyKey'] },
+    ],
+  },
+  {
+    key: 'pdf_output_log',
+    displayName: '帳票出力ログ',
+    resolve: () => envOr('VITE_SP_LIST_PDF_OUTPUT_LOG', fromConfig(ListKeys.PdfOutputLog)),
+    operations: ['R', 'W'],
+    category: 'other',
+    lifecycle: 'optional',
+    provisioningFields: [
+      { internalName: 'OutputType', type: 'Text', displayName: 'Output Type', candidates: ['OutputType'] },
+      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', candidates: ['UserCode'] },
+      { internalName: 'OutputDate', type: 'DateTime', displayName: 'Output Date', dateTimeFormat: 'DateTime', candidates: ['OutputDate'] },
+      { internalName: 'TargetPeriod', type: 'Text', displayName: 'Target Period', indexed: true, candidates: ['TargetPeriod'] },
+      { internalName: 'FileName', type: 'Text', displayName: 'File Name', candidates: ['FileName'] },
+      { internalName: 'FileUrl', type: 'Text', displayName: 'File URL', isSilent: true, candidates: ['FileUrl'] },
+      { internalName: 'OutputBy', type: 'Text', displayName: 'Output By', candidates: ['OutputBy'] },
+      { internalName: 'Status', type: 'Text', displayName: 'Status', candidates: ['Status'] },
+      { internalName: 'ErrorMessage', type: 'Note', displayName: 'Error Message', richText: false, isSilent: true, candidates: ['ErrorMessage'] },
+      { internalName: 'Source', type: 'Text', displayName: 'Source', candidates: ['Source'] },
+    ],
+  },
+  {
+    key: 'call_logs',
+    displayName: '電話受付ログ',
+    resolve: () => envOr('VITE_SP_LIST_CALL_LOGS', fromConfig(ListKeys.CallLogs)),
+    operations: ['R', 'W'],
+    category: 'other',
+    lifecycle: 'optional',
+    provisioningFields: [
+      { internalName: 'ReceivedAt', type: 'DateTime', displayName: 'Received At', candidates: ['ReceivedAt'] },
+      { internalName: 'CallerName', type: 'Text', displayName: 'Caller Name', candidates: ['CallerName'] },
+      { internalName: 'CallerOrg', type: 'Text', displayName: 'Caller Org', isSilent: true, candidates: ['CallerOrg'] },
+      { internalName: 'TargetStaffName', type: 'Text', displayName: 'Target Staff Name', candidates: ['TargetStaffName'] },
+      { internalName: 'ReceivedByName', type: 'Text', displayName: 'Received By Name', candidates: ['ReceivedByName'] },
+      { internalName: 'MessageBody', type: 'Note', displayName: 'Message Body', richText: false, candidates: ['MessageBody'] },
+      { internalName: 'NeedCallback', type: 'Boolean', displayName: 'Need Callback', isSilent: true, candidates: ['NeedCallback'] },
+      { internalName: 'Urgency', type: 'Text', displayName: 'Urgency', candidates: ['Urgency'] },
+      { internalName: 'Status', type: 'Text', displayName: 'Status', candidates: ['Status'] },
+      { internalName: 'RelatedUserId', type: 'Text', displayName: 'Related User ID', isSilent: true, candidates: ['RelatedUserId'] },
+      { internalName: 'RelatedUserName', type: 'Text', displayName: 'Related User Name', isSilent: true, candidates: ['RelatedUserName'] },
+      { internalName: 'CallbackDueAt', type: 'DateTime', displayName: 'Callback Due At', isSilent: true, candidates: ['CallbackDueAt'] },
+      { internalName: 'CompletedAt', type: 'DateTime', displayName: 'Completed At', isSilent: true, candidates: ['CompletedAt'] },
+    ],
+  },
+];
