+++ src/sharepoint/spListRegistry.definitions.ts
@@ -1,1290 +1,23 @@
-import { ListKeys } from '@/sharepoint/fields';
-
-import { envOr, fromConfig } from './spListRegistry.shared';
 import type { SpListEntry } from './spListRegistry.shared';
+import { masterListEntries } from './definitions/master';
+import { dailyListEntries } from './definitions/daily';
+import { attendanceListEntries } from './definitions/attendance';
+import { scheduleListEntries } from './definitions/schedule';
+import { meetingListEntries } from './definitions/meeting';
+import { handoffListEntries } from './definitions/handoff';
+import { complianceListEntries } from './definitions/compliance';
+import { otherListEntries } from './definitions/other';
 
-export const masterListEntries: readonly SpListEntry[] = [
-// ── 1. マスタ系 ──────────────────────────────────────────
-  {
-    key: 'users_master',
-    displayName: '利用者マスタ',
-    resolve: () => envOr('VITE_SP_LIST_USERS', fromConfig(ListKeys.UsersMaster)),
-    operations: ['R', 'W'],
-    category: 'master',
-    lifecycle: 'required', // 制度管理の基盤となるため required
-    essentialFields: [
-      'UserID', 'FullName', 'IsActive', 'UsageStatus'
-    ],
-    provisioningFields: [
-      { internalName: 'UserID', type: 'Text', displayName: 'User Code', required: true, indexed: true, candidates: ['UserID', 'userCode', 'UserCode'] },
-      { internalName: 'FullName', type: 'Text', displayName: 'Full Name', required: true, candidates: ['FullName', 'Full_x0020_Name', 'Title'] },
-      { internalName: 'Furigana', type: 'Text', displayName: 'Furigana' },
-      { internalName: 'FullNameKana', type: 'Text', displayName: 'Full Name Kana', candidates: ['FullNameKana', 'Full_x0020_Name_x0020_Kana'] },
-      { internalName: 'ContractDate', type: 'DateTime', displayName: 'Contract Date', dateTimeFormat: 'DateOnly', candidates: ['ContractDate', 'Contract_x0020_Date'] },
-      { internalName: 'ServiceStartDate', type: 'DateTime', displayName: 'Service Start Date', dateTimeFormat: 'DateOnly', candidates: ['ServiceStartDate', 'Service_x0020_Start_x0020_Date'] },
-      { internalName: 'ServiceEndDate', type: 'DateTime', displayName: 'Service End Date', dateTimeFormat: 'DateOnly', candidates: ['ServiceEndDate', 'Service_x0020_End_x0020_Date', 'EndDate'] },
-      { internalName: 'IsActive', type: 'Boolean', displayName: 'Is Active', default: true, candidates: ['IsActive', 'Active'] },
-      { internalName: 'UsageStatus', type: 'Text', displayName: 'Usage Status', candidates: ['UsageStatus', 'Usage_x0020_Status'] },
-      { internalName: 'IsHighIntensitySupportTarget', type: 'Boolean', displayName: 'High Intensity Target', candidates: ['IsHighIntensitySupportTarget', 'HighIntensityTarget'] },
-      { internalName: 'IsSupportProcedureTarget', type: 'Boolean', displayName: 'Support Procedure Target', candidates: ['IsSupportProcedureTarget', 'SupportProcedureTarget'] },
-      { internalName: 'severeFlag', type: 'Boolean', displayName: 'Severe Flag', candidates: ['severeFlag', 'HighSeverityFlag', 'cr013_severeFlag'] },
-      { internalName: 'TransportToDays', type: 'MultiChoice', displayName: 'Transport To Days (Legacy)', choices: ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'], governance: 'allow', isSilent: true, candidates: ['TransportToDays', 'Transport_x0020_To_x0020_Days', 'cr013_transportToDays'] },
-      { internalName: 'TransportFromDays', type: 'MultiChoice', displayName: 'Transport From Days (Legacy)', choices: ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'], governance: 'allow', isSilent: true, candidates: ['TransportFromDays', 'Transport_x0020_From_x0020_Days', 'cr013_transportFromDays'] },
-      { internalName: 'LastAssessmentDate', type: 'DateTime', displayName: 'Last Assessment Date', dateTimeFormat: 'DateOnly', candidates: ['LastAssessmentDate', 'Last_x0020_Assessment_x0020_Date'] },
-      { internalName: 'BehaviorScore', type: 'Number', displayName: 'Behavior Score', candidates: ['BehaviorScore', 'Behavior_x0020_Score'] },
-      { internalName: 'ChildBehaviorScore', type: 'Number', displayName: 'Child Behavior Score', candidates: ['ChildBehaviorScore', 'Child_x0020_Behavior_x0020_Score'] },
-      { internalName: 'ServiceTypesJson', type: 'Note', displayName: 'Service Types JSON', richText: false, candidates: ['ServiceTypesJson', 'Service_x0020_Types_x0020_JSON'] },
-      { internalName: 'EligibilityCheckedAt', type: 'DateTime', displayName: 'Eligibility Checked At', candidates: ['EligibilityCheckedAt', 'Eligibility_x0020_Checked_x0020_'] },
-      { internalName: 'Last_x0020_Assessment_x0020_Date', type: 'DateTime', displayName: 'Last Assessment Date (Legacy SP-encoded)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['Last_x0020_Assessment_x0020_Date'] },
-      { internalName: 'Behavior_x0020_Score', type: 'Number', displayName: 'Behavior Score (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Behavior_x0020_Score'] },
-      { internalName: 'Child_x0020_Behavior_x0020_Score', type: 'Number', displayName: 'Child Behavior Score (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Child_x0020_Behavior_x0020_Score'] },
-      { internalName: 'Service_x0020_Types_x0020_JSON', type: 'Note', displayName: 'Service Types JSON (Legacy SP-encoded)', richText: false, governance: 'allow', isSilent: true, candidates: ['Service_x0020_Types_x0020_JSON'] },
-      { internalName: 'Eligibility_x0020_Checked_x0020_', type: 'DateTime', displayName: 'Eligibility Checked At (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Eligibility_x0020_Checked_x0020_'] },
-      { internalName: 'GrantMunicipality', type: 'Text', displayName: 'Grant Municipality', isSilent: true, candidates: ['GrantMunicipality', 'GrantMunicipality0', 'Grant_x0020_Municipality'] },
-      { internalName: 'GrantPeriodStart', type: 'DateTime', displayName: 'Grant Period Start', dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['GrantPeriodStart', 'GrantPeriodStart0', 'Grant_x0020_Period_x0020_Start'] },
-      { internalName: 'GrantPeriodEnd', type: 'DateTime', displayName: 'Grant Period End', dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['GrantPeriodEnd', 'GrantPeriodEnd0', 'Grant_x0020_Period_x0020_End'] },
-      { internalName: 'UserCopayLimit', type: 'Text', displayName: 'User Copay Limit', isSilent: true, candidates: ['UserCopayLimit', 'UserCopayLimit0', 'User_x0020_Copay_x0020_Limit'] },
-      { internalName: 'MealAddition', type: 'Text', displayName: 'Meal Addition', isSilent: true, candidates: ['MealAddition', 'MealAddition0', 'Meal_x0020_Addition'] },
-      { internalName: 'CopayPaymentMethod', type: 'Text', displayName: 'Copay Payment Method', isSilent: true, candidates: ['CopayPaymentMethod', 'CopayPaymentMethod0', 'Copay_x0020_Payment_x0020_Method'] },
-      { internalName: 'AttendanceDays', type: 'Number', displayName: 'Attendance Days (Legacy)', governance: 'allow', isSilent: true },
-      { internalName: 'RecipientCertNumber', type: 'Text', displayName: 'Recipient Cert Number (Legacy)', governance: 'allow', isSilent: true },
-      { internalName: 'RecipientCertExpiry', type: 'DateTime', displayName: 'Recipient Cert Expiry (Legacy)', governance: 'allow', isSilent: true },
-      { internalName: 'email', type: 'Text', displayName: 'Email (Legacy)', governance: 'allow', isSilent: true },
-      { internalName: 'phone', type: 'Text', displayName: 'Phone (Legacy)', governance: 'allow', isSilent: true, candidates: ['phone', 'Phone'] },
-    ],
-  },
-  {
-    key: 'user_feature_flags',
-    displayName: 'ユーザー機能フラグ',
-    resolve: () => envOr('VITE_SP_LIST_USER_FLAGS', fromConfig(ListKeys.User_Feature_Flags)),
-    operations: ['R', 'W'],
-    category: 'master',
-    lifecycle: 'optional',
-    essentialFields: ['UserCode', 'FlagKey'],
-    provisioningFields: [
-      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true, candidates: ['UserCode', 'User_x0020_Code'] },
-      { internalName: 'FlagKey', type: 'Text', displayName: 'Flag Key', required: true, candidates: ['FlagKey', 'Flag_x0020_Key'] },
-      { internalName: 'FlagValue', type: 'Text', displayName: 'Flag Value', candidates: ['FlagValue', 'Flag_x0020_Value'] },
-      { internalName: 'ExpiresAt', type: 'DateTime', displayName: 'Expires At', dateTimeFormat: 'DateTime', candidates: ['ExpiresAt', 'Expires_x0020_At'] },
-    ],
-  },
-  {
-    key: 'user_transport_settings',
-    displayName: '利用者送迎設定',
-    resolve: () => envOr('VITE_SP_LIST_USER_TRANSPORT', fromConfig(ListKeys.UserTransportSettings)),
-    operations: ['R', 'W'],
-    category: 'master',
-    lifecycle: 'required',
-    essentialFields: [
-      'UserID', 'TransportToDays', 'TransportFromDays', 'TransportAdditionType'
-    ],
-    provisioningFields: [
-      { internalName: 'UserID', type: 'Text', displayName: 'User ID', required: true, indexed: true, candidates: ['UserID'] },
-      { internalName: 'TransportToDays', type: 'MultiChoice', displayName: 'Transport To Days', choices: ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'], candidates: ['TransportToDays'] },
-      { internalName: 'TransportFromDays', type: 'MultiChoice', displayName: 'Transport From Days', choices: ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'], candidates: ['TransportFromDays'] },
-      { internalName: 'TransportCourse', type: 'Text', displayName: 'Transport Course', candidates: ['TransportCourse'] },
-      { internalName: 'TransportSchedule', type: 'Note', displayName: 'Transport Schedule', richText: false, candidates: ['TransportSchedule'] },
-      { internalName: 'TransportAdditionType', type: 'Text', displayName: 'Transport Addition Type', required: true, candidates: ['TransportAdditionType', 'Transport_x0020_Addition_x0020_Type', 'Transport_x0020_Addition_x0020_T', 'cr013_transportAdditionType', 'cr013_transportadditiontype'] },
-    ],
-  },
-  {
-    key: 'user_benefit_profile',
-    displayName: '利用者支給量プロファイル',
-    resolve: () => envOr('VITE_SP_LIST_USER_BENEFIT', fromConfig(ListKeys.UserBenefitProfile)),
-    operations: ['R', 'W'],
-    category: 'master',
-    lifecycle: 'required',
-    essentialFields: [
-      'UserID'
-    ],
-    provisioningFields: [
-      { internalName: 'UserID', type: 'Text', displayName: 'User ID', required: true, indexed: true, isSilent: true, candidates: ['UserID', 'Title'] },
-      // RecipientCertNumber moved to _Ext to avoid 8KB limit
-      { internalName: 'RecipientCertExpiry', type: 'DateTime', displayName: 'Recipient Cert Expiry', dateTimeFormat: 'DateOnly', isSilent: true },
-      { internalName: 'GrantMunicipality', type: 'Text', displayName: 'Grant Municipality', isSilent: true, candidates: ['GrantMunicipality', 'Grant_x0020_Municipality'] },
-      { internalName: 'GrantPeriodStart', type: 'DateTime', displayName: 'Grant Period Start', dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['GrantPeriodStart', 'Grant_x0020_Period_x0020_Start'] },
-      { internalName: 'GrantPeriodEnd', type: 'DateTime', displayName: 'Grant Period End', dateTimeFormat: 'DateOnly', indexed: true, isSilent: true, candidates: ['GrantPeriodEnd', 'Grant_x0020_Period_x0020_End'] },
-      { internalName: 'DisabilitySupportLevel', type: 'Text', displayName: 'Disability Support Level', isSilent: true },
-      { internalName: 'GrantedDaysPerMonth', type: 'Text', displayName: 'Granted Days Per Month', isSilent: true },
-      { internalName: 'UserCopayLimit', type: 'Text', displayName: 'User Copay Limit', isSilent: true, candidates: ['UserCopayLimit', 'User_x0020_Copay_x0020_Limit'] },
-      { internalName: 'MealAddition', type: 'Text', displayName: 'Meal Addition', isSilent: true, candidates: ['MealAddition', 'Meal_x0020_Addition'] },
-      { internalName: 'CopayPaymentMethod', type: 'Text', displayName: 'Copay Payment Method', isSilent: true, candidates: ['CopayPaymentMethod', 'Copay_x0020_Payment_x0020_Method'] },
-    ],
-  },
-  {
-    key: 'user_benefit_profile_ext',
-    displayName: '利用者支給量プロファイル (拡充)',
-    resolve: () => envOr('VITE_SP_LIST_USER_BENEFIT_EXT', fromConfig(ListKeys.UserBenefitProfileExt)),
-    operations: ['R', 'W'],
-    category: 'master',
-    lifecycle: 'required',
-    essentialFields: [
-      'UserID', 'RecipientCertNumber'
-    ],
-    provisioningFields: [
-      { internalName: 'UserID', type: 'Text', displayName: 'User ID', required: true, indexed: true, candidates: ['UserID', 'Title'] },
-      { internalName: 'RecipientCertNumber', type: 'Text', displayName: 'Recipient Cert Number', required: true, isSilent: true, candidates: ['RecipientCertNumber', 'RecipientCertNumber0', 'Recipient_x0020_Cert_x0020_Numbe'] },
-    ],
-  },
-  {
-    key: 'staff_master',
-    displayName: '職員マスタ',
-    resolve: () => envOr('VITE_SP_LIST_STAFF', fromConfig(ListKeys.StaffMaster)),
-    operations: ['R'],
-    category: 'master',
-    lifecycle: 'required',
-    essentialFields: [
-      'StaffID', 'FullName', 'Role', 'RBACRole', 'IsActive', 'Department'
-    ],
-    provisioningFields: [
-      { internalName: 'StaffID', type: 'Text', displayName: 'Staff ID', required: true, candidates: ['StaffID', 'Staff_x0020_ID', 'StaffCode'] },
-      { internalName: 'FullName', type: 'Text', displayName: 'Full Name', required: true, candidates: ['FullName', 'Full_x0020_Name', 'Title'] },
-      { internalName: 'Furigana', type: 'Text', displayName: 'Furigana' },
-      { internalName: 'FullNameKana', type: 'Text', displayName: 'Full Name Kana', candidates: ['FullNameKana', 'Full_x0020_Name_x0020_Kana'] },
-      { internalName: 'JobTitle', type: 'Text', displayName: 'Job Title', candidates: ['JobTitle', 'Job_x0020_Title'] },
-      { internalName: 'EmploymentType', type: 'Text', displayName: 'Employment Type', candidates: ['EmploymentType', 'Employment_x0020_Type'] },
-      { internalName: 'RBACRole', type: 'Text', displayName: 'RBAC Role', candidates: ['RBACRole', 'RBAC_x0020_Role'] },
-      { internalName: 'Role', type: 'Text', displayName: 'Role' },
-      { internalName: 'IsActive', type: 'Boolean', displayName: 'Is Active', default: true, candidates: ['IsActive', 'Active'] },
-      { internalName: 'Department', type: 'Text', displayName: 'Department' },
-      { internalName: 'Work_x0020_Days', type: 'Text', displayName: 'Work Days (Legacy SP-encoded)', isSilent: true, governance: 'allow' },
-      { internalName: 'WorkDays', type: 'MultiChoice', displayName: 'Work Days', choices: ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'] },
-      { internalName: 'BaseWorkingDays', type: 'Number', displayName: 'Base Working Days', governance: 'allow', isSilent: true, candidates: ['BaseWorkingDays', 'BaseWorkDays'] },
-      { internalName: 'StaffCode', type: 'Text', displayName: 'Staff Code (Legacy)', governance: 'allow', isSilent: true, candidates: ['StaffCode'] },
-      { internalName: 'ActiveFlag', type: 'Boolean', displayName: 'Active Flag (Legacy)', governance: 'allow', isSilent: true, candidates: ['ActiveFlag'] },
-      { internalName: 'BaseShiftStartTime', type: 'Text', displayName: 'Base Shift Start' },
-      { internalName: 'BaseShiftEndTime', type: 'Text', displayName: 'Base Shift End' },
-      { internalName: 'HireDate', type: 'DateTime', displayName: 'Hire Date', dateTimeFormat: 'DateOnly' },
-      { internalName: 'ResignDate', type: 'DateTime', displayName: 'Resign Date', dateTimeFormat: 'DateOnly' },
-      { internalName: 'Email', type: 'Text', displayName: 'Email' },
-      { internalName: 'Phone', type: 'Text', displayName: 'Phone' },
-      { internalName: 'Certifications', type: 'Note', displayName: 'Certifications' },
-    ],
-  },
-  {
-    key: 'org_master',
-    displayName: '組織マスタ',
-    resolve: () => envOr('VITE_SP_LIST_ORG_MASTER', fromConfig(ListKeys.OrgMaster)),
-    operations: ['R'],
-    category: 'master',
-    lifecycle: 'required',
-    essentialFields: ['OrgCode', 'OrgType', 'Audience'],
-    provisioningFields: [
-      { internalName: 'OrgCode', type: 'Text', displayName: 'Organization Code', required: true },
-      { internalName: 'OrgType', type: 'Text', displayName: 'Organization Type' },
-      { internalName: 'Audience', type: 'Text', displayName: 'Audience' },
-      { internalName: 'SortOrder', type: 'Number', displayName: 'Sort Order' },
-      { internalName: 'IsActive', type: 'Boolean', displayName: 'Active', default: true },
-      { internalName: 'Notes', type: 'Note', displayName: 'Notes' },
-    ],
-  },
-  {
-    key: 'holiday_master',
-    displayName: '祝日・休業日マスタ',
-    resolve: () => envOr('VITE_SP_LIST_HOLIDAY_MASTER', fromConfig(ListKeys.HolidayMaster)),
-    operations: ['R'],
-    category: 'master',
-    lifecycle: 'required',
-    essentialFields: ['Date', 'Type'],
-    provisioningFields: [
-      { internalName: 'Date', type: 'DateTime', displayName: 'Date', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['Date', 'HolidayDate', 'TargetDate', 'RecordDate'] },
-      { internalName: 'Label', type: 'Text', displayName: 'Label', candidates: ['Label', 'Title', 'Name'] },
-      { internalName: 'Type', type: 'Text', displayName: 'Type', candidates: ['Type', 'Category', 'HolidayType'] },
-      { internalName: 'FiscalYear', type: 'Text', displayName: 'Fiscal Year', candidates: ['FiscalYear', 'Year', 'BusinessYear'] },
-      { internalName: 'IsActive', type: 'Boolean', displayName: 'Is Active', default: true, candidates: ['IsActive', 'Active', 'Enabled'] },
-    ],
-  },
-];
-
-export const dailyListEntries: readonly SpListEntry[] = [
-// ── 2. 日々の記録系 ─────────────────────────────────────
-  {
-    key: 'support_record_daily',
-    displayName: '日次支援記録 (親/統括)',
-    resolve: () => envOr('VITE_SP_LIST_DAILY_RECORD', fromConfig(ListKeys.DailyRecordParent)),
-    operations: ['R', 'W'],
-    category: 'daily',
-    lifecycle: 'required',
-    // DAILY_RECORD_CANONICAL_ESSENTIALS と一致させる:
-    //   Title / RecordDate / UserRowsJSON が必須。
-    //   ReporterName は欠落時でも記録者不明として継続可能なためオプション。
-    essentialFields: ['Title', 'RecordDate', 'UserRowsJSON'],
-    provisioningFields: [
-      { internalName: 'RecordDate', type: 'DateTime', displayName: 'Record Date', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['RecordDate', 'Record_x0020_Date'] },
-      { internalName: 'ReporterName', type: 'Text', displayName: 'Reporter Name', candidates: ['ReporterName', 'Reporter_x0020_Name'] },
-      { internalName: 'ReporterRole', type: 'Text', displayName: 'Reporter Role', candidates: ['ReporterRole', 'Reporter_x0020_Role'] },
-      { internalName: 'UserRowsJSON', type: 'Note', displayName: 'User Rows JSON', required: true, richText: false, candidates: ['UserRowsJSON', 'User_x0020_Rows_x0020_JSON'] },
-      { internalName: 'UserCount', type: 'Number', displayName: 'User Count', candidates: ['UserCount', 'User_x0020_Count'] },
-      { internalName: 'ApprovalStatus', type: 'Text', displayName: 'Approval Status', candidates: ['ApprovalStatus', 'Approval_x0020_Status'] },
-      { internalName: 'ApprovedBy', type: 'Text', displayName: 'Approved By', candidates: ['ApprovedBy', 'Approved_x0020_By'] },
-      { internalName: 'ApprovedAt', type: 'DateTime', displayName: 'Approved At', dateTimeFormat: 'DateTime', candidates: ['ApprovedAt', 'Approved_x0020_At'] },
-      { internalName: 'Completed', type: 'Boolean', displayName: 'Completed (Legacy)', governance: 'allow', isSilent: true, candidates: ['Completed'] },
-      { internalName: 'Incident', type: 'Text', displayName: 'Incident (Legacy)', governance: 'allow', isSilent: true, candidates: ['Incident'] },
-      { internalName: 'UserId', type: 'Text', displayName: 'User ID (Legacy Row Aggregate)', governance: 'allow', isSilent: true, candidates: ['UserId', 'UserID', 'UserCode', 'User_x0020_ID', 'cr013_personId', 'cr013_usercode'] },
-      { internalName: 'Status', type: 'Text', displayName: 'Status (Legacy Row Aggregate)', governance: 'allow', isSilent: true, candidates: ['Status', 'status', 'cr013_status'] },
-      { internalName: 'cr013_recorddate', type: 'DateTime', displayName: 'Record Date (Legacy)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['cr013_recorddate', 'RecordDate', 'Date'] },
-      { internalName: 'cr013_specialnote', type: 'Note', displayName: 'Special Note (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['cr013_specialnote', 'SpecialNote'] },
-    ],
-  },
-  {
-    key: 'support_procedure_record_daily',
-    displayName: '支援手順記録 (ISP詳細)',
-    resolve: () => envOr('VITE_SP_LIST_PROCEDURE_RECORD', fromConfig(ListKeys.ProcedureRecordDaily)),
-    operations: ['R', 'W'],
-    category: 'daily',
-    lifecycle: 'required', // 三層モデルの核となるため required
-    essentialFields: [
-      'UserCode', 'RecordDate', 'ExecutionStatus', 'PlanningSheetId'
-    ],
-    provisioningFields: [
-      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true, candidates: ['UserCode', 'User_ID', 'cr013_userCode'] },
-      { internalName: 'RecordDate', type: 'DateTime', displayName: 'Record Date', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['RecordDate', 'Date', 'cr013_recordDate'] },
-      { internalName: 'ISP_x0020_ID', type: 'Number', displayName: 'ISP ID', candidates: ['ISP_x0020_ID', 'ISPId', 'cr013_ispId'] },
-      { internalName: 'PlanningSheetId', type: 'Text', displayName: 'Planning Sheet ID', required: true, candidates: ['PlanningSheetId', 'cr013_planningSheetId'] },
-      { internalName: 'ProcedureText', type: 'Note', displayName: 'Procedure Text', richText: false, candidates: ['ProcedureText', 'Procedure', 'Procedure_x0020_Text', 'cr013_procedureText'] },
-      { internalName: 'ExecutionStatus', type: 'Text', displayName: 'Execution Status', candidates: ['ExecutionStatus', 'Status', 'Execution_x0020_Status', 'cr013_executionStatus'] },
-      { internalName: 'TimeSlot', type: 'Text', displayName: 'Time Slot', candidates: ['TimeSlot', 'Time', 'Time_x0020_Slot', 'cr013_timeSlot'] },
-      { internalName: 'Activity', type: 'Text', displayName: 'Activity', candidates: ['Activity', 'Action', 'cr013_activity'] },
-      { internalName: 'PerformedBy', type: 'Text', displayName: 'Performed By', candidates: ['PerformedBy', 'Staff', 'Performed_x0020_By', 'cr013_performedBy'] },
-      { internalName: 'PerformedAt', type: 'DateTime', displayName: 'Performed At', candidates: ['PerformedAt', 'Time', 'Performed_x0020_At', 'cr013_performedAt'] },
-      { internalName: 'UserResponse', type: 'Note', displayName: 'User Response', richText: false, candidates: ['UserResponse', 'User_x0020_Response'] },
-      { internalName: 'SpecialNotes', type: 'Note', displayName: 'Special Notes', richText: false, candidates: ['SpecialNotes', 'Special_x0020_Notes'] },
-      { internalName: 'Handoff_x0020_Notes', type: 'Note', displayName: 'Handoff Notes', richText: false, candidates: ['HandoffNotes', 'Handoff_x0020_Notes', 'cr013_handoffNotes'] },
-    ],
-  },
-  {
-    key: 'support_record_rows',
-    displayName: '日次支援記録 (詳細行)',
-    resolve: () => envOr('VITE_SP_LIST_PROCEDURE_RECORD_ROWS', 'DailyRecordRows'),
-    operations: ['R', 'W'],
-    category: 'daily',
-    lifecycle: 'required',
-    essentialFields: [
-      'ParentID', 'UserID'
-    ],
-    provisioningFields: [
-      { internalName: 'ParentID', type: 'Number', displayName: 'Parent ID', required: true, indexed: true, candidates: ['ParentID', 'Parent_x0020_ID'] },
-      { internalName: 'UserID', type: 'Text', displayName: 'User ID', required: true, indexed: true, candidates: ['UserID', 'User_x0020_ID', 'Title'] },
-      { internalName: 'Version', type: 'Number', displayName: 'Version', default: 1, governance: 'allow', isSilent: true, candidates: ['Version', 'VersionNo', 'cr013_version'] },
-      { internalName: 'Status', type: 'Text', displayName: 'Status' },
-      { internalName: 'Status0', type: 'Text', displayName: 'Status (Legacy)', governance: 'allow', isSilent: true, candidates: ['Status0'] },
-      { internalName: 'Observation', type: 'Note', displayName: 'Observation (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['Observation'] },
-      { internalName: 'User_x0020_ID', type: 'Text', displayName: 'User ID (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['User_x0020_ID'] },
-      { internalName: 'RowNo', type: 'Number', displayName: 'Row No', candidates: ['RowNo', 'cr013_rowNo', 'RowIndex'] },
-      { internalName: 'Payload', type: 'Note', displayName: 'Row Data JSON', richText: false, governance: 'allow', isSilent: true, candidates: ['Payload', 'payload', 'cr013_payload', 'cr013_draftJson', 'PayloadJSON', 'SupportRecordPayload', 'Payload_x0020_JSON', 'Observation'] },
-      { internalName: 'RecordedAt', type: 'DateTime', displayName: 'Recorded At', candidates: ['RecordedAt', 'Recorded_x0020_At'] },
-    ],
-  },
-  {
-    key: 'daily_activity_records',
-    displayName: '日次活動記録',
-    resolve: () => envOr('VITE_SP_LIST_DAILY_ACTIVITY_RECORDS', fromConfig(ListKeys.DailyActivityRecords)),
-    operations: ['R', 'W'],
-    category: 'daily',
-    lifecycle: 'required',
-    essentialFields: ['UserCode', 'RecordDate', 'TimeSlot', 'Observation'],
-    provisioningFields: [
-      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, candidates: ['UserCode', 'User_x0020_Code'] },
-      { internalName: 'RecordDate', type: 'DateTime', displayName: 'Record Date', required: true, dateTimeFormat: 'DateOnly', candidates: ['RecordDate', 'Record_x0020_Date'] },
-      { internalName: 'TimeSlot', type: 'Text', displayName: 'Time Slot', candidates: ['TimeSlot', 'Time_x0020_Slot'] },
-      { internalName: 'PlanSlotKey', type: 'Text', displayName: 'Plan Slot Key', candidates: ['PlanSlotKey', 'Plan_x0020_Slot_x0020_Key'] },
-      { internalName: 'PlannedActivity', type: 'Text', displayName: 'Planned Activity', candidates: ['PlannedActivity', 'Planned_x0020_Activity'] },
-      { internalName: 'RecordedAtText', type: 'Text', displayName: 'Recorded At', candidates: ['RecordedAtText', 'RecordedAt'] },
-      { internalName: 'Observation', type: 'Note', displayName: 'Observation' },
-      { internalName: 'Behavior', type: 'Note', displayName: 'Behavior' },
-      { internalName: 'version', type: 'Text', displayName: 'Intensity/Version', governance: 'allow' },
-      { internalName: 'duration', type: 'Text', displayName: 'Duration' },
-      { internalName: 'Order', type: 'Number', displayName: 'Order' },
-    ],
-  },
-  {
-    key: 'service_provision_records',
-    displayName: 'サービス提供実績',
-    resolve: () => envOr('VITE_SP_LIST_SERVICE_PROVISION', 'ServiceProvisionRecords'),
-    operations: ['R', 'W'],
-    category: 'daily',
-    lifecycle: 'required', // 業務継続に不可欠なため required へ昇格
-    essentialFields: [
-      'EntryKey', 'UserCode', 'RecordDate', 'Status', 'StartHHMM', 'EndHHMM'
-    ],
-    provisioningFields: [
-      { internalName: 'EntryKey', type: 'Text', displayName: 'Entry Key', required: true, indexed: true, candidates: ['EntryKey', 'Entry_x0020_Key'] },
-      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true, candidates: ['UserCode', 'User_x0020_Code'] },
-      { internalName: 'RecordDate', type: 'DateTime', displayName: 'Record Date', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['RecordDate', 'Record_x0020_Date'] },
-      { internalName: 'Status', type: 'Text', displayName: 'Status' },
-      { internalName: 'StartHHMM', type: 'Text', displayName: 'Start Time (HHMM)', candidates: ['StartHHMM', 'Start_x0020_Time_x0020__x0028_HH'] },
-      { internalName: 'EndHHMM', type: 'Text', displayName: 'End Time (HHMM)', candidates: ['EndHHMM', 'End_x0020_Time_x0020__x0028_HHMM'] },
-      { internalName: 'HasTransport', type: 'Boolean', displayName: 'Has Transport', candidates: ['HasTransport', 'Has_x0020_Transport'] },
-      { internalName: 'HasTransportPickup', type: 'Boolean', displayName: 'Has Transport Pickup' },
-      { internalName: 'HasTransportDropoff', type: 'Boolean', displayName: 'Has Transport Dropoff' },
-      { internalName: 'HasMeal', type: 'Boolean', displayName: 'Has Meal', candidates: ['HasMeal', 'Has_x0020_Meal'] },
-      { internalName: 'HasBath', type: 'Boolean', displayName: 'Has Bath', candidates: ['HasBath', 'Has_x0020_Bath'] },
-      { internalName: 'HasExtended', type: 'Boolean', displayName: 'Has Extended', candidates: ['HasExtended', 'Has_x0020_Extended'] },
-      { internalName: 'HasAbsentSupport', type: 'Boolean', displayName: 'Has Absent Support', candidates: ['HasAbsentSupport', 'Has_x0020_Absent_x0020_Support'] },
-      { internalName: 'Note', type: 'Note', displayName: 'Note', richText: false },
-      { internalName: 'Status0', type: 'Text', displayName: 'Status (Legacy)', governance: 'allow', isSilent: true, candidates: ['Status0'] },
-      { internalName: 'Source', type: 'Text', displayName: 'Source' },
-      { internalName: 'UpdatedByUPN', type: 'Text', displayName: 'UpdatedByUPN' },
-    ],
-  },
-  {
-    key: 'activity_diary',
-    displayName: '活動日誌',
-    resolve: () => envOr('VITE_SP_LIST_ACTIVITY_DIARY', 'ActivityDiary'),
-    operations: ['R', 'W'],
-    category: 'daily',
-    lifecycle: 'required', // 業務継続に不可欠なため required へ昇格
-    essentialFields: ['UserID', 'Date', 'Shift', 'Category'],
-    provisioningFields: [
-      { internalName: 'UserID', type: 'Text', displayName: 'User ID', required: true },
-      { internalName: 'Date', type: 'DateTime', displayName: 'Date', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['Date', 'ServiceDate', 'RecordDate', 'EntryDate', 'cr013_date'] },
-      { internalName: 'Shift', type: 'Choice', displayName: 'Shift', choices: ['AM', 'PM', '1日'], required: true },
-      { internalName: 'Category', type: 'Choice', displayName: 'Category', choices: ['請負', '個別', '外活動', '余暇'], required: true },
-      { internalName: 'LunchAmount', type: 'Choice', displayName: 'Lunch Amount', choices: ['完食', '8割', '半分', '少量', 'なし'] },
-      { internalName: 'ProblemBehavior', type: 'Boolean', displayName: 'Problem Behavior' },
-      { internalName: 'Seizure', type: 'Boolean', displayName: 'Seizure' },
-      { internalName: 'Goals', type: 'Note', displayName: 'Goals JSON', richText: false, governance: 'allow', isSilent: true, candidates: ['Goals', 'goals', 'GoalIds', 'cr013_goals'] },
-      { internalName: 'Notes', type: 'Note', displayName: 'Notes', richText: false },
-    ],
-  },
-  {
-    key: 'abc_behavior_records',
-    displayName: 'ABC行動記録',
-    resolve: () => envOr('VITE_SP_LIST_ABC_BEHAVIOR_RECORDS', 'AbcBehaviorRecords'),
-    operations: ['R', 'W'],
-    category: 'daily',
-    lifecycle: 'required',
-    essentialFields: ['AbcRecordId', 'UserId', 'RecordDate', 'Behavior', 'IsDeleted'],
-    provisioningFields: [
-      { internalName: 'AbcRecordId', type: 'Text', displayName: 'ABC記録ID (UUID)', required: true, indexed: true, candidates: ['AbcRecordId', 'abcRecordId', 'cr013_abcRecordId'] },
-      { internalName: 'UserId', type: 'Text', displayName: '利用者コード', required: true, indexed: true, candidates: ['UserId', 'userId', 'cr013_userId', 'UserID'] },
-      { internalName: 'RecordDate', type: 'DateTime', displayName: '記録日', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['RecordDate', 'recordDate', 'cr013_recordDate'] },
-      { internalName: 'OccurredAt', type: 'Text', displayName: '発生時刻', candidates: ['OccurredAt', 'occurredAt', 'cr013_occurredAt'] },
-      { internalName: 'Setting', type: 'Text', displayName: '環境・状況(S)', candidates: ['Setting', 'setting'] },
-      { internalName: 'Antecedent', type: 'Note', displayName: '直前の状況(A)', richText: false, candidates: ['Antecedent', 'antecedent'] },
-      { internalName: 'Behavior', type: 'Note', displayName: '行動の様子(B)', required: true, richText: false, candidates: ['Behavior', 'behavior'] },
-      { internalName: 'Consequence', type: 'Note', displayName: '直後の結果(C)', richText: false, candidates: ['Consequence', 'consequence'] },
-      { internalName: 'Intensity', type: 'Text', displayName: '強度', candidates: ['Intensity', 'intensity'] },
-      { internalName: 'DurationMinutes', type: 'Number', displayName: '継続時間(分)', candidates: ['DurationMinutes', 'durationMinutes'] },
-      { internalName: 'RiskFlag', type: 'Boolean', displayName: '自他害・物損フラグ', candidates: ['RiskFlag', 'riskFlag'] },
-      { internalName: 'TagsJson', type: 'Note', displayName: 'タグJSON', richText: false, candidates: ['TagsJson', 'tagsJson'] },
-      { internalName: 'Notes', type: 'Note', displayName: '備考・特記事項', richText: false, candidates: ['Notes', 'notes'] },
-      { internalName: 'SourcePage', type: 'Text', displayName: '遷移元画面', candidates: ['SourcePage', 'sourcePage'] },
-      { internalName: 'SourceDate', type: 'Text', displayName: '遷移元日付', candidates: ['SourceDate', 'sourceDate'] },
-      { internalName: 'SourceSlotId', type: 'Text', displayName: '遷移元スロットID', candidates: ['SourceSlotId', 'sourceSlotId'] },
-      { internalName: 'SourceSlotLabel', type: 'Text', displayName: '遷移元スロット名', candidates: ['SourceSlotLabel', 'sourceSlotLabel'] },
-      { internalName: 'ReturnUrl', type: 'Text', displayName: '戻り先URL', candidates: ['ReturnUrl', 'returnUrl'] },
-      { internalName: 'RecorderName', type: 'Text', displayName: '記録者氏名', candidates: ['RecorderName', 'recorderName'] },
-      { internalName: 'CreatedByCode', type: 'Text', displayName: '作成者コード', required: true, candidates: ['CreatedByCode', 'createdByCode', 'CreatedBy'] },
-      { internalName: 'UpdatedByCode', type: 'Text', displayName: '更新者コード', candidates: ['UpdatedByCode', 'updatedByCode', 'ModifiedBy'] },
-      { internalName: 'CreatedAt', type: 'Text', displayName: '作成日時', required: true, candidates: ['CreatedAt', 'createdAt', 'Created'] },
-      { internalName: 'UpdatedAt', type: 'Text', displayName: '更新日時', candidates: ['UpdatedAt', 'updatedAt', 'Modified'] },
-      { internalName: 'IsDeleted', type: 'Boolean', displayName: '論理削除フラグ', default: false, candidates: ['IsDeleted', 'isDeleted'] },
-      { internalName: 'DeletedAt', type: 'Text', displayName: '削除日時', candidates: ['DeletedAt', 'deletedAt'] },
-      { internalName: 'DeletedByCode', type: 'Text', displayName: '削除者コード', candidates: ['DeletedByCode', 'deletedByCode'] },
-    ],
-  },
-];
-
-/**
- * DailyRecordRows 17行モデル向け 追加予定列（dry-run 専用）
- *
- * 注意:
- * - 本定義は guarded apply 前の候補整理用途。
- * - 現時点では support_record_rows の provisioningFields へは反映しない。
- */
-export const DAILY_RECORD_ROWS_17ROW_PROPOSED_FIELDS = [
-  { internalName: 'UserCode0', type: 'Text', displayName: '利用者コード', candidates: ['UserCode0', 'UserCode', 'User_x0020_Code'] },
-  { internalName: 'RecordDate0', type: 'DateTime', displayName: '記録日', dateTimeFormat: 'DateOnly', candidates: ['RecordDate0', 'RecordDate', 'Record_x0020_Date'] },
-  { internalName: 'RowNo0', type: 'Number', displayName: '行番号', candidates: ['RowNo0', 'RowNo'] },
-  { internalName: 'TimeSlot0', type: 'Text', displayName: '時間帯', candidates: ['TimeSlot0', 'TimeSlot', 'Time_x0020_Slot'] },
-  { internalName: 'Activity0', type: 'Text', displayName: '活動内容', candidates: ['Activity0', 'Activity'] },
-  { internalName: 'PersonManual0', type: 'Note', displayName: '本人の動き', richText: false, candidates: ['PersonManual0', 'PersonManual'] },
-  { internalName: 'SupporterManual0', type: 'Note', displayName: '支援者の動き', richText: false, candidates: ['SupporterManual0', 'SupporterManual'] },
-  { internalName: 'Situation0', type: 'Note', displayName: '当日の様子・記録', richText: false, candidates: ['Situation0', 'Situation'] },
-  { internalName: 'Completed0', type: 'Text', displayName: '実行ステータス', candidates: ['Completed0', 'Completed'] },
-  { internalName: 'ProcedureType0', type: 'Text', displayName: 'ブロックカテゴリ', candidates: ['ProcedureType0', 'ProcedureType'] },
-  { internalName: 'ParentRowNo0', type: 'Number', displayName: '親行番号', candidates: ['ParentRowNo0', 'ParentRowNo'] },
-  { internalName: 'CreatedByName0', type: 'Text', displayName: '記録者氏名', candidates: ['CreatedByName0', 'CreatedByName'] },
-  { internalName: 'SourceFileName0', type: 'Text', displayName: '取込元ファイル名', required: true, candidates: ['SourceFileName0', 'SourceFileName'] },
-] as const;
-
-export const attendanceListEntries: readonly SpListEntry[] = [
-// ── 3. 出席管理系 ──────────────────────────────────────
-  {
-    key: 'daily_attendance',
-    displayName: '日次出欠',
-    resolve: () => envOr('VITE_SP_LIST_ATTENDANCE', fromConfig(ListKeys.DailyAttendance)),
-    operations: ['R', 'W'],
-    category: 'attendance',
-    lifecycle: 'required',
-    essentialFields: ['UserID', 'Date', 'Status'],
-    provisioningFields: [
-      { internalName: 'UserID', type: 'Text', displayName: 'User ID', required: true, candidates: ['UserID', 'User_x0020_ID'] },
-      { internalName: 'User_x0020_ID', type: 'Text', displayName: 'User ID (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['User_x0020_ID'] },
-      { internalName: 'Date', type: 'DateTime', displayName: 'Date', required: true, dateTimeFormat: 'DateOnly', indexed: true },
-      { internalName: 'Status', type: 'Choice', displayName: 'Status', choices: ['通常', '欠席', '振替', '休止'], required: true },
-      { internalName: 'IsTrial', type: 'Boolean', displayName: 'Trial', candidates: ['IsTrial', 'Trial'] },
-      { internalName: 'Trial', type: 'Boolean', displayName: 'Trial (Legacy)', governance: 'allow', isSilent: true, candidates: ['Trial'] },
-      { internalName: 'Notes', type: 'Note', displayName: 'Notes', richText: false },
-    ],
-  },
-  {
-    key: 'attendance_users',
-    displayName: '出席管理ユーザー',
-    resolve: () => envOr('VITE_SP_LIST_ATTENDANCE_USERS', fromConfig(ListKeys.AttendanceUsers)),
-    operations: ['R'],
-    category: 'attendance',
-    lifecycle: 'required',
-    essentialFields: [
-      'UserCode', 'IsActive', 'UsageStatus', 'ServiceEndDate'
-    ],
-    provisioningFields: [
-      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true },
-      { internalName: 'IsActive', type: 'Boolean', displayName: 'Is Active', default: true },
-      { internalName: 'UsageStatus', type: 'Text', displayName: 'Usage Status' },
-      { internalName: 'ServiceEndDate', type: 'DateTime', displayName: 'Service End Date', dateTimeFormat: 'DateOnly' },
-      { internalName: 'Service_x0020_End_x0020_Date', type: 'DateTime', displayName: 'Service End Date (Legacy SP-encoded)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['Service_x0020_End_x0020_Date'] },
-      { internalName: 'IsTransportTarget', type: 'Boolean', displayName: 'Is Transport Target' },
-      { internalName: 'StandardMinutes', type: 'Number', displayName: 'Standard Minutes' },
-      { internalName: 'DefaultTransportToMethod', type: 'Text', displayName: 'Default Transport To Method (Legacy)', governance: 'allow', isSilent: true, candidates: ['DefaultTransportToMethod'] },
-      { internalName: 'DefaultTransportFromMethod', type: 'Text', displayName: 'Default Transport From Method (Legacy)', governance: 'allow', isSilent: true, candidates: ['DefaultTransportFromMethod'] },
-      { internalName: 'DefaultTransportToNote', type: 'Note', displayName: 'Default Transport To Note (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['DefaultTransportToNote'] },
-      { internalName: 'DefaultTransportFromNote', type: 'Note', displayName: 'Default Transport From Note (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['DefaultTransportFromNote'] },
-    ],
-  },
-  {
-    key: 'attendance_daily',
-    displayName: '日次出席詳細',
-    resolve: () => envOr('VITE_SP_LIST_ATTENDANCE_DAILY', fromConfig(ListKeys.AttendanceDaily)),
-    operations: ['R', 'W'],
-    category: 'attendance',
-    lifecycle: 'required',
-    // ATTENDANCE_DAILY_ESSENTIALS と同期
-    essentialFields: ['UserCode', 'RecordDate', 'Status', 'CheckInAt'],
-    provisioningFields: [
-      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true, candidates: ['UserCode'] },
-      { internalName: 'RecordDate', type: 'DateTime', displayName: 'Record Date', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['RecordDate'] },
-      { internalName: 'Status', type: 'Text', displayName: 'Status', required: true },
-      { internalName: 'CheckInAt', type: 'DateTime', displayName: 'Check-in Time', dateTimeFormat: 'DateTime', candidates: ['CheckInAt'] },
-      { internalName: 'CheckOutAt', type: 'DateTime', displayName: 'Check-out Time', dateTimeFormat: 'DateTime', candidates: ['CheckOutAt'] },
-      { internalName: 'CntAttendIn', type: 'Number', displayName: 'Attendance In Count', isSilent: true, candidates: ['CntAttendIn'] },
-      { internalName: 'CntAttendOut', type: 'Number', displayName: 'Attendance Out Count', isSilent: true, candidates: ['CntAttendOut'] },
-      { internalName: 'TransportTo', type: 'Boolean', displayName: 'Transport To', candidates: ['TransportTo'] },
-      { internalName: 'TransportFrom', type: 'Boolean', displayName: 'Transport From', candidates: ['TransportFrom'] },
-      { internalName: 'ProvidedMinutes', type: 'Number', displayName: 'Provided Minutes', candidates: ['ProvidedMinutes'] },
-      { internalName: 'IsEarlyLeave', type: 'Boolean', displayName: 'Early Leave', isSilent: true, candidates: ['IsEarlyLeave'] },
-      { internalName: 'UserConfirmedAt', type: 'DateTime', displayName: 'User Confirmed Time', dateTimeFormat: 'DateTime', isSilent: true, candidates: ['UserConfirmedAt'] },
-      { internalName: 'AbsentMorningContacted', type: 'Boolean', displayName: 'Absent Morning Contacted', isSilent: true, candidates: ['AbsentMorningContacted'] },
-      { internalName: 'AbsentMorningMethod', type: 'Text', displayName: 'Absent Morning Method', isSilent: true, candidates: ['AbsentMorningMethod'] },
-      { internalName: 'EveningChecked', type: 'Boolean', displayName: 'Evening Checked', candidates: ['EveningChecked'] },
-      { internalName: 'EveningNote', type: 'Note', displayName: 'Evening Note', candidates: ['EveningNote'] },
-      { internalName: 'IsAbsenceAddonClaimable', type: 'Boolean', displayName: 'Absence Add-on Claimable', isSilent: true, candidates: ['IsAbsenceAddonClaimable'] },
-    ],
-  },
-  {
-    key: 'staff_attendance',
-    displayName: '職員出勤管理',
-    resolve: () => envOr('VITE_SP_LIST_STAFF_ATTENDANCE', fromConfig(ListKeys.StaffAttendance)),
-    operations: ['R', 'W'],
-    category: 'attendance',
-    lifecycle: 'required',
-    essentialFields: ['StaffId', 'RecordDate', 'Status'],
-    provisioningFields: [
-      { internalName: 'StaffId', type: 'Text', displayName: 'Staff ID', required: true, indexed: true, candidates: ['StaffId', 'Staff_x0020_ID'] },
-      { internalName: 'Staff_x0020_ID', type: 'Text', displayName: 'Staff ID (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Staff_x0020_ID'] },
-      { internalName: 'RecordDate', type: 'DateTime', displayName: 'Record Date', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['RecordDate', 'Record_x0020_Date'] },
-      { internalName: 'Status', type: 'Text', displayName: 'Status', required: true },
-      { internalName: 'CheckInAt', type: 'DateTime', displayName: 'Check-in', dateTimeFormat: 'DateTime', candidates: ['CheckInAt', 'Check_x002d_in'] },
-      { internalName: 'CheckOutAt', type: 'DateTime', displayName: 'Check-out', dateTimeFormat: 'DateTime', candidates: ['CheckOutAt', 'Check_x002d_out'] },
-      { internalName: 'IsFinalized', type: 'Boolean', displayName: 'Finalized', default: false, candidates: ['IsFinalized', 'Finalized'] },
-      { internalName: 'Finalized', type: 'Boolean', displayName: 'Finalized (Legacy)', governance: 'allow', isSilent: true, candidates: ['Finalized'] },
-      { internalName: 'Note', type: 'Note', displayName: 'Note' },
-    ],
-  },
-  {
-    key: 'transport_log',
-    displayName: '送迎ステータスログ',
-    resolve: () => envOr('VITE_SP_LIST_TRANSPORT_LOG', fromConfig(ListKeys.TransportLog)),
-    operations: ['R', 'W'],
-    category: 'attendance',
-    lifecycle: 'required',
-    essentialFields: ['UserCode', 'RecordDate', 'Direction', 'Status'],
-    provisioningFields: [
-      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true, candidates: ['UserCode', 'User_x0020_Code'] },
-      { internalName: 'RecordDate', type: 'DateTime', displayName: 'Record Date', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['RecordDate', 'Record_x0020_Date'] },
-      { internalName: 'Direction', type: 'Choice', displayName: 'Direction', choices: ['to', 'from'], required: true },
-      { internalName: 'Status', type: 'Choice', displayName: 'Status', choices: ['pending', 'in-progress', 'arrived', 'absent', 'self'], required: true },
-      { internalName: 'Method', type: 'Choice', displayName: 'Method', choices: ['facility-vehicle', 'family', 'taxi', 'walk', 'self', 'other'] },
-      { internalName: 'ScheduledTime', type: 'Text', displayName: 'Scheduled Time', candidates: ['ScheduledTime', 'Scheduled_x0020_Time'] },
-      { internalName: 'ActualTime', type: 'Text', displayName: 'Actual Time', candidates: ['ActualTime', 'Actual_x0020_Time'] },
-      { internalName: 'DriverName', type: 'Text', displayName: 'Driver Name', candidates: ['DriverName', 'Driver_x0020_Name'] },
-      { internalName: 'Notes', type: 'Note', displayName: 'Notes', candidates: ['Notes', 'Note'] },
-      { internalName: 'UpdatedBy', type: 'Text', displayName: 'Updated By' },
-      { internalName: 'UpdatedAt', type: 'DateTime', displayName: 'Updated At', dateTimeFormat: 'DateTime' },
-    ],
-  },
-];
-
-export const scheduleListEntries: readonly SpListEntry[] = [
-// ── 4. スケジュール系 ──────────────────────────────────
-  {
-    key: 'schedule_events',
-    displayName: 'スケジュール',
-    resolve: () => envOr('VITE_SP_LIST_SCHEDULES', fromConfig(ListKeys.Schedules)),
-    operations: ['R', 'W', 'D'],
-    category: 'schedule',
-    essentialFields: ['Title', 'EventDate', 'EndDate'],
-    provisioningFields: [
-      { internalName: 'EventDate', type: 'DateTime', displayName: 'Start Time', required: true, dateTimeFormat: 'DateTime' },
-      { internalName: 'EndDate', type: 'DateTime', displayName: 'End Time', required: true, dateTimeFormat: 'DateTime' },
-      { internalName: 'Status', type: 'Choice', displayName: 'Status', choices: ['Planned', 'Postponed', 'Cancelled'], candidates: ['Status', 'cr014_status'] },
-      { internalName: 'ServiceType', type: 'Text', displayName: 'Service Type', candidates: ['ServiceType', 'Category', 'cr014_serviceType'] },
-      { internalName: 'TargetUserId', type: 'Text', displayName: 'User ID', candidates: ['TargetUserId', 'TargetUser', 'UserCode', 'cr013_usercode', 'cr013_personId', 'cr014_personId', 'UserId', 'UserID'] },
-      { internalName: 'cr014_personId', type: 'Text', displayName: 'Person ID (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_personId'] },
-      { internalName: 'AssignedStaffId', type: 'Text', displayName: 'Staff ID' },
-      { internalName: 'UserName', type: 'Text', displayName: 'User Name', isSilent: true, candidates: ['cr014_personName', 'UserName', 'PersonName'] },
-      { internalName: 'RowKey', type: 'Text', displayName: 'Row Key', candidates: ['RowKey', 'cr014_rowKey'] },
-      { internalName: 'DayKey', type: 'Text', displayName: 'Day Key', isSilent: true, candidates: ['DayKey', 'cr014_dayKey'] },
-      { internalName: 'MonthKey', type: 'Text', displayName: 'Month Key', isSilent: true, candidates: ['MonthKey', 'cr014_monthKey'] },
-      { internalName: 'FiscalYear', type: 'Text', displayName: 'Fiscal Year', isSilent: true, candidates: ['FiscalYear', 'cr014_fiscalYear'] },
-      { internalName: 'LocationName', type: 'Text', displayName: 'Location Name', isSilent: true, candidates: ['LocationName', 'Location', 'cr014_locationName'] },
-      { internalName: 'Note', type: 'Note', displayName: 'Note', candidates: ['Note', 'Notes', 'cr014_note'] },
-      { internalName: 'OrgAudience', type: 'Text', displayName: 'Org Audience', isSilent: true, candidates: ['OrgAudience', 'cr014_orgAudience'] },
-      { internalName: 'CreatedAt', type: 'DateTime', displayName: 'Created At (Legacy)', governance: 'allow', isSilent: true, candidates: ['CreatedAt', 'Created'] },
-      { internalName: 'UpdatedAt', type: 'DateTime', displayName: 'Updated At (Legacy)', governance: 'allow', isSilent: true, candidates: ['UpdatedAt', 'Modified'] },
-      { internalName: 'Visibility', type: 'Choice', displayName: 'Visibility', choices: ['org', 'team', 'private'] },
-      { internalName: 'Category', type: 'Choice', displayName: 'Category (Legacy)', governance: 'allow', isSilent: true, candidates: ['Category', 'cr014_category', 'PersonType', 'cr014_personType'] },
-      { internalName: 'cr014_personType', type: 'Text', displayName: 'Person Type (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_personType'] },
-    ],
-    lifecycle: 'required',
-  },
-  {
-    key: 'support_procedure_results',
-    displayName: '支援手順結果 (ISP)',
-    resolve: () => envOr('VITE_SP_LIST_PROCEDURE_RESULTS', fromConfig(ListKeys.SupportProcedure_Results)),
-    operations: ['R', 'W'],
-    category: 'schedule',
-    lifecycle: 'optional',
-    essentialFields: ['ParentScheduleId', 'ResultDate'],
-    provisioningFields: [
-      { internalName: 'ParentScheduleId', type: 'Number', displayName: 'Parent Schedule ID', required: true, indexed: true, candidates: ['ParentScheduleId', 'Parent_x0020_Schedule_x0020_ID'] },
-      { internalName: 'ResultDate', type: 'DateTime', displayName: 'Result Date', required: true, dateTimeFormat: 'DateOnly', candidates: ['ResultDate', 'Result_x0020_Date'] },
-      { internalName: 'ResultStatus', type: 'Text', displayName: 'Result Status', isSilent: true, candidates: ['ResultStatus', 'Result_x0020_Status'] },
-      { internalName: 'ResultNote', type: 'Note', displayName: 'Result Note', richText: false, isSilent: true, candidates: ['ResultNote', 'Result_x0020_Note'] },
-      { internalName: 'StaffCode', type: 'Text', displayName: 'Staff Code', isSilent: true, candidates: ['StaffCode', 'Staff_x0020_Code'] },
-    ],
-  },
-  {
-    key: 'approval_logs',
-    displayName: '承認ログ',
-    resolve: () => envOr('VITE_SP_LIST_APPROVAL_LOGS', fromConfig(ListKeys.Approval_Logs)),
-    operations: ['R', 'W'],
-    category: 'schedule',
-    lifecycle: 'optional',
-    essentialFields: ['ParentScheduleId', 'ApprovedBy', 'ApprovedAt', 'ApprovalAction'],
-    provisioningFields: [
-      { internalName: 'ParentScheduleId', type: 'Number', displayName: 'Parent Schedule ID', required: true, indexed: true, candidates: ['ParentScheduleId', 'Parent_x0020_Schedule_x0020_ID'] },
-      { internalName: 'ApprovedBy', type: 'Text', displayName: 'Approved By', required: true, candidates: ['ApprovedBy', 'Approved_x0020_By'] },
-      { internalName: 'ApprovedAt', type: 'DateTime', displayName: 'Approved At', required: true, dateTimeFormat: 'DateTime', indexed: true, candidates: ['ApprovedAt', 'Approved_x0020_At'] },
-      { internalName: 'ApprovalNote', type: 'Note', displayName: 'Approval Note', richText: false, candidates: ['ApprovalNote', 'Approval_x0020_Note'] },
-      { internalName: 'ApprovalAction', type: 'Text', displayName: 'Approval Action', required: true, candidates: ['ApprovalAction', 'Approval_x0020_Action'] },
-    ],
-  },
-];
-
-export const meetingListEntries: readonly SpListEntry[] = [
-// ── 5. 会議系 ──────────────────────────────────────────
-  {
-    key: 'meeting_sessions',
-    displayName: '会議セッション',
-    resolve: () => envOr('VITE_SP_LIST_MEETING_SESSIONS', 'MeetingSessions'),
-    operations: ['R', 'W', 'D'],
-    category: 'meeting',
-    lifecycle: 'optional',
-    essentialFields: ['SessionKey', 'MeetingKind', 'MeetingDate'],
-    provisioningFields: [
-      { internalName: 'SessionKey', type: 'Text', displayName: 'Session Key', required: true, indexed: true, candidates: ['SessionKey', 'Session_x0020_Key'] },
-      { internalName: 'MeetingKind', type: 'Choice', displayName: 'Meeting Kind', choices: ['morning', 'evening'], required: true, candidates: ['MeetingKind', 'Meeting_x0020_Kind'] },
-      { internalName: 'MeetingDate', type: 'DateTime', displayName: 'Meeting Date', required: true, dateTimeFormat: 'DateOnly', candidates: ['MeetingDate', 'Meeting_x0020_Date'] },
-      { internalName: 'StartTime', type: 'Text', displayName: 'Start Time', candidates: ['StartTime', 'Start_x0020_Time'] },
-      { internalName: 'EndTime', type: 'Text', displayName: 'End Time', candidates: ['EndTime', 'End_x0020_Time'] },
-      { internalName: 'Session_x0020_Key', type: 'Text', displayName: 'Session Key (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Session_x0020_Key'] },
-      { internalName: 'Meeting_x0020_Kind', type: 'Text', displayName: 'Meeting Kind (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Meeting_x0020_Kind'] },
-      { internalName: 'Meeting_x0020_Date', type: 'DateTime', displayName: 'Meeting Date (Legacy SP-encoded)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['Meeting_x0020_Date'] },
-      { internalName: 'Start_x0020_Time', type: 'Text', displayName: 'Start Time (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Start_x0020_Time'] },
-      { internalName: 'End_x0020_Time', type: 'Text', displayName: 'End Time (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['End_x0020_Time'] },
-      { internalName: 'ChairpersonUserId', type: 'Text', displayName: 'Chairperson ID', isSilent: true, candidates: ['ChairpersonUserId', 'Chairperson_x0020_ID'] },
-      { internalName: 'ChairpersonName', type: 'Text', displayName: 'Chairperson Name', candidates: ['ChairpersonName', 'Chairperson_x0020_Name'] },
-      { internalName: 'Chairperson_x0020_Name', type: 'Text', displayName: 'Chairperson Name (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Chairperson_x0020_Name'] },
-      { internalName: 'Status', type: 'Choice', displayName: 'Status', choices: ['scheduled', 'in-progress', 'completed', 'cancelled'] },
-      { internalName: 'TotalParticipants', type: 'Number', displayName: 'Total Participants', candidates: ['TotalParticipants', 'Total_x0020_Participants'] },
-      { internalName: 'Total_x0020_Participants', type: 'Number', displayName: 'Total Participants (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Total_x0020_Participants'] },
-    ],
-  },
-  {
-    key: 'meeting_steps',
-    displayName: '会議ステップ',
-    resolve: () => envOr('VITE_SP_LIST_MEETING_STEPS', 'MeetingSteps'),
-    operations: ['R', 'W', 'D'],
-    category: 'meeting',
-    lifecycle: 'optional',
-    essentialFields: ['SessionId', 'StepId', 'StepTitle'],
-    provisioningFields: [
-      { internalName: 'SessionId', type: 'Number', displayName: 'Session ID', required: true, indexed: true, isSilent: true, candidates: ['SessionId', 'Session_x0020_ID'] },
-      { internalName: 'SessionKey', type: 'Text', displayName: 'Session Key', required: true, isSilent: true, candidates: ['SessionKey', 'Session_x0020_Key'] },
-      { internalName: 'StepId', type: 'Number', displayName: 'Step ID', required: true, isSilent: true, candidates: ['StepId', 'Step_x0020_ID'] },
-      { internalName: 'StepTitle', type: 'Text', displayName: 'Step Title', required: true, isSilent: true, candidates: ['StepTitle', 'Step_x0020_Title'] },
-      { internalName: 'Completed', type: 'Boolean', displayName: 'Completed', default: false },
-      { internalName: 'TimeSpentMinutes', type: 'Number', displayName: 'Minutes Spent', isSilent: true, candidates: ['TimeSpentMinutes', 'Minutes_x0020_Spent'] },
-      { internalName: 'StepNotes', type: 'Note', displayName: 'Step Notes', richText: false, isSilent: true, candidates: ['StepNotes', 'Step_x0020_Notes'] },
-      { internalName: 'Session_x0020_ID', type: 'Number', displayName: 'Session ID (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Session_x0020_ID'] },
-      { internalName: 'Session_x0020_Key', type: 'Text', displayName: 'Session Key (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Session_x0020_Key'] },
-      { internalName: 'Step_x0020_ID', type: 'Number', displayName: 'Step ID (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Step_x0020_ID'] },
-      { internalName: 'Step_x0020_Title', type: 'Text', displayName: 'Step Title (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Step_x0020_Title'] },
-      { internalName: 'Step_x0020_Notes', type: 'Note', displayName: 'Step Notes (Legacy SP-encoded)', richText: false, governance: 'allow', isSilent: true, candidates: ['Step_x0020_Notes'] },
-    ],
-  },
-  {
-    key: 'meeting_minutes',
-    displayName: '議事録',
-    resolve: () => envOr('VITE_SP_LIST_MEETING_MINUTES', fromConfig(ListKeys.MeetingMinutes)),
-    operations: ['R', 'W', 'D'],
-    category: 'meeting',
-    lifecycle: 'required',
-    essentialFields: ['MeetingDate', 'Category'],
-    provisioningFields: [
-      { internalName: 'MeetingDate', type: 'DateTime', displayName: 'Meeting Date', required: true, dateTimeFormat: 'DateOnly', indexed: true, candidates: ['MeetingDate', 'meetingDate', 'cr013_meetingDate', 'Date'] },
-      { internalName: 'Category', type: 'Text', displayName: 'Category', required: true, candidates: ['Category', 'category', 'cr013_category', 'MeetingCategory'] },
-      { internalName: 'Summary', type: 'Note', displayName: 'Summary', candidates: ['Summary', 'summary', 'cr013_summary', 'Notes'] },
-      { internalName: 'Decisions', type: 'Note', displayName: 'Decisions', candidates: ['Decisions', 'decisions', 'cr013_decisions', 'DecisionSummary'] },
-      { internalName: 'Actions', type: 'Note', displayName: 'Actions', candidates: ['Actions', 'actions', 'cr013_actions', 'ActionItems'] },
-      { internalName: 'Attendees', type: 'Note', displayName: 'Attendees', candidates: ['Attendees', 'attendees', 'cr013_attendees'] },
-      { internalName: 'StaffAttendance', type: 'Note', displayName: 'Staff Attendance', candidates: ['StaffAttendance', 'staffAttendance', 'cr013_staffAttendance'] },
-      { internalName: 'Tags', type: 'Text', displayName: 'Tags (Legacy)', governance: 'allow', isSilent: true, candidates: ['Tags'] },
-      { internalName: 'UserHealthNotes', type: 'Note', displayName: 'User Health Notes (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['UserHealthNotes'] },
-      { internalName: 'RelatedLinks', type: 'Note', displayName: 'Related Links (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['RelatedLinks'] },
-      { internalName: 'Chair', type: 'Text', displayName: 'Chair (Legacy)', governance: 'allow', isSilent: true, candidates: ['Chair'] },
-      { internalName: 'Scribe', type: 'Text', displayName: 'Scribe (Legacy)', governance: 'allow', isSilent: true, candidates: ['Scribe'] },
-      { internalName: 'IsPublished', type: 'Boolean', displayName: 'Published', default: false, candidates: ['IsPublished', 'isPublished', 'cr013_isPublished', 'Published'] },
-    ],
-  },
-  {
-    key: 'monitoring_meetings',
-    displayName: 'モニタリング会議',
-    resolve: () => envOr('VITE_SP_LIST_MONITORING_MEETINGS', 'MonitoringMeetings'),
-    operations: ['R', 'W', 'D'],
-    category: 'meeting',
-    lifecycle: 'required',
-    essentialFields: ['cr014_recordId', 'cr014_userId', 'cr014_meetingDate'],
-    provisioningFields: [
-      { internalName: 'cr014_recordId', type: 'Text', displayName: 'Record ID', required: true, indexed: true, candidates: ['cr014_recordId', 'RecordId', 'recordId', 'Record_x0020_Id', 'Title'] },
-      { internalName: 'cr014_userId', type: 'Text', displayName: 'User ID', required: true, indexed: true, candidates: ['cr014_userId', 'UserId', 'UserID', 'UserCode', 'userId', 'User_x0020_Id', 'cr013_userId'] },
-      { internalName: 'cr014_ispId', type: 'Text', displayName: 'ISP ID', required: true, candidates: ['cr014_ispId', 'IspId', 'ISPId', 'ispId'] },
-      { internalName: 'cr014_planningSheetId', type: 'Text', displayName: 'Planning Sheet ID', candidates: ['cr014_planningSheetId', 'PlanningSheetId', 'planningSheetId'] },
-      { internalName: 'cr014_meetingType', type: 'Text', displayName: 'Meeting Type', candidates: ['cr014_meetingType', 'MeetingType', 'meetingType'] },
-      { internalName: 'cr014_status', type: 'Text', displayName: 'Status', candidates: ['cr014_status', 'Status', 'status'] },
-      { internalName: 'cr014_meetingDate', type: 'DateTime', displayName: 'Meeting Date', required: true, dateTimeFormat: 'DateOnly', candidates: ['cr014_meetingDate', 'MeetingDate', 'Meeting_x0020_Date', 'meetingDate', 'Date', 'RecordDate', 'cr013_date'] },
-      { internalName: 'cr014_venue', type: 'Text', displayName: 'Venue', candidates: ['cr014_venue', 'Venue', 'venue'] },
-      { internalName: 'cr014_attendeesJson', type: 'Note', displayName: 'Attendees (JSON)', richText: false, candidates: ['cr014_attendeesJson', 'AttendeesJson', 'Attendees', 'attendeesJson'] },
-      { internalName: 'cr014_goalEvaluationsJson', type: 'Note', displayName: 'Goal Evaluations (JSON)', richText: false, candidates: ['cr014_goalEvaluationsJson', 'GoalEvaluationsJson', 'GoalEvaluations', 'goalEvaluationsJson'] },
-      { internalName: 'cr014_overallAssessment', type: 'Note', displayName: 'Overall Assessment', richText: false, candidates: ['cr014_overallAssessment', 'OverallAssessment', 'overallAssessment'] },
-      { internalName: 'cr014_userFeedback', type: 'Note', displayName: 'User Feedback', richText: false, candidates: ['cr014_userFeedback', 'UserFeedback', 'userFeedback'] },
-      { internalName: 'cr014_familyFeedback', type: 'Note', displayName: 'Family Feedback', richText: false, candidates: ['cr014_familyFeedback', 'FamilyFeedback', 'familyFeedback'] },
-      { internalName: 'cr014_planChangeDecision', type: 'Text', displayName: 'Plan Change Decision', candidates: ['cr014_planChangeDecision', 'PlanChangeDecision', 'planChangeDecision'] },
-      { internalName: 'cr014_changeReason', type: 'Note', displayName: 'Change Reason', richText: false, candidates: ['cr014_changeReason', 'ChangeReason', 'changeReason'] },
-      { internalName: 'cr014_decisionsJson', type: 'Note', displayName: 'Decisions (JSON)', richText: false, candidates: ['cr014_decisionsJson', 'DecisionsJson', 'Decisions', 'decisionsJson'] },
-      { internalName: 'cr014_nextMonitoringDate', type: 'DateTime', displayName: 'Next Monitoring Date', dateTimeFormat: 'DateOnly', candidates: ['cr014_nextMonitoringDate', 'NextMonitoringDate', 'Next_x0020_Monitoring_x0020_Date', 'nextMonitoringDate'] },
-      { internalName: 'cr014_recordedBy', type: 'Text', displayName: 'Recorded By', candidates: ['cr014_recordedBy', 'RecordedBy', 'recordedBy'] },
-      { internalName: 'cr014_recordedAt', type: 'DateTime', displayName: 'Recorded At', candidates: ['cr014_recordedAt', 'RecordedAt', 'recordedAt'] },
-      { internalName: 'cr014_implSummary', type: 'Note', displayName: 'Implementation Summary (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['cr014_implSummary'] },
-      { internalName: 'cr014_behaviorChange', type: 'Note', displayName: 'Behavior Change (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['cr014_behaviorChange'] },
-      { internalName: 'cr014_effSupport', type: 'Note', displayName: 'Effective Support (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['cr014_effSupport'] },
-      { internalName: 'cr014_issueSummary', type: 'Note', displayName: 'Issue Summary (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['cr014_issueSummary'] },
-      { internalName: 'cr014_discussionSummary', type: 'Note', displayName: 'Discussion Summary (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['cr014_discussionSummary'] },
-      { internalName: 'cr014_reqPlanSheetUpd', type: 'Boolean', displayName: 'Request Plan Sheet Update (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_reqPlanSheetUpd'] },
-      { internalName: 'cr014_reqIspUpd', type: 'Boolean', displayName: 'Request ISP Update (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_reqIspUpd'] },
-      { internalName: 'cr014_hasBasicTrained', type: 'Boolean', displayName: 'Has Basic Trained (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_hasBasicTrained'] },
-      { internalName: 'cr014_hasPractTrained', type: 'Boolean', displayName: 'Has Practical Trained (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_hasPractTrained'] },
-      { internalName: 'cr014_qualCheckStatus', type: 'Text', displayName: 'Qualification Check Status (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_qualCheckStatus'] },
-      { internalName: 'cr014_nextActions', type: 'Note', displayName: 'Next Actions (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['cr014_nextActions'] },
-      { internalName: 'cr014_finalizedAt', type: 'DateTime', displayName: 'Finalized At (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_finalizedAt'] },
-      { internalName: 'cr014_finalizedBy', type: 'Text', displayName: 'Finalized By (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_finalizedBy'] },
-      { internalName: 'cr014_prevMeetingId', type: 'Text', displayName: 'Previous Meeting ID (Legacy)', governance: 'allow', isSilent: true, candidates: ['cr014_prevMeetingId'] },
-    ],
-  },
-];
-
-export const handoffListEntries: readonly SpListEntry[] = [
-// ── 6. 引き継ぎ・支援計画系 ────────────────────────────
-  {
-    key: 'handoff',
-    displayName: '引き継ぎ',
-    resolve: () => envOr('VITE_SP_HANDOFF_LIST_TITLE', 'Handoff'),
-    operations: ['R', 'W', 'D'],
-    category: 'handoff',
-    lifecycle: 'required',
-    essentialFields: ['Message', 'UserCode', 'Category'],
-    provisioningFields: [
-      { internalName: 'Message', type: 'Note', displayName: 'Message', required: true },
-      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true },
-      { internalName: 'Category', type: 'Text', displayName: 'Category', required: true },
-      { internalName: 'Severity', type: 'Text', displayName: 'Severity' },
-      { internalName: 'Status', type: 'Text', displayName: 'Status' },
-      { internalName: 'TimeBand', type: 'Text', displayName: 'Time Band' },
-      { internalName: 'SourceType', type: 'Text', displayName: 'Source Type' },
-      { internalName: 'SourceKey', type: 'Text', displayName: 'Source Key' },
-      { internalName: 'UserDisplayName', type: 'Text', displayName: 'User Display Name', governance: 'allow', isSilent: true, candidates: ['UserDisplayName'] },
-      { internalName: 'CreatedAt', type: 'DateTime', displayName: 'Created At', governance: 'allow', isSilent: true, candidates: ['CreatedAt'] },
-      { internalName: 'CreatedByName', type: 'Text', displayName: 'Created By Name', governance: 'allow', isSilent: true, candidates: ['CreatedByName'] },
-      { internalName: 'IsDraft', type: 'Boolean', displayName: 'Is Draft', governance: 'allow', isSilent: true, candidates: ['IsDraft'] },
-      { internalName: 'MeetingSessionKey', type: 'Text', displayName: 'Meeting Session Key', governance: 'allow', isSilent: true, candidates: ['MeetingSessionKey'] },
-      { internalName: 'SourceId', type: 'Text', displayName: 'Source ID', governance: 'allow', isSilent: true, candidates: ['SourceId'] },
-      { internalName: 'SourceUrl', type: 'Text', displayName: 'Source URL', governance: 'allow', isSilent: true, candidates: ['SourceUrl'] },
-      { internalName: 'SourceLabel', type: 'Text', displayName: 'Source Label', governance: 'allow', isSilent: true, candidates: ['SourceLabel'] },
-      { internalName: 'CarryOverDate', type: 'DateTime', displayName: 'Carry Over Date', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['CarryOverDate'] },
-      { internalName: 'CreatedBy', type: 'Text', displayName: 'Created By (Legacy)', governance: 'allow', isSilent: true, candidates: ['CreatedBy'] },
-      { internalName: 'ModifiedBy', type: 'Text', displayName: 'Modified By (Legacy)', governance: 'allow', isSilent: true, candidates: ['ModifiedBy'] },
-      { internalName: 'ModifiedAt', type: 'DateTime', displayName: 'Modified At (Legacy)', governance: 'allow', isSilent: true, candidates: ['ModifiedAt'] },
-    ],
-  },
-  {
-    key: 'support_templates',
-    displayName: '支援手順テンプレート',
-    resolve: () => envOr('VITE_SP_LIST_SUPPORT_TEMPLATES', fromConfig(ListKeys.SupportTemplates)),
-    operations: ['R'],
-    category: 'handoff',
-    lifecycle: 'required',
-    essentialFields: ['UserCode0', 'RowNo0', 'TimeSlot0', 'Activity0'],
-    provisioningFields: [
-      { internalName: 'UserCode0', type: 'Text', displayName: '利用者コード', indexed: true, candidates: ['UserCode0', 'UserCode'] },
-      { internalName: 'RowNo0', type: 'Number', displayName: '行番号', candidates: ['RowNo0', 'RowNo'] },
-      { internalName: 'TimeSlot0', type: 'Text', displayName: '時間帯', candidates: ['TimeSlot0', 'TimeSlot'] },
-      { internalName: 'Activity0', type: 'Text', displayName: '活動内容', candidates: ['Activity0', 'Activity'] },
-      { internalName: 'PersonManual0', type: 'Note', displayName: '本人手順', candidates: ['PersonManual0', 'PersonManual'] },
-      { internalName: 'SupporterManual0', type: 'Note', displayName: '支援者手順', candidates: ['SupporterManual0', 'SupporterManual'] },
-      { internalName: 'UserID', type: 'Text', displayName: 'User ID (Legacy)', governance: 'allow', isSilent: true },
-      { internalName: 'IsActive', type: 'Boolean', displayName: 'Is Active (Legacy)', governance: 'allow', isSilent: true },
-      { internalName: 'modifiedBy', type: 'Text', displayName: 'Modified By (Legacy)', governance: 'allow', isSilent: true, candidates: ['modifiedBy'] },
-      { internalName: 'Note', type: 'Note', displayName: 'Note (Legacy)', governance: 'allow', isSilent: true },
-      { internalName: 'Notes', type: 'Note', displayName: 'Notes (Legacy)', governance: 'allow', isSilent: true },
-    ],
-  },
-  {
-    key: 'plan_goals',
-    displayName: '支援計画目標',
-    resolve: () => envOr('VITE_SP_LIST_PLAN_GOAL', fromConfig(ListKeys.PlanGoal)),
-    operations: ['R', 'W'],
-    category: 'handoff',
-    lifecycle: 'required',
-    essentialFields: ['UserCode', 'GoalType', 'GoalText', 'PlanStatus'],
-    provisioningFields: [
-      { internalName: 'UserCode', type: 'Text', displayName: '利用者コード', required: true, indexed: true, candidates: ['UserCode', 'cr013_usercode', 'UserID'] },
-      { internalName: 'GoalType', type: 'Text', displayName: '目標種別', required: true, candidates: ['GoalType', 'cr013_goaltype'] },
-      { internalName: 'GoalLabel', type: 'Text', displayName: '目標ラベル', isSilent: true, candidates: ['GoalLabel', 'cr013_goallabel', 'Title'] },
-      { internalName: 'GoalText', type: 'Note', displayName: '目標内容', richText: false, candidates: ['GoalText', 'cr013_goaltext'] },
-      { internalName: 'Domains', type: 'Text', displayName: '領域', isSilent: true, candidates: ['Domains', 'cr013_domains'] },
-      { internalName: 'PlanPeriod', type: 'Text', displayName: '計画期間', isSilent: true, candidates: ['PlanPeriod', 'cr013_planperiod'] },
-      { internalName: 'PlanStatus', type: 'Text', displayName: '計画状態', candidates: ['PlanStatus', 'cr013_planstatus'] },
-      { internalName: 'CertExpiry', type: 'Text', displayName: '受給者証期限', isSilent: true, candidates: ['CertExpiry', 'cr013_certexpiry'] },
-      { internalName: 'SortOrder', type: 'Number', displayName: '表示順', isSilent: true, candidates: ['SortOrder', 'cr013_sortorder'] },
-    ],
-  },
-  {
-    key: 'support_plans',
-    displayName: '個別支援計画',
-    resolve: () => envOr('VITE_SP_LIST_SUPPORT_PLANS', fromConfig(ListKeys.SupportPlans)),
-    operations: ['R', 'W', 'D'],
-    category: 'handoff',
-    lifecycle: 'required', // 制度上必須のため required
-    essentialFields: ['DraftId', 'UserCode', 'FormDataJson'],
-    provisioningFields: [
-      { internalName: 'DraftId', type: 'Text', displayName: 'Draft ID', required: true, indexed: true },
-      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true },
-      { internalName: 'DraftName', type: 'Text', displayName: 'Draft Name' },
-      { internalName: 'FormDataJson', type: 'Note', displayName: 'Form Data (JSON)', required: true, richText: false },
-      { internalName: 'Status', type: 'Text', displayName: 'Status' },
-      { internalName: 'SchemaVersion', type: 'Number', displayName: 'Schema Version', default: 2 },
-      { internalName: 'UserCode0', type: 'Text', displayName: 'User Code (Legacy)', governance: 'allow', isSilent: true, candidates: ['UserCode0'] },
-      { internalName: 'Status0', type: 'Text', displayName: 'Status (Legacy)', governance: 'allow', isSilent: true, candidates: ['Status0'] },
-    ],
-  },
-  {
-    key: 'iceberg_pdca',
-    displayName: '氷山モデルPDCA',
-    resolve: () => envOr('VITE_SP_LIST_ICEBERG_PDCA', fromConfig(ListKeys.IcebergPdca)),
-    operations: ['R', 'W', 'D'],
-    category: 'handoff',
-    lifecycle: 'optional',
-    essentialFields: ['Title', 'UserID0', 'Phase0'],
-    provisioningFields: [
-      { internalName: 'Title', type: 'Text', displayName: 'Title', required: true },
-      { internalName: 'UserID0', type: 'Text', displayName: 'User ID', required: true, indexed: true, candidates: ['UserID0', 'userId', 'UserID', 'UserCode'] },
-      { internalName: 'PlanningSheetId', type: 'Text', displayName: 'Planning Sheet ID' },
-      { internalName: 'Summary0', type: 'Note', displayName: 'Summary', richText: false, candidates: ['Summary0', 'summary'] },
-      { internalName: 'Phase0', type: 'Text', displayName: 'Phase', candidates: ['Phase0', 'phase'] },
-      { internalName: 'userId', type: 'Text', displayName: 'User ID (Legacy)', governance: 'allow', isSilent: true },
-      { internalName: 'summary', type: 'Note', displayName: 'Summary (Legacy)', governance: 'allow', richText: false, isSilent: true },
-      { internalName: 'phase', type: 'Text', displayName: 'Phase (Legacy)', governance: 'allow', isSilent: true },
-    ],
-  },
-  {
-    key: 'iceberg_analysis',
-    displayName: '氷山モデル分析',
-    resolve: () => envOr('VITE_SP_LIST_ICEBERG_ANALYSIS', fromConfig(ListKeys.IcebergAnalysis)),
-    operations: ['R', 'W'],
-    category: 'handoff',
-    lifecycle: 'optional',
-    essentialFields: ['Title', 'EntryHash', 'UserId', 'PayloadJson'],
-    provisioningFields: [
-      { internalName: 'Title', type: 'Text', displayName: 'Title', required: true },
-      { internalName: 'EntryHash', type: 'Text', displayName: 'Entry Hash', required: true, indexed: true },
-      { internalName: 'SessionId', type: 'Text', displayName: 'Session ID', indexed: true },
-      { internalName: 'UserId', type: 'Text', displayName: 'User ID', indexed: true },
-      { internalName: 'PayloadJson', type: 'Note', displayName: 'Payload JSON', required: true, richText: false },
-      { internalName: 'SchemaVersion', type: 'Number', displayName: 'Schema Version' },
-      { internalName: 'UpdatedAt', type: 'DateTime', displayName: 'Updated At', indexed: true },
-    ],
-  },
-  {
-    key: 'isp_master',
-    displayName: '個別支援計画（ISP）',
-    resolve: () => envOr('VITE_SP_LIST_ISP_MASTER', fromConfig(ListKeys.IspMaster)),
-    operations: ['R', 'W'],
-    category: 'handoff',
-    lifecycle: 'required', // 三層モデルの核となるため required
-    essentialFields: ['UserCode', 'PlanStartDate', 'Status'],
-    provisioningFields: [
-      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true, candidates: ['UserCode', 'UserID', 'User_ID', 'cr013_userCode'] },
-      { internalName: 'PlanStartDate', type: 'DateTime', displayName: 'Plan Start Date', required: true, dateTimeFormat: 'DateOnly', candidates: ['PlanStartDate', 'StartDate', 'cr013_planStartDate'] },
-      { internalName: 'PlanEndDate', type: 'DateTime', displayName: 'Plan End Date', dateTimeFormat: 'DateOnly', candidates: ['PlanEndDate', 'EndDate', 'cr013_planEndDate'] },
-      { internalName: 'Status', type: 'Text', displayName: 'Status', required: true, candidates: ['Status', 'UsageStatus', 'cr013_status'] },
-      { internalName: 'NextReviewAt', type: 'DateTime', displayName: 'Next Review At (Legacy)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['NextReviewAt'] },
-      { internalName: 'UserIntent', type: 'Note', displayName: 'User Intent (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['UserIntent'] },
-      { internalName: 'Precautions', type: 'Note', displayName: 'Precautions (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['Precautions'] },
-      { internalName: 'FamilyIntent', type: 'Note', displayName: 'Family Intent (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['FamilyIntent'] },
-      { internalName: 'OverallSupportPolicy', type: 'Note', displayName: 'Overall Support Policy (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['OverallSupportPolicy'] },
-      { internalName: 'QolIssues', type: 'Note', displayName: 'QOL Issues (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['QolIssues'] },
-      { internalName: 'LongTermGoalsJson', type: 'Note', displayName: 'Long Term Goals JSON (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['LongTermGoalsJson'] },
-      { internalName: 'ShortTermGoalsJson', type: 'Note', displayName: 'Short Term Goals JSON (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['ShortTermGoalsJson'] },
-      { internalName: 'SupportSummary', type: 'Note', displayName: 'Support Summary (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['SupportSummary'] },
-      { internalName: 'MonitoringSummary', type: 'Note', displayName: 'Monitoring Summary (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['MonitoringSummary'] },
-      { internalName: 'ConsentAt', type: 'DateTime', displayName: 'Consent At (Legacy)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['ConsentAt'] },
-      { internalName: 'DeliveredAt', type: 'DateTime', displayName: 'Delivered At (Legacy)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['DeliveredAt'] },
-      { internalName: 'LastMonitoringAt', type: 'DateTime', displayName: 'Last Monitoring At (Legacy)', dateTimeFormat: 'DateOnly', governance: 'allow', isSilent: true, candidates: ['LastMonitoringAt'] },
-      { internalName: 'VersionNo', type: 'Number', displayName: 'Version No', default: 1, governance: 'allow' },
-      { internalName: 'IsCurrent', type: 'Boolean', displayName: 'Is Current', default: true },
-      { internalName: 'FormDataJson', type: 'Note', displayName: 'Form Data (JSON)', richText: false },
-      { internalName: 'UserSnapshotJson', type: 'Note', displayName: 'User Snapshot (JSON)', richText: false },
-    ],
-  },
-  {
-    key: 'planning_sheet_master',
-    displayName: '支援計画シート',
-    resolve: () => envOr('VITE_SP_LIST_PLANNING_SHEET', fromConfig(ListKeys.PlanningSheetMaster)),
-    operations: ['R', 'W'],
-    category: 'handoff',
-    lifecycle: 'optional',
-    essentialFields: ['Status', 'UserCode', 'UserId'],
-    provisioningFields: [
-      { internalName: 'Status', type: 'Text', displayName: 'Status', required: true, candidates: ['Status', 'UsageStatus', 'cr013_status'] },
-      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', required: true, indexed: true, candidates: ['UserCode', 'User_ID', 'cr013_userCode'] },
-      { internalName: 'UserId', type: 'Text', displayName: 'User ID', required: true, indexed: true, candidates: ['UserId', 'UserID', 'cr013_userId'] },
-      { internalName: 'DecisionStatus', type: 'Text', displayName: 'Decision Status', candidates: ['DecisionStatus'] },
-      { internalName: 'IsCurrent', type: 'Boolean', displayName: 'Is Current', default: true },
-      { internalName: 'ISPId', type: 'Number', displayName: 'ISP ID', candidates: ['ISPId', 'ISP_x0020_ID', 'cr013_ispId'] },
-      { internalName: 'TargetScene', type: 'Text', displayName: 'Target Scene', candidates: ['TargetScene', 'Scene', 'cr013_targetScene'] },
-      { internalName: 'TargetDomain', type: 'Text', displayName: 'Target Domain', candidates: ['TargetDomain', 'Domain', 'cr013_targetDomain'] },
-      { internalName: 'ObservationFacts', type: 'Note', displayName: 'Observation Facts', richText: false, candidates: ['ObservationFacts', 'cr013_observationFacts'] },
-      { internalName: 'CollectedInformation', type: 'Note', displayName: 'Collected Information', richText: false, candidates: ['CollectedInformation', 'cr013_collectedInformation'] },
-      { internalName: 'InterpretationHypothesis', type: 'Note', displayName: 'Interpretation Hypothesis', richText: false, candidates: ['InterpretationHypothesis', 'cr013_interpretationHypothesis'] },
-      { internalName: 'SupportIssues', type: 'Note', displayName: 'Support Issues', richText: false, candidates: ['SupportIssues', 'cr013_supportIssues'] },
-      { internalName: 'SupportPolicy', type: 'Note', displayName: 'Support Policy', richText: false, candidates: ['SupportPolicy', 'cr013_supportPolicy'] },
-      { internalName: 'EnvironmentalAdjustments', type: 'Note', displayName: 'Environmental Adjustments', richText: false, candidates: ['EnvironmentalAdjustments', 'cr013_environmentalAdjustments'] },
-      { internalName: 'ConcreteApproaches', type: 'Note', displayName: 'Concrete Approaches', richText: false, candidates: ['ConcreteApproaches', 'cr013_concreteApproaches'] },
-      { internalName: 'AppliedFrom', type: 'DateTime', displayName: 'Applied From', dateTimeFormat: 'DateOnly', candidates: ['AppliedFrom', 'cr013_appliedFrom'] },
-      { internalName: 'NextReviewAt', type: 'DateTime', displayName: 'Next Review At', dateTimeFormat: 'DateOnly', candidates: ['NextReviewAt', 'cr013_nextReviewAt'] },
-      { internalName: 'AuthoredByStaffId', type: 'Text', displayName: 'Authored By Staff ID', candidates: ['AuthoredByStaffId', 'cr013_authoredByStaffId'] },
-      { internalName: 'AuthoredByQualification', type: 'Text', displayName: 'Authored By Qualification', candidates: ['AuthoredByQualification', 'cr013_authoredByQualification'] },
-      { internalName: 'AuthoredAt', type: 'DateTime', displayName: 'Authored At', dateTimeFormat: 'DateTime', candidates: ['AuthoredAt', 'cr013_authoredAt'] },
-      { internalName: 'ApplicableServiceType', type: 'Text', displayName: 'Applicable Service Type', candidates: ['ApplicableServiceType', 'cr013_applicableServiceType'] },
-      { internalName: 'ApplicableAddOnTypesJson', type: 'Note', displayName: 'Applicable Add-on Types JSON', richText: false, candidates: ['ApplicableAddOnTypesJson', 'cr013_applicableAddOnTypesJson'] },
-      { internalName: 'DeliveredToUserAt', type: 'DateTime', displayName: 'Delivered To User At', dateTimeFormat: 'DateOnly', candidates: ['DeliveredToUserAt', 'cr013_deliveredToUserAt'] },
-      { internalName: 'ReviewedAt', type: 'DateTime', displayName: 'Reviewed At', dateTimeFormat: 'DateOnly', candidates: ['ReviewedAt', 'cr013_reviewedAt'] },
-      { internalName: 'HasMedicalCoordination', type: 'Boolean', displayName: 'Has Medical Coordination', candidates: ['HasMedicalCoordination', 'cr013_hasMedicalCoordination'] },
-      { internalName: 'HasEducationCoordination', type: 'Boolean', displayName: 'Has Education Coordination', candidates: ['HasEducationCoordination', 'cr013_hasEducationCoordination'] },
-      { internalName: 'RegulatoryBasisSnapshotJson', type: 'Note', displayName: 'Regulatory Basis Snapshot JSON', richText: false, candidates: ['RegulatoryBasisSnapshotJson', 'cr013_regulatoryBasisSnapshotJson'] },
-      { internalName: 'IntakeJson', type: 'Note', displayName: 'Intake JSON', richText: false, candidates: ['IntakeJson', 'cr013_intakeJson'] },
-      { internalName: 'AssessmentJson', type: 'Note', displayName: 'Assessment JSON', richText: false, candidates: ['AssessmentJson', 'cr013_assessmentJson'] },
-      { internalName: 'PlanningJson', type: 'Note', displayName: 'Planning JSON', richText: false, candidates: ['PlanningJson', 'cr013_planningJson'] },
-      { internalName: 'GoalId', type: 'Text', displayName: 'Goal ID', candidates: ['GoalId'] },
-      { internalName: 'GoalLabel', type: 'Text', displayName: 'Goal Label', candidates: ['GoalLabel'] },
-      { internalName: 'DecisionNote', type: 'Note', displayName: 'Decision Note', richText: false, candidates: ['DecisionNote'] },
-      { internalName: 'DecisionBy', type: 'Text', displayName: 'Decision By', candidates: ['DecisionBy'] },
-      { internalName: 'DecisionAt', type: 'DateTime', displayName: 'Decision At', dateTimeFormat: 'DateTime', candidates: ['DecisionAt'] },
-      { internalName: 'RecommendationLevel', type: 'Text', displayName: 'Recommendation Level', candidates: ['RecommendationLevel'] },
-      { internalName: 'SnapshotJson', type: 'Note', displayName: 'Snapshot JSON', richText: false, candidates: ['SnapshotJson'] },
-      { internalName: 'FormDataJson', type: 'Note', displayName: 'Form Data (JSON)', richText: false, candidates: ['FormDataJson', 'cr013_formDataJson'] },
-    ],
-  },
-  {
-    key: 'behavior_monitoring_master',
-    displayName: '行動モニタリング記録',
-    resolve: () => envOr('VITE_SP_LIST_BEHAVIOR_MONITORING', fromConfig(ListKeys.BehaviorMonitoringMaster)),
-    operations: ['R', 'W'],
-    category: 'handoff',
-    lifecycle: 'required',
-    essentialFields: ['UserId', 'PlanningSheetId', 'PeriodEnd'],
-    provisioningFields: [
-      { internalName: 'UserId', type: 'Text', displayName: 'User ID', required: true, indexed: true, isSilent: true, candidates: ['UserId', 'UserCode', 'User_x0020_ID', 'cr013_userId'] },
-      { internalName: 'PlanningSheetId', type: 'Text', displayName: 'Planning Sheet ID', required: true, isSilent: true, candidates: ['PlanningSheetId', 'Planning_x0020_Sheet_x0020_ID', 'cr013_planningSheetId'] },
-      { internalName: 'PeriodStart', type: 'DateTime', displayName: 'Period Start', dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['PeriodStart', 'PeriodStart0', 'cr013_periodStart'] },
-      { internalName: 'PeriodEnd', type: 'DateTime', displayName: 'Period End', required: true, dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['PeriodEnd', 'Period_x0020_End', 'cr013_periodEnd'] },
-      { internalName: 'SupportEvaluationsJson', type: 'Note', displayName: 'Support Evaluations JSON', richText: false, isSilent: true, candidates: ['SupportEvaluationsJson', 'SupportEvaluationsJson0', 'cr013_supportEvaluationsJson'] },
-      { internalName: 'EnvironmentFindingsJson', type: 'Note', displayName: 'Environment Findings JSON', richText: false, isSilent: true, candidates: ['EnvironmentFindingsJson', 'EnvironmentFindingsJson0', 'cr013_environmentFindingsJson'] },
-      { internalName: 'EffectiveSupports', type: 'Note', displayName: 'Effective Supports', richText: false, isSilent: true, candidates: ['EffectiveSupports', 'EffectiveSupports0', 'cr013_effectiveSupports'] },
-      { internalName: 'DifficultiesObserved', type: 'Note', displayName: 'Difficulties Observed', richText: false, isSilent: true, candidates: ['DifficultiesObserved', 'DifficultiesObserved0', 'cr013_difficultiesObserved'] },
-      { internalName: 'NewTriggersJson', type: 'Note', displayName: 'New Triggers JSON', richText: false, isSilent: true, candidates: ['NewTriggersJson', 'NewTriggersJson0', 'cr013_newTriggersJson'] },
-      { internalName: 'MedicalSafetyNotes', type: 'Note', displayName: 'Medical Safety Notes', richText: false, isSilent: true, candidates: ['MedicalSafetyNotes', 'MedicalSafetyNotes0', 'cr013_medicalSafetyNotes'] },
-      { internalName: 'UserFeedback', type: 'Note', displayName: 'User Feedback', richText: false, isSilent: true, candidates: ['UserFeedback', 'UserFeedback0', 'cr013_userFeedback'] },
-      { internalName: 'FamilyFeedback', type: 'Note', displayName: 'Family Feedback', richText: false, isSilent: true, candidates: ['FamilyFeedback', 'FamilyFeedback0', 'cr013_familyFeedback'] },
-      { internalName: 'RecommendedChangesJson', type: 'Note', displayName: 'Recommended Changes JSON', richText: false, isSilent: true, candidates: ['RecommendedChangesJson', 'RecommendedChangesJson0', 'cr013_recommendedChangesJson'] },
-      { internalName: 'Summary', type: 'Note', displayName: 'Summary', richText: false, isSilent: true, candidates: ['Summary', 'Summary0', 'cr013_summary'] },
-      { internalName: 'RecordedBy', type: 'Text', displayName: 'Recorded By', isSilent: true, candidates: ['RecordedBy', 'RecordedBy0', 'cr013_recordedBy'] },
-      { internalName: 'RecordedAt', type: 'DateTime', displayName: 'Recorded At', isSilent: true, candidates: ['RecordedAt', 'RecordedAt0', 'cr013_recordedAt'] },
-    ],
-  },
-  {
-    key: 'planning_sheet_reassessment_master',
-    displayName: 'アセスメント・再アセスメント履歴',
-    resolve: () => envOr('VITE_SP_LIST_PLANNING_REASSESSMENT', fromConfig(ListKeys.PlanningSheetReassessmentMaster)),
-    operations: ['R', 'W'],
-    category: 'handoff',
-    lifecycle: 'required',
-    essentialFields: ['PlanningSheetId', 'ReassessmentDate'],
-    provisioningFields: [
-      { internalName: 'PlanningSheetId', type: 'Text', displayName: 'Planning Sheet ID', required: true, indexed: true, isSilent: true, candidates: ['PlanningSheetId', 'Planning_x0020_Sheet_x0020_ID', 'cr013_planningSheetId'] },
-      { internalName: 'UserId', type: 'Text', displayName: 'User ID', isSilent: true, candidates: ['UserId', 'UserCode', 'cr013_userId'] },
-      { internalName: 'ReassessmentTrigger', type: 'Text', displayName: 'Reassessment Trigger', isSilent: true, candidates: ['ReassessmentTrigger', 'ReassessmentTrigger0', 'cr013_reassessmentTrigger'] },
-      { internalName: 'ReassessmentDate', type: 'DateTime', displayName: 'Reassessment Date', required: true, dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['ReassessmentDate', 'Reassessment_x0020_Date', 'cr013_reassessmentDate'] },
-      { internalName: 'Summary', type: 'Note', displayName: 'Summary', richText: false, isSilent: true, candidates: ['Summary', 'Summary0', 'cr013_summary'] },
-      { internalName: 'CreatedByText', type: 'Text', displayName: 'Created By (Text)', isSilent: true, candidates: ['CreatedByText', 'CreatedBy', 'cr013_createdByText'] },
-      { internalName: 'CreatedAtText', type: 'Text', displayName: 'Created At (Text)', isSilent: true, candidates: ['CreatedAtText', 'CreatedAt', 'cr013_createdAtText'] },
-      { internalName: 'VersionNo', type: 'Number', displayName: 'Version No', governance: 'allow', isSilent: true, candidates: ['VersionNo', 'Version', 'cr013_versionNo'] },
-      { internalName: 'PlanChangeDecision', type: 'Text', displayName: 'Plan Change Decision', isSilent: true, candidates: ['PlanChangeDecision', 'PlanChangeDecision0', 'cr013_planChangeDecision'] },
-      { internalName: 'AbcSummary', type: 'Note', displayName: 'ABC Summary', richText: false, isSilent: true, candidates: ['AbcSummary', 'AbcSummary0', 'cr013_abcSummary'] },
-      { internalName: 'HypothesisReview', type: 'Note', displayName: 'Hypothesis Review', richText: false, isSilent: true, candidates: ['HypothesisReview', 'HypothesisReview0', 'cr013_hypothesisReview'] },
-      { internalName: 'ProcedureEffectiveness', type: 'Note', displayName: 'Procedure Effectiveness', richText: false, isSilent: true, candidates: ['ProcedureEffectiveness', 'ProcedureEffectiveness0', 'cr013_procedureEffectiveness'] },
-      { internalName: 'EnvironmentChange', type: 'Note', displayName: 'Environment Change', richText: false, isSilent: true, candidates: ['EnvironmentChange', 'EnvironmentChange0', 'cr013_environmentChange'] },
-      { internalName: 'NextReassessmentAt', type: 'DateTime', displayName: 'Next Reassessment At', dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['NextReassessmentAt', 'NextReassessmentAt0', 'cr013_nextReassessmentAt'] },
-      { internalName: 'Notes', type: 'Note', displayName: 'Notes', richText: false, isSilent: true, candidates: ['Notes', 'Notes0', 'cr013_notes'] },
-      { internalName: 'ReassessedBy', type: 'Text', displayName: 'Reassessed By', isSilent: true, candidates: ['ReassessedBy', 'ReassessedBy0', 'cr013_reassessedBy'] },
-    ],
-  },
-];
-
-export const complianceListEntries: readonly SpListEntry[] = [
-// ── 7. コンプライアンス・診断系 ────────────────────────
-  {
-    key: 'compliance_check_rules',
-    displayName: '監査チェックルール',
-    resolve: () => envOr('VITE_SP_LIST_COMPLIANCE', fromConfig(ListKeys.ComplianceCheckRules)),
-    operations: ['R'],
-    category: 'compliance',
-    lifecycle: 'optional',
-    provisioningFields: [
-      { internalName: 'RuleID', type: 'Text', displayName: 'Rule ID', required: true, indexed: true },
-      { internalName: 'Checkpoint', type: 'Text', displayName: 'Checkpoint', required: true },
-      { internalName: 'Criteria', type: 'Note', displayName: 'Criteria', richText: false },
-      { internalName: 'EvidenceRequired', type: 'Note', displayName: 'Evidence Required', richText: false },
-      { internalName: 'RuleName', type: 'Text', displayName: 'Rule Name (Legacy)', governance: 'allow', isSilent: true, candidates: ['RuleName'] },
-      { internalName: 'EvaluationLogic', type: 'Note', displayName: 'Evaluation Logic (Legacy)', richText: false, governance: 'allow', isSilent: true, candidates: ['EvaluationLogic'] },
-      { internalName: 'SeverityLevel', type: 'Text', displayName: 'Severity Level (Legacy)', governance: 'allow', isSilent: true, candidates: ['SeverityLevel'] },
-      { internalName: 'ValidFrom', type: 'DateTime', displayName: 'Valid From (Legacy)', governance: 'allow', isSilent: true, candidates: ['ValidFrom'] },
-      { internalName: 'ValidTo', type: 'DateTime', displayName: 'Valid To (Legacy)', governance: 'allow', isSilent: true, candidates: ['ValidTo'] },
-      { internalName: 'SortOrder', type: 'Number', displayName: 'Sort Order' },
-      { internalName: 'IsActive', type: 'Boolean', displayName: 'Active', default: true },
-    ],
-  },
-  {
-    key: 'diagnostics_reports',
-    displayName: '環境診断レポート',
-    resolve: () => envOr('VITE_SP_LIST_DIAGNOSTICS_REPORTS', fromConfig(ListKeys.DiagnosticsReports)),
-    operations: ['R', 'W'],
-    category: 'compliance',
-    lifecycle: 'optional',
-    provisioningFields: [
-      { internalName: 'Status', type: 'Text', displayName: 'Status (Legacy)', governance: 'allow', isSilent: true },
-      { internalName: 'Notified', type: 'Boolean', displayName: 'Notified (Legacy)', governance: 'allow', isSilent: true },
-      { internalName: 'Overall', type: 'Text', displayName: 'Overall (Legacy)', governance: 'allow', isSilent: true },
-      { internalName: 'Top_x0020_Issue', type: 'Text', displayName: 'Top Issue', candidates: ['Top_x0020_Issue', 'TopIssue'] },
-      { internalName: 'Summary_x0020_Text', type: 'Note', displayName: 'Summary Text (Legacy SP-encoded)', richText: false, governance: 'allow', isSilent: true, candidates: ['Summary_x0020_Text'] },
-      { internalName: 'Report_x0020_Link', type: 'Text', displayName: 'Report Link (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Report_x0020_Link'] },
-      { internalName: 'Notified_x0020_At', type: 'DateTime', displayName: 'Notified At (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Notified_x0020_At'] },
-    ],
-  },
-  {
-    key: 'drift_events_log',
-    displayName: 'ドリフトイベント記録',
-    resolve: () => envOr('VITE_SP_LIST_DRIFT_LOG', 'DriftEventsLog_v2'),
-    operations: ['R', 'W'],
-    category: 'compliance',
-    lifecycle: 'optional',
-    // Severity は SharePointDriftEventRepository が任意扱い（fail-open）で、
-    // かつ本リストは lifecycle: 'optional' な観測用途のため essential から除外する。
-    // 診断契約を実装契約に合わせ、宣言だけが厳しい状態を解消する。
-    essentialFields: ['ListName', 'FieldName', 'DetectedAt', 'DriftType'],
-    provisioningFields: [
-      { internalName: 'ListName', type: 'Text', displayName: 'List Name', required: true, indexed: true, candidates: ['ListName', 'List_x0020_Name', 'cr013_listName'] },
-      { internalName: 'FieldName', type: 'Text', displayName: 'Field Name', required: true, indexed: true, candidates: ['FieldName', 'Field_x0020_Name', 'cr013_fieldName'] },
-      { internalName: 'DetectedAt', type: 'DateTime', displayName: 'Detected At', required: true, candidates: ['DetectedAt', 'Detected_x0020_At', 'cr013_detectedAt'] },
-      { internalName: 'List_x0020_Name', type: 'Text', displayName: 'List Name (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['List_x0020_Name'] },
-      { internalName: 'Field_x0020_Name', type: 'Text', displayName: 'Field Name (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Field_x0020_Name'] },
-      { internalName: 'Detected_x0020_At', type: 'DateTime', displayName: 'Detected At (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Detected_x0020_At'] },
-      { internalName: 'LoggedAt', type: 'DateTime', displayName: 'Logged At', candidates: ['Logged_x0020_At', 'LoggedAt', 'cr013_loggedAt'] },
-      { internalName: 'Severity', type: 'Text', displayName: 'Severity', isSilent: true, candidates: ['Severity', 'Level', 'cr013_severity'] },
-      { internalName: 'ResolutionType', type: 'Text', displayName: 'Resolution Type', isSilent: true, candidates: ['Resolution_x0020_Type', 'ResolutionType', 'cr013_resolutionType'] },
-      { internalName: 'DriftType', type: 'Text', displayName: 'Drift Type', isSilent: true, candidates: ['Drift_x0020_Type', 'DriftType', 'cr013_driftType'] },
-      { internalName: 'Resolved', type: 'Boolean', displayName: 'Resolved', default: false, isSilent: true, candidates: ['Resolved', 'IsResolved', 'cr013_resolved'] },
-      { internalName: 'Description', type: 'Note', displayName: 'Description', richText: false, isSilent: true, candidates: ['Description', 'Details', 'cr013_description'] },
-      { internalName: 'RemediationSource', type: 'Text', displayName: 'Remediation Source', isSilent: true, candidates: ['RemediationSource', 'Source', 'cr013_remediationSource'] },
-    ],
-  },
-  {
-    key: 'remediation_audit_log',
-    displayName: '自動修復・ガバナンス監査ログ',
-    resolve: () => envOr('VITE_SP_LIST_REMEDIATION_LOG', fromConfig(ListKeys.RemediationAuditLog)),
-    operations: ['R', 'W'],
-    category: 'compliance',
-    lifecycle: 'optional',
-    essentialFields: ['PlanId', 'Phase', 'ListKey', 'Action', 'Timestamp'],
-    provisioningFields: [
-      { internalName: 'PlanId', type: 'Text', displayName: 'Plan ID', required: true, indexed: true, candidates: ['PlanId', 'Plan_x0020_ID', 'cr013_planId'] },
-      { internalName: 'Phase', type: 'Text', displayName: 'Phase', required: true, candidates: ['Phase', 'Phase0', 'cr013_phase'] },
-      { internalName: 'ListKey', type: 'Text', displayName: 'List Key', required: true, indexed: true, isSilent: true, candidates: ['ListKey', 'List_x0020_Key', 'cr013_listKey'] },
-      { internalName: 'Action', type: 'Text', displayName: 'Action', required: true, candidates: ['Action', 'Action0', 'cr013_action'] },
-      { internalName: 'Timestamp', type: 'DateTime', displayName: 'Timestamp', required: true, candidates: ['Timestamp', 'Timestamp0', 'cr013_timestamp'] },
-      { internalName: 'Plan_x0020_ID', type: 'Text', displayName: 'Plan ID (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Plan_x0020_ID'] },
-      { internalName: 'CorrelationId', type: 'Text', displayName: 'Correlation ID (Legacy)', governance: 'allow', isSilent: true, candidates: ['CorrelationId'] },
-      { internalName: 'Correlation_x0020_ID', type: 'Text', displayName: 'Correlation ID (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Correlation_x0020_ID'] },
-      { internalName: 'Status', type: 'Text', displayName: 'Status', isSilent: true, candidates: ['Status', 'ExecutionStatus', 'cr013_status'] },
-      { internalName: 'Payload', type: 'Note', displayName: 'Payload JSON', richText: false, governance: 'allow', isSilent: true, candidates: ['Payload', 'payload', 'cr013_payload', 'cr013_draftJson', 'PayloadJSON', 'SupportRecordPayload', 'Payload_x0020_JSON', 'Observation'] },
-    ],
-  },
-];
-
-export const otherListEntries: readonly SpListEntry[] = [
-// ── 8. その他 ─────────────────────────────────────────
-  {
-    key: 'survey_tokusei',
-    displayName: '特性アンケート',
-    resolve: () => envOr('VITE_SP_LIST_SURVEY_TOKUSEI', fromConfig(ListKeys.SurveyTokusei)),
-    operations: ['R'],
-    category: 'other',
-    lifecycle: 'optional',
-    provisioningFields: [
-      { internalName: 'Notes', type: 'Note', displayName: 'Notes', richText: false },
-      { internalName: 'StartTime', type: 'DateTime', displayName: 'Start Time' },
-      { internalName: 'EndTime', type: 'DateTime', displayName: 'End Time' },
-      { internalName: 'ResponderEmail', type: 'Text', displayName: 'Responder Email', candidates: ['ResponderEmail', 'Responder_x0020_Email'] },
-      { internalName: 'ResponderName', type: 'Text', displayName: 'Responder Name', candidates: ['ResponderName', 'Responder_x0020_Name'] },
-      { internalName: 'FillDate', type: 'DateTime', displayName: 'Fill Date', candidates: ['FillDate', 'Fill_x0020_Date'] },
-      { internalName: 'TargetUserName', type: 'Text', displayName: 'Target User Name', candidates: ['TargetUserName', 'Target_x0020_User_x0020_Name'] },
-      { internalName: 'FormRowId', type: 'Text', displayName: 'Form Row ID', candidates: ['FormRowId', 'Form_x0020_Row_x0020_Id'] },
-      { internalName: 'Strengths', type: 'Note', displayName: 'Strengths', richText: false, candidates: ['Strengths', 'UserStrengths'] },
-      { internalName: 'RelationalDifficulties', type: 'Note', displayName: 'Relational Difficulties', richText: false, candidates: ['RelationalDifficulties', 'Relational_x0020_Difficulties'] },
-      { internalName: 'SituationalUnderstanding', type: 'Note', displayName: 'Situational Understanding', richText: false, candidates: ['SituationalUnderstanding', 'Situational_x0020_Understanding'] },
-      { internalName: 'Hearing', type: 'Note', displayName: 'Hearing', richText: false, candidates: ['Hearing', 'Sensory_x0020_Hearing'] },
-      { internalName: 'Vision', type: 'Note', displayName: 'Vision', richText: false, candidates: ['Vision', 'Sensory_x0020_Vision'] },
-      { internalName: 'Touch', type: 'Note', displayName: 'Touch', richText: false, candidates: ['Touch', 'Sensory_x0020_Touch'] },
-      { internalName: 'Smell', type: 'Note', displayName: 'Smell', richText: false, candidates: ['Smell', 'Sensory_x0020_Smell'] },
-      { internalName: 'Taste', type: 'Note', displayName: 'Taste', richText: false, candidates: ['Taste', 'Sensory_x0020_Taste'] },
-      { internalName: 'SensoryMultiSelect', type: 'Text', displayName: 'Sensory Multi Select', candidates: ['SensoryMultiSelect', 'Sensory_x0020_Multiple'] },
-      { internalName: 'SensoryFreeText', type: 'Note', displayName: 'Sensory Free Text', richText: false, candidates: ['SensoryFreeText', 'Sensory_x0020_Notes'] },
-      { internalName: 'DifficultyWithChanges', type: 'Note', displayName: 'Difficulty With Changes', richText: false, candidates: ['DifficultyWithChanges', 'Change_x0020_Difficulties'] },
-      { internalName: 'InterestInParts', type: 'Note', displayName: 'Interest In Parts', richText: false, candidates: ['InterestInParts', 'Part_x0020_Interests'] },
-      { internalName: 'RepetitiveBehaviors', type: 'Note', displayName: 'Repetitive Behaviors', richText: false, candidates: ['RepetitiveBehaviors', 'Repetitive_x0020_Actions'] },
-      { internalName: 'FixedHabits', type: 'Note', displayName: 'Fixed Habits', richText: false, candidates: ['FixedHabits', 'Fixed_x0020_Routines'] },
-      { internalName: 'ComprehensionDifficulty', type: 'Note', displayName: 'Comprehension Difficulty', richText: false, candidates: ['ComprehensionDifficulty', 'Comprehension_x0020_Issues'] },
-      { internalName: 'ExpressionDifficulty', type: 'Note', displayName: 'Expression Difficulty', richText: false, candidates: ['ExpressionDifficulty', 'Expression_x0020_Issues'] },
-      { internalName: 'InteractionDifficulty', type: 'Note', displayName: 'Interaction Difficulty', richText: false, candidates: ['InteractionDifficulty', 'Interaction_x0020_Issues'] },
-      { internalName: 'BehaviorMultiSelect', type: 'Text', displayName: 'Behavior Multi Select', candidates: ['BehaviorMultiSelect', 'Behavior_x0020_Multiple'] },
-      { internalName: 'BehaviorEpisodes', type: 'Note', displayName: 'Behavior Episodes', richText: false, candidates: ['BehaviorEpisodes', 'Behavior_x0020_Details'] },
-    ],
-  },
-  {
-    key: 'survey_tokusei_v2',
-    displayName: '特性確認シート',
-    resolve: () => envOr('VITE_SP_LIST_SURVEY_TOKUSEI_V2', fromConfig(ListKeys.SurveyTokuseiV2)),
-    operations: ['R'],
-    category: 'other',
-    lifecycle: 'optional',
-    provisioningFields: [
-      { internalName: 'Notes', type: 'Note', displayName: 'Notes', richText: false },
-      { internalName: 'StartTime', type: 'DateTime', displayName: 'Start Time' },
-      { internalName: 'EndTime', type: 'DateTime', displayName: 'End Time' },
-      { internalName: 'ResponderEmail', type: 'Text', displayName: 'Responder Email', candidates: ['ResponderEmail', 'Responder_x0020_Email'] },
-      { internalName: 'ResponderName', type: 'Text', displayName: 'Responder Name', candidates: ['ResponderName', 'Responder_x0020_Name'] },
-      { internalName: 'FillDate', type: 'DateTime', displayName: 'Fill Date', candidates: ['FillDate', 'Fill_x0020_Date'] },
-      { internalName: 'TargetUserName', type: 'Text', displayName: 'Target User Name', candidates: ['TargetUserName', 'Target_x0020_User_x0020_Name'] },
-      { internalName: 'FormRowId', type: 'Text', displayName: 'Form Row ID', candidates: ['FormRowId', 'Form_x0020_Row_x0020_Id'] },
-      { internalName: 'Strengths', type: 'Note', displayName: 'Strengths', richText: false, candidates: ['Strengths', 'UserStrengths'] },
-      { internalName: 'RelationalDifficulties', type: 'Note', displayName: 'Relational Difficulties', richText: false, candidates: ['RelationalDifficulties', 'Relational_x0020_Difficulties'] },
-      { internalName: 'SituationalUnderstanding', type: 'Note', displayName: 'Situational Understanding', richText: false, candidates: ['SituationalUnderstanding', 'Situational_x0020_Understanding'] },
-      { internalName: 'Hearing', type: 'Note', displayName: 'Hearing', richText: false, candidates: ['Hearing', 'Sensory_x0020_Hearing'] },
-      { internalName: 'Vision', type: 'Note', displayName: 'Vision', richText: false, candidates: ['Vision', 'Sensory_x0020_Vision'] },
-      { internalName: 'Touch', type: 'Note', displayName: 'Touch', richText: false, candidates: ['Touch', 'Sensory_x0020_Touch'] },
-      { internalName: 'Smell', type: 'Note', displayName: 'Smell', richText: false, candidates: ['Smell', 'Sensory_x0020_Smell'] },
-      { internalName: 'Taste', type: 'Note', displayName: 'Taste', richText: false, candidates: ['Taste', 'Sensory_x0020_Taste'] },
-      { internalName: 'SensoryMultiSelect', type: 'Text', displayName: 'Sensory Multi Select', candidates: ['SensoryMultiSelect', 'Sensory_x0020_Multiple'] },
-      { internalName: 'SensoryFreeText', type: 'Note', displayName: 'Sensory Free Text', richText: false, candidates: ['SensoryFreeText', 'Sensory_x0020_Notes'] },
-      { internalName: 'DifficultyWithChanges', type: 'Note', displayName: 'Difficulty With Changes', richText: false, candidates: ['DifficultyWithChanges', 'Change_x0020_Difficulties'] },
-      { internalName: 'InterestInParts', type: 'Note', displayName: 'Interest In Parts', richText: false, candidates: ['InterestInParts', 'Part_x0020_Interests'] },
-      { internalName: 'RepetitiveBehaviors', type: 'Note', displayName: 'Repetitive Behaviors', richText: false, candidates: ['RepetitiveBehaviors', 'Repetitive_x0020_Actions'] },
-      { internalName: 'FixedHabits', type: 'Note', displayName: 'Fixed Habits', richText: false, candidates: ['FixedHabits', 'Fixed_x0020_Routines'] },
-      { internalName: 'ComprehensionDifficulty', type: 'Note', displayName: 'Comprehension Difficulty', richText: false, candidates: ['ComprehensionDifficulty', 'Comprehension_x0020_Issues'] },
-      { internalName: 'ExpressionDifficulty', type: 'Note', displayName: 'Expression Difficulty', richText: false, candidates: ['ExpressionDifficulty', 'Expression_x0020_Issues'] },
-      { internalName: 'InteractionDifficulty', type: 'Note', displayName: 'Interaction Difficulty', richText: false, candidates: ['InteractionDifficulty', 'Interaction_x0020_Issues'] },
-      { internalName: 'BehaviorMultiSelect', type: 'Text', displayName: 'Behavior Multi Select', candidates: ['BehaviorMultiSelect', 'Behavior_x0020_Multiple'] },
-      { internalName: 'BehaviorEpisodes', type: 'Note', displayName: 'Behavior Episodes', richText: false, candidates: ['BehaviorEpisodes', 'Behavior_x0020_Details'] },
-    ],
-  },
-  {
-    key: 'nurse_observations',
-    displayName: '看護観察記録',
-    resolve: () => envOr('VITE_SP_LIST_NURSE_OBSERVATION', fromConfig(ListKeys.NurseObservations)),
-    operations: ['R', 'W'],
-    category: 'other',
-    lifecycle: 'optional',
-    essentialFields: ['UserLookupId', 'ObservedAt', 'Temperature'],
-    provisioningFields: [
-      { internalName: 'UserLookupId', type: 'Number', displayName: 'User Lookup ID', required: true, isSilent: true, candidates: ['UserLookupId', 'User_x0020_Lookup_x0020_ID'] },
-      { internalName: 'ObservedAt', type: 'DateTime', displayName: 'Observed At', required: true, dateTimeFormat: 'DateTime', isSilent: true, candidates: ['ObservedAt', 'Observed_x0020_At', 'Created'] },
-      { internalName: 'Temperature', type: 'Number', displayName: 'Temperature', required: true, candidates: ['Temperature', 'Temperature_x0020_Value'] },
-      { internalName: 'Pulse', type: 'Number', displayName: 'Pulse', candidates: ['Pulse'] },
-      { internalName: 'Systolic', type: 'Number', displayName: 'Systolic BP', isSilent: true, candidates: ['Systolic', 'Systolic_x0020_BP'] },
-      { internalName: 'Diastolic', type: 'Number', displayName: 'Diastolic BP', isSilent: true, candidates: ['Diastolic', 'Diastolic_x0020_BP'] },
-      { internalName: 'SpO2', type: 'Number', displayName: 'SpO2', isSilent: true, candidates: ['SpO2', 'SPO2', '_x0053_pO2'] },
-      { internalName: 'Weight', type: 'Number', displayName: 'Weight', candidates: ['Weight'] },
-      { internalName: 'Memo', type: 'Note', displayName: 'Memo', richText: false, candidates: ['Memo', 'Notes'] },
-      { internalName: 'Tags', type: 'Text', displayName: 'Tags', candidates: ['Tags'] },
-      { internalName: 'IdempotencyKey', type: 'Text', displayName: 'Idempotency Key', isSilent: true, candidates: ['IdempotencyKey', 'Idempotency_x0020_Key'] },
-      { internalName: 'Source', type: 'Text', displayName: 'Source', candidates: ['Source'] },
-      { internalName: 'LocalTimeZone', type: 'Text', displayName: 'Local Time Zone', isSilent: true, candidates: ['LocalTimeZone', 'Local_x0020_Time_x0020_Zone'] },
-      { internalName: 'CreatedBy', type: 'Text', displayName: 'Created By', isSilent: true, candidates: ['CreatedBy', 'Created_x0020_By'] },
-      { internalName: 'DeviceId', type: 'Text', displayName: 'Device ID', isSilent: true, candidates: ['DeviceId', 'Device_x0020_ID'] },
-      { internalName: 'VitalsJson', type: 'Note', displayName: 'Vitals JSON', richText: false, isSilent: true, candidates: ['VitalsJson', 'Vitals_x0020_JSON'] },
-    ],
-  },
-  {
-    key: 'official_forms',
-    displayName: '公式帳票ライブラリ',
-    resolve: () => envOr('VITE_SP_LIST_OFFICIAL_FORMS', 'OfficialForms'),
-    operations: ['W'],
-    category: 'other',
-    lifecycle: 'optional',
-    baseTemplate: 101, // Document Library
-    provisioningFields: [
-      { internalName: 'Title', type: 'Text', displayName: 'Title' },
-      { internalName: 'TemplateName', type: 'Text', displayName: 'Template Name', candidates: ['Template_x0020_Name', 'TemplateName'] },
-      { internalName: 'GeneratedAt', type: 'DateTime', displayName: 'Generated At', candidates: ['Generated_x0020_At', 'GeneratedAt'] },
-    ],
-  },
-  {
-    key: 'billing_orders',
-    displayName: '請求オーダー',
-    resolve: () => envOr('VITE_SP_LIST_BILLING_ORDERS', 'BillingOrders'),
-    operations: ['R'],
-    category: 'other',
-    lifecycle: 'optional',
-    provisioningFields: [
-      { internalName: 'Title', type: 'Text', displayName: 'Order Date' },
-      { internalName: 'OrdererCode', type: 'Text', displayName: 'Orderer Code', required: true, governance: 'allow', isSilent: true, candidates: ['OrdererCode', 'Orderer_x0020_Code'] },
-      { internalName: 'OrdererName', type: 'Text', displayName: 'Orderer Name', governance: 'allow', isSilent: true, candidates: ['OrdererName', 'Orderer_x0020_Name'] },
-      { internalName: 'OrderCount', type: 'Number', displayName: 'Order Count', governance: 'allow', isSilent: true, candidates: ['OrderCount', 'Order_x0020_Count'] },
-      { internalName: 'Item', type: 'Text', displayName: 'Item', governance: 'allow' },
-    ],
-  },
-  {
-    key: 'billing_summary',
-    displayName: '月次請求サマリー',
-    resolve: () => envOr('VITE_SP_LIST_BILLING_SUMMARY', fromConfig(ListKeys.MonthlyRecordSummary)),
-    operations: ['R', 'W'],
-    category: 'other',
-    lifecycle: 'optional',
-    essentialFields: ['UserId', 'YearMonth', 'KPI_TotalDays'],
-    provisioningFields: [
-      { internalName: 'UserId', type: 'Text', displayName: 'User ID', required: true, indexed: true, candidates: ['UserId', 'User_x0020_ID'] },
-      { internalName: 'YearMonth', type: 'Text', displayName: 'Year Month', required: true, indexed: true, candidates: ['YearMonth', 'Year_x0020_Month'] },
-      { internalName: 'DisplayName', type: 'Text', displayName: 'Display Name', isSilent: true, candidates: ['Display_x0020_Name', 'DisplayName', 'cr013_displayName'] },
-      { internalName: 'LastUpdated', type: 'DateTime', displayName: 'Last Updated', isSilent: true, candidates: ['LastAggregatedAt', 'LastUpdated', 'Last_x0020_Updated'] },
-      { internalName: 'KPI_TotalDays', type: 'Number', displayName: 'Total Days', candidates: ['KPI_TotalDays', 'TotalDays', 'WorkingDays', 'Total_x0020_Days'], legacyCandidates: ['TotalDays', 'Total_x0020_Days'] },
-      { internalName: 'KPI_PlannedRows', type: 'Number', displayName: 'Planned Rows', isSilent: true, candidates: ['Planned_x0020_Rows', 'KPI_PlannedRows', 'PlannedRows', 'cr013_plannedRows'] },
-      { internalName: 'KPI_CompletedRows', type: 'Number', displayName: 'Completed Rows', isSilent: true, candidates: ['Completed_x0020_Rows', 'KPI_CompletedRows', 'CompletedRows', 'CompletedCount', 'cr013_completedRows'] },
-      { internalName: 'KPI_InProgressRows', type: 'Number', displayName: 'In Progress Rows', isSilent: true, candidates: ['In_x0020_Progress_x0020_Rows', 'KPI_InProgressRows', 'InProgressRows', 'PendingCount', 'cr013_inProgressRows'] },
-      { internalName: 'KPI_EmptyRows', type: 'Number', displayName: 'Empty Rows', isSilent: true, candidates: ['Empty_x0020_Rows', 'KPI_EmptyRows', 'EmptyRows', 'EmptyCount', 'cr013_emptyRows'] },
-      { internalName: 'KPI_SpecialNotes', type: 'Number', displayName: 'Special Notes', isSilent: true, candidates: ['Special_x0020_Notes', 'KPI_SpecialNotes', 'SpecialNotes', 'SpecialNoteCount', 'cr013_specialNotes'] },
-      { internalName: 'KPI_Incidents', type: 'Number', displayName: 'Incidents', isSilent: true, candidates: ['Incidents', 'KPI_Incidents', 'IncidentCount', 'cr013_incidents'] },
-      { internalName: 'CompletionRate', type: 'Number', displayName: 'Completion Rate' },
-      { internalName: 'Key', type: 'Text', displayName: 'Idempotency Key (Legacy)', governance: 'allow', isSilent: true, candidates: ['Key'] },
-      { internalName: 'IsLocked', type: 'Boolean', displayName: 'Locked Flag (Legacy)', governance: 'allow', isSilent: true, candidates: ['IsLocked'] },
-      { internalName: 'TotalDays', type: 'Number', displayName: 'Total Days (Legacy)', governance: 'allow', isSilent: true, candidates: ['TotalDays'] },
-      { internalName: 'Total_x0020_Days', type: 'Number', displayName: 'Total Days (Legacy SP-encoded)', governance: 'allow', isSilent: true, candidates: ['Total_x0020_Days'] },
-      { internalName: 'CompletedCount', type: 'Number', displayName: 'Completed Count (Legacy)', governance: 'allow', isSilent: true, candidates: ['CompletedCount'] },
-      { internalName: 'PendingCount', type: 'Number', displayName: 'Pending Count (Legacy)', governance: 'allow', isSilent: true, candidates: ['PendingCount'] },
-      { internalName: 'EmptyCount', type: 'Number', displayName: 'Empty Count (Legacy)', governance: 'allow', isSilent: true, candidates: ['EmptyCount'] },
-      { internalName: 'IncidentCount', type: 'Number', displayName: 'Incident Count (Legacy)', governance: 'allow', isSilent: true, candidates: ['IncidentCount'] },
-      { internalName: 'SpecialNoteCount', type: 'Number', displayName: 'Special Note Count (Legacy)', governance: 'allow', isSilent: true, candidates: ['SpecialNoteCount'] },
-      { internalName: 'FirstEntryDate', type: 'DateTime', displayName: 'First Entry Date', dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['First_x0020_Entry_x0020_Date', 'FirstEntryDate', 'cr013_firstEntryDate'] },
-      { internalName: 'LastEntryDate', type: 'DateTime', displayName: 'Last Entry Date', dateTimeFormat: 'DateOnly', isSilent: true, candidates: ['Last_x0020_Entry_x0020_Date', 'LastEntryDate', 'cr013_lastEntryDate'] },
-      { internalName: 'IdempotencyKey', type: 'Text', displayName: 'Idempotency Key', isSilent: true, candidates: ['Idempotency_x0020_Key', 'IdempotencyKey', 'Key', 'cr013_idempotencyKey'] },
-    ],
-  },
-
-  {
-    key: 'pdf_output_log',
-    displayName: '帳票出力ログ',
-    resolve: () => envOr('VITE_SP_LIST_PDF_OUTPUT_LOG', fromConfig(ListKeys.PdfOutputLog)),
-    operations: ['R', 'W'],
-    category: 'other',
-    lifecycle: 'optional',
-    provisioningFields: [
-      { internalName: 'OutputType', type: 'Text', displayName: 'Output Type', candidates: ['OutputType'] },
-      { internalName: 'UserCode', type: 'Text', displayName: 'User Code', candidates: ['UserCode'] },
-      { internalName: 'OutputDate', type: 'DateTime', displayName: 'Output Date', dateTimeFormat: 'DateTime', candidates: ['OutputDate'] },
-      { internalName: 'TargetPeriod', type: 'Text', displayName: 'Target Period', indexed: true, candidates: ['TargetPeriod'] },
-      { internalName: 'FileName', type: 'Text', displayName: 'File Name', candidates: ['FileName'] },
-      { internalName: 'FileUrl', type: 'Text', displayName: 'File URL', isSilent: true, candidates: ['FileUrl'] },
-      { internalName: 'OutputBy', type: 'Text', displayName: 'Output By', candidates: ['OutputBy'] },
-      { internalName: 'Status', type: 'Text', displayName: 'Status', candidates: ['Status'] },
-      { internalName: 'ErrorMessage', type: 'Note', displayName: 'Error Message', richText: false, isSilent: true, candidates: ['ErrorMessage'] },
-      { internalName: 'Source', type: 'Text', displayName: 'Source', candidates: ['Source'] },
-    ],
-  },
-  {
-    key: 'call_logs',
-    displayName: '電話受付ログ',
-    resolve: () => envOr('VITE_SP_LIST_CALL_LOGS', fromConfig(ListKeys.CallLogs)),
-    operations: ['R', 'W'],
-    category: 'other',
-    lifecycle: 'optional',
-    provisioningFields: [
-      { internalName: 'ReceivedAt', type: 'DateTime', displayName: 'Received At', candidates: ['ReceivedAt'] },
-      { internalName: 'CallerName', type: 'Text', displayName: 'Caller Name', candidates: ['CallerName'] },
-      { internalName: 'CallerOrg', type: 'Text', displayName: 'Caller Org', isSilent: true, candidates: ['CallerOrg'] },
-      { internalName: 'TargetStaffName', type: 'Text', displayName: 'Target Staff Name', candidates: ['TargetStaffName'] },
-      { internalName: 'ReceivedByName', type: 'Text', displayName: 'Received By Name', candidates: ['ReceivedByName'] },
-      { internalName: 'MessageBody', type: 'Note', displayName: 'Message Body', richText: false, candidates: ['MessageBody'] },
-      { internalName: 'NeedCallback', type: 'Boolean', displayName: 'Need Callback', isSilent: true, candidates: ['NeedCallback'] },
-      { internalName: 'Urgency', type: 'Text', displayName: 'Urgency', candidates: ['Urgency'] },
-      { internalName: 'Status', type: 'Text', displayName: 'Status', candidates: ['Status'] },
-      { internalName: 'RelatedUserId', type: 'Text', displayName: 'Related User ID', isSilent: true, candidates: ['RelatedUserId'] },
-      { internalName: 'RelatedUserName', type: 'Text', displayName: 'Related User Name', isSilent: true, candidates: ['RelatedUserName'] },
-      { internalName: 'CallbackDueAt', type: 'DateTime', displayName: 'Callback Due At', isSilent: true, candidates: ['CallbackDueAt'] },
-      { internalName: 'CompletedAt', type: 'DateTime', displayName: 'Completed At', isSilent: true, candidates: ['CompletedAt'] },
-    ],
-  },
-];
+export {
+  masterListEntries,
+  dailyListEntries,
+  attendanceListEntries,
+  scheduleListEntries,
+  meetingListEntries,
+  handoffListEntries,
+  complianceListEntries,
+  otherListEntries,
+};
 
 /** 全リスト定義の統合（参照等価性を維持するため） */
 export const listDefinitions: readonly SpListEntry[] = [
